### 2018-01-10 #184 @jeffreyhanson
Sure, I'm happy to review this. I'm pretty busy at the moment though, so it might be a few weeks before I can get around to it if that's ok?
