### 2018-01-10 #184 @karthik
### Editor checks:

---

#### Editor comments

Work in progress (dealing with installation issues)

---

Reviewers: @lwasser and @jeffreyhanson
Due date:  Feb 10, 2018
