### 2018-01-10 #184 @lbusett
@karthik Sorry to hear that you having problems in installation.... :worried: Unfortunately, as I reported in the submission, that is a common problem related to the requirements of the `gWidgetsRGtk2` package. Instructions reported here: http://lbusett.github.io/MODIStsp/articles/installation.html are usually sufficient to solve the issue (unless you are working on a mac with os sierra, in which case it can be more dificult...).

Let me know if I can help in any way!
