### 2018-01-31 #184 @jeffreyhanson
Awesome work @lbusett! This package looks absolutely amazing. I've just started working on my review, but I thought I would let you know that some of the unit tests fail on my system in case you wanted to address this before I submit my review. I've included the output below and a copy of my session info.

```
> test()
Loading MODIStsp
Testing MODIStsp
✔ | OK F W S | Context
✔ |  4       | Check proper functioning of bbox_from_file [1.6 s]
✔ |  4       | get_yeardates works as expected when download_range == "Full"
✔ |  6       | get_yeardates works as expected when download_range == "Seasonal"
✔ |  7       | Check proper functioning of MODIStsp_addindex [0.2 s]
✔ | 21       | MODIStsp_extract works as expected [58.7 s]
⠏ |  0       | Check proper functioning of the main MODIStsp GUI
(R:19949): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 83: Element 'markup' was closed, but the currently open element is 'span'

(R:19949): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 83: Element 'markup' was closed, but the currently open element is 'span'

(R:19949): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 87: Element 'markup' was closed, but the currently open element is 'span'

(R:19949): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 87: Element 'markup' was closed, but the currently open element is 'span'

(R:19949): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 83: Element 'markup' was closed, but the currently open element is 'span'

(R:19949): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 83: Element 'markup' was closed, but the currently open element is 'span'

(R:19949): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 83: Element 'markup' was closed, but the currently open element is 'span'

(R:19949): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 83: Element 'markup' was closed, but the currently open element is 'span'
✔ |  0       | Check proper functioning of the main MODIStsp GUI [4.1 s]
✔ |  6   2   | MODIStsp Test 0: Gracefully fail on input problems [14.9 s]
────────────────────────────────────────────────────────────────────────────────
test_MODIStsp.R:13: warning: Tests on MODIStsp
input string 1 is invalid in this locale

test_MODIStsp.R:18: warning: Tests on MODIStsp
input string 1 is invalid in this locale
────────────────────────────────────────────────────────────────────────────────
⠏ |  0       | MODIStsp Test 1: Basic processing on bands and quality
⠋ |  1       | MODIStsp Test 1: Basic processing on bands and quality
⠙ |  2       | MODIStsp Test 1: Basic processing on bands and quality
✔ |  2       | MODIStsp Test 1: Basic processing on bands and quality
        indicators [5.3 s]
✔ |  4       | MODIStsp Test 2: Geometric operations [3.2 s]
⠏ |  0       | MODIStsp Test 3: Computation of spectral indices and
            creation of time seriesWarning 1: Metadata exceeding 32000 bytes cannot be written into GeoTIFF. Transferred to PAM instead.
⠋ |  1       | MODIStsp Test 3: Computation of spectral indices and
⠙ |  2       | MODIStsp Test 3: Computation of spectral indices and
⠹ |  3       | MODIStsp Test 3: Computation of spectral indices and
⠸ |  4       | MODIStsp Test 3: Computation of spectral indices and
⠼ |  5       | MODIStsp Test 3: Computation of spectral indices and
            creation of time seriesWarning 1: Metadata exceeding 32000 bytes cannot be written into GeoTIFF. Transferred to PAM instead.
⠴ |  6       | MODIStsp Test 3: Computation of spectral indices and
⠦ |  7       | MODIStsp Test 3: Computation of spectral indices and
✔ |  7       | MODIStsp Test 3: Computation of spectral indices and
            creation of time series [22.0 s]
✖ |  0 1     | MODIStsp Test 4: HTTP download from NSIDC (seasonal) [5.8 s]
────────────────────────────────────────────────────────────────────────────────
test_MODIStsp.R:193: error: Tests on MODIStsp
argument is of length zero
1: MODIStsp(test = 4) at /home/jeff/GitHub/MODIStsp/tests/testthat/test_MODIStsp.R:193
2: MODIStsp_process(sel_prod = general_opts$sel_prod, start_date = general_opts$start_date,
       end_date = general_opts$end_date, out_folder = general_opts$out_folder, out_folder_mod = general_opts$out_folder_mod,
       reprocess = general_opts$reprocess, delete_hdf = general_opts$delete_hdf, sensor = general_opts$sensor,
       download_server = general_opts$download_server, user = general_opts$user, password = general_opts$password,
       https = prod_opts$http, ftps = prod_opts$ftp, start_x = general_opts$start_x,
       start_y = general_opts$start_y, end_x = general_opts$end_x, end_y = general_opts$end_y,
       full_ext = general_opts$full_ext, bbox = general_opts$bbox, out_format = general_opts$out_format,
       out_res_sel = general_opts$out_res_sel, out_res = as.numeric(general_opts$out_res),
       native_res = prod_opts$native_res, tiled = prod_opts$tiled, resampling = general_opts$resampling,
       ts_format = general_opts$ts_format, compress = general_opts$compress, mod_proj_str = mod_proj_str,
       outproj_str = general_opts$user_proj4, nodata_in = prod_opts$nodata_in, nodata_out = prod_opts$nodata_out,
       rts = general_opts$rts, nodata_change = general_opts$nodata_change, scale_val = general_opts$scale_val,
       scale_factor = prod_opts$scale_factor, offset = prod_opts$offset, datatype = prod_opts$datatype,
       bandsel = general_opts$bandsel, bandnames = prod_opts$bandnames, indexes_bandsel = c(general_opts$indexes_bandsel),
       indexes_bandnames = c(prod_opts$indexes_bandnames, custom_idx$indexes_bandnames),
       indexes_formula = c(prod_opts$indexes_formula, custom_idx$indexes_formulas),
       indexes_nodata_out = c(prod_opts$indexes_nodata_out, custom_idx$indexes_nodata_out),
       quality_bandnames = prod_opts$quality_bandnames, quality_bandsel = general_opts$quality_bandsel,
       quality_bitN = prod_opts$quality_bitN, quality_source = prod_opts$quality_source,
       quality_nodata_in = prod_opts$quality_nodata_in, quality_nodata_out = prod_opts$quality_nodata_out,
       file_prefixes = prod_opts$file_prefix, main_out_folder = prod_opts$main_out_folder,
       gui = gui, use_aria = general_opts$use_aria, download_range = general_opts$download_range,
       n_retries = n_retries, verbose = verbose) at /home/jeff/GitHub/MODIStsp/R/MODIStsp.R:464
3: MODIStsp_download(modislist, out_folder_mod, download_server, http, ftp, n_retries,
       use_aria, date_dirs[date], year, DOY, user, password, sens_sel, date_name, gui,
       mess_lab, verbose) at /home/jeff/GitHub/MODIStsp/R/MODIStsp_process.R:388
────────────────────────────────────────────────────────────────────────────────
✖ |  0 1     | MODIStsp Test 5: HTTP download from USGS, resize and reproject [14.5 s]
────────────────────────────────────────────────────────────────────────────────
test_MODIStsp.R:249: error: Tests on MODIStsp
argument is of length zero
1: MODIStsp(test = 5) at /home/jeff/GitHub/MODIStsp/tests/testthat/test_MODIStsp.R:249
2: MODIStsp_process(sel_prod = general_opts$sel_prod, start_date = general_opts$start_date,
       end_date = general_opts$end_date, out_folder = general_opts$out_folder, out_folder_mod = general_opts$out_folder_mod,
       reprocess = general_opts$reprocess, delete_hdf = general_opts$delete_hdf, sensor = general_opts$sensor,
       download_server = general_opts$download_server, user = general_opts$user, password = general_opts$password,
       https = prod_opts$http, ftps = prod_opts$ftp, start_x = general_opts$start_x,
       start_y = general_opts$start_y, end_x = general_opts$end_x, end_y = general_opts$end_y,
       full_ext = general_opts$full_ext, bbox = general_opts$bbox, out_format = general_opts$out_format,
       out_res_sel = general_opts$out_res_sel, out_res = as.numeric(general_opts$out_res),
       native_res = prod_opts$native_res, tiled = prod_opts$tiled, resampling = general_opts$resampling,
       ts_format = general_opts$ts_format, compress = general_opts$compress, mod_proj_str = mod_proj_str,
       outproj_str = general_opts$user_proj4, nodata_in = prod_opts$nodata_in, nodata_out = prod_opts$nodata_out,
       rts = general_opts$rts, nodata_change = general_opts$nodata_change, scale_val = general_opts$scale_val,
       scale_factor = prod_opts$scale_factor, offset = prod_opts$offset, datatype = prod_opts$datatype,
       bandsel = general_opts$bandsel, bandnames = prod_opts$bandnames, indexes_bandsel = c(general_opts$indexes_bandsel),
       indexes_bandnames = c(prod_opts$indexes_bandnames, custom_idx$indexes_bandnames),
       indexes_formula = c(prod_opts$indexes_formula, custom_idx$indexes_formulas),
       indexes_nodata_out = c(prod_opts$indexes_nodata_out, custom_idx$indexes_nodata_out),
       quality_bandnames = prod_opts$quality_bandnames, quality_bandsel = general_opts$quality_bandsel,
       quality_bitN = prod_opts$quality_bitN, quality_source = prod_opts$quality_source,
       quality_nodata_in = prod_opts$quality_nodata_in, quality_nodata_out = prod_opts$quality_nodata_out,
       file_prefixes = prod_opts$file_prefix, main_out_folder = prod_opts$main_out_folder,
       gui = gui, use_aria = general_opts$use_aria, download_range = general_opts$download_range,
       n_retries = n_retries, verbose = verbose) at /home/jeff/GitHub/MODIStsp/R/MODIStsp.R:464
3: MODIStsp_download(modislist, out_folder_mod, download_server, http, ftp, n_retries,
       use_aria, date_dirs[date], year, DOY, user, password, sens_sel, date_name, gui,
       mess_lab, verbose) at /home/jeff/GitHub/MODIStsp/R/MODIStsp_process.R:388
────────────────────────────────────────────────────────────────────────────────
  |======================================================================| 100%
  |======================================================================| 100%
  |======================================================================| 100%
  |======================================================================| 100%
✔ |  6       | MODIStsp Test 6: FTP download and mosaicing of MODIS tiles [77.0 s]
✔ |  1       | MODIStsp Test 7: Passing the extent with a spatial file [3.0 s]
⠏ |  0       | MODIStsp Test 8: Fail gracefully on missing connectionError : GET http://e4ftl01.cr.usgs.gov/MOTA/MCD43A3.006/
In addition: Warning message:
`encoding` is deprecated; all files now assumed to be UTF-8
Error : GET ftp://ladsweb.nascom.nasa.gov/allData/6/MCD43A3/2017/
⠋ |  1       | MODIStsp Test 8: Fail gracefully on missing connectionError : GET ftp://ladsweb.nascom.nasa.gov/allData/6/MCD15A2H/2016/
✔ |  2       | MODIStsp Test 8: Fail gracefully on missing connection [0.5 s]
✖ |  0 1     | Check proper functioning of MODIStsp_readxml [1.9 s]
────────────────────────────────────────────────────────────────────────────────
test_MODIStsp_readxml.R:8: error: MODIStsp_readxml works as expected
attempt to select less than one element in OneIndex
1: MODIStsp_read_xml(tmpfile, system.file("ExtData", "MODIStsp_ProdOpts.xml", package = "MODIStsp")) at /home/jeff/GitHub/MODIStsp/tests/testthat/test_MODIStsp_readxml.R:8
────────────────────────────────────────────────────────────────────────────────
✔ |  2       | MODIStsp_resetindexes

══ Results ═════════════════════════════════════════════════════════════════════
Duration: 213.2 s

OK:       72
Failed:   3
Warnings: 2
Skipped:  0
```

Session information
```
R version 3.4.3 (2017-11-30)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 16.04.3 LTS

Matrix products: default
BLAS: /usr/lib/libblas/libblas.so.3.6.0
LAPACK: /usr/lib/lapack/liblapack.so.3.6.0

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C
 [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
 [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] httptest_3.0.0       gWidgetsRGtk2_0.0-85 cairoDevice_2.24
[4] gWidgets_0.0-54      RGtk2_2.20.34        MODIStsp_1.3.3.9000
[7] testthat_2.0.0       devtools_1.13.4

 [1] Rcpp_0.12.15        compiler_3.4.3      R.methodsS3_1.7.1
 [4] bitops_1.0-6        R.utils_2.6.0       iterators_1.0.9
 [7] tools_3.4.3         xts_0.10-1          gdalUtils_2.0.1.7
[10] digest_0.6.15       jsonlite_1.5        memoise_1.1.0
[13] lattice_0.20-35     rlang_0.1.6         foreach_1.4.4
[16] cli_1.0.0           commonmark_1.4      curl_3.1
[19] rgdal_1.2-16        parallel_3.4.3      withr_2.1.1
[22] stringr_1.2.0       httr_1.3.1          roxygen2_6.0.1
[25] raster_2.6-7        xml2_1.2.0          rgeos_0.3-26
[28] grid_3.4.3          data.table_1.10.4-3 R6_2.2.2
[31] pacman_0.4.6        sp_1.2-7            magrittr_1.5
[34] codetools_0.2-15    assertthat_0.2.0    mime_0.5
[37] stringi_1.1.6       RCurl_1.95-4.10     crayon_1.3.4
[40] zoo_1.8-1           R.oo_1.21.0
```

Let me know if there's any other info I can provide.
