### 2018-01-31 68f578e
Fix errors reported in ropensci review
See (https://github.com/ropensci/onboarding/issues/184#issuecomment-362080409)

- FIx bug related to new xml version 
- Fix bug related to changes in metadata retreived from lpdaac
- bumps required xml2 version to 1.2.0
