### 2018-01-31 b286037
Fixes problem on Rd macros
changed from c:\Progra... to c:/Progra
