+++ man/install_MODIStsp_launcher.Rd
@@ -17,7 +17,7 @@ Default: Start Menu -> Programs -> MODIStsp.
 }}
 
 \item{rscript_dir}{'character' in Windows only, the path of the directory in which
-Rscript is installed (default is "C:\Progra~1\R\R-<version>\bin<arch>").
+Rscript is installed (default is "C:/Progra~1/R/R-<version>/bin/<arch>").
 Edit this parameter if R is installed in a custom directory.}
 
 \item{desktop_dir}{\code{character}
