### 2018-01-31 #184 @lbusett
@jeffreyhanson

Thanks! Glad you like  our package!

Concerning the tests, thanks for reporting the issues. They should now be fixed. One was due to a recent change in `xml2` functionality, the other (apparently) to a change in the "standard" http response of NASA servers. If you pull the new github version, the tests should pass (and you should also be able to download data from `http` also outside of the tests, which I think ws not possible due to the NASA response change. I'll have to also test the CRAN version to see if it still works over http.
