### 2018-02-02 e45292e
Fixes regression caused by last commit
There was a small typo in the regex.
