+++ R/MODIStsp.R
@@ -64,9 +64,9 @@
 #' MODIStsp(gui = FALSE, options_file = "X:/yourpath/youroptions.json")}
 #'
 #' # Running the tool using the settings previously saved in a specific option file
-#' # and specifying the extent from a spatial file allows to re-use the same 
+#' # and specifying the extent from a spatial file allows to re-use the same
 #' # processing settings to perform download and reprocessing on a different area
-#' 
+#'
 #' \dontrun{
 #' MODIStsp(gui = FALSE, options_file = "X:/yourpath/youroptions.json",
 #'   spatial_file_path = "X:/yourpath/yourspatialfile.shp" )}
@@ -80,8 +80,8 @@
 #' for (single_shape in extent_list)
 #'   MODIStsp(gui = FALSE, options_file = "X:/yourpath/youroptions.json",
 #'     spatial_file_path = single_shape )}
-#'     
-#' # (See http://lbusett.github.io/MODIStsp/articles for additional details and 
+#'
+#' # (See http://lbusett.github.io/MODIStsp/articles for additional details and
 #' # examples)
 
 MODIStsp <- function(gui               = TRUE,
@@ -89,22 +89,22 @@ MODIStsp <- function(gui               = TRUE,
                      spatial_file_path = NULL,
                      scroll_window     = FALSE,
                      test              = -1,
-                     n_retries         = 20, 
+                     n_retries         = 20,
                      verbose           = TRUE) {
-  
+
   options("guiToolkit" = "RGtk2")
   # Make so that "raster" functions does not automatically add extensions on
   # output files. This is automatically reset to TRUE at the end of the session
   raster::rasterOptions(setfileext = FALSE)
-  
+
   #   __________________________________________________________________________
   #   Initialize processing                                                 ####
-  
+
   #   __________________________________________________________________________
   #   If test mode is selected, select the options file for testing         ####
   #   and set other parameters
   if (test > -1) {
-    
+
     gui <- FALSE
     message("MODIStsp is running in test mode.")
     # read names of available json test files
@@ -132,19 +132,19 @@ MODIStsp <- function(gui               = TRUE,
               unzip   = "internal")
       }
     }
-    
+
     # Assign the selected test Option File
     options_file <- test_files[test]
     start <- TRUE
   }
-  
+
   #   __________________________________________________________________________
   #   On interactive execution, ensure that gWidgetsRGtk2 is avauilable     ####
-  
+
   if (gui) {
     #nocov start
     if (!pacman::p_exists("gWidgetsRGtk2", local = TRUE)) {
-      
+
       message(strwrap("Library 'gWidgetsRGtk2' is not installed. It is required
                       to run MODIStsp in interactive mode!\n\n",
                       "Do you want to install it now?"),
@@ -156,15 +156,15 @@ MODIStsp <- function(gui               = TRUE,
         stop(strwrap("MODIStsp can not work in Interactive mode without
                      gWidgetsRGtk2! Aborting!"))
       }
-      
+
     }
     options("guiToolkit" = "RGtk2")
     #nocov end
   }
-  
+
   # Check GDAL version ----
   if (is.null(getOption("gdalUtils_gdalPath"))) {
-    
+
     welcome_text <- strwrap("Welcome to MODIStsp!\n\nWe will now search for a
                             valid GDAL installation - please wait! (this will
                             happen only once)", width = 60)
@@ -172,7 +172,7 @@ MODIStsp <- function(gui               = TRUE,
       #nocov start
       welcome_win       <- gWidgets::gwindow(title  = "Welcome", width = 400,
                                              height = 100)
-      
+
       welcome_lab       <- gWidgets::glabel(text      = welcome_text,
[…]
