### 2018-02-05 #184 @lwasser
hey @lbusett or @jeffreyhanson i'm a bit confused on the quartz install

i'm here:

next, edit the configure options for GTK to require x11 rather than Quartz:
```
brew edit gtk+
```
in the def install section, remove the reference to quartz and switch to:
```
--with-gdktarget=x11,
--enable-x11-backend
```
And would like to understand what part of the gtk+ file i need to edit. do i just replace with-gdktarget=quartz with two lines as follows:

```
"--with-gdktarget=x11",
"--enable-x11-backend",
```
it's taking me some time to get everything installed and this step is a bit confusing. thank you for any advice!
