### 2018-02-05 #184 @lbusett
hi @lwasser ,

   sorry you are having trouble... unfortunately, at the moent RGtk2 installation on mac is troublesome  (and the solution to the trouble seems to change with time.. ;-(.
Moreover, I know almost nothing about mac, so I admit that I never really tested the instructions provided in the help.... However,  those instructions mean that indeed you need to modify the gtk file so that it reads:

> def install
args = ["--disable-dependency-tracking",
"--disable-silent-rules",
"--prefix=#{prefix}",
"--disable-glibtest",
"--enable-introspection=yes",
"--with-gdktarget=x11",
"--enable-x11-backend"
]

and then proceed with `brew install`.

Concerning the "R session is headless; GTK+ not initialized." :  this does not bode well. It (should) mean that RGtk2 does not find a "graphical environment" to use. I fear that if you try to load `gWidgetsRGtk2` you will have problems.

Again, sorry for the trouble. I hope you'll be able to solve it easily (this link may also help : https://gist.github.com/sebkopf/9405675). Otherwise, I'll try to get my hands on a (modern) mac pc and see if I manage to provide better instructions.
