### 2018-02-05 #184 @lwasser
ok!! i think i just got it. THANK YOU @lbusett  i just installed those packages manually but also i hadn't run that line above that you suggested so i did that too. phew! i'll play with the package more tomorrow. thank you for the help.
