### 2018-02-06 lbusett/MODIStsp#109 @domras65
Problems with GUI window
Dear friends, first of all, congratulation for your MODIStsp package. I installed successfully and worked with it a little bit in order to familiarize with the environment.  However, after an initial trial, using the GUI window, I am not able to reopen it. Each time I try, the machine sends me the following message:

> library(MODIStsp)
> MODIStsp()
Loading required package: gWidgetsRGtk2
Loading required package: RGtk2
Loading required package: gWidgets
Loading required package: cairoDevice
[Tue Feb 06 21:30:12 2018] Welcome to MODIStsp!We will now search for a valid GDAL installation - pleasewait! (this will happen only once)
GDAL version in use: 2.2.0

(rsession.exe:3688): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error en la l�nea 1, car�cter 100: Se cerr� el elemento �markup�, pero el elemento que est� abierto actualmente es �span�

(rsession.exe:3688): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error en la l�nea 1, car�cter 100: Se cerr� el elemento �markup�, pero el elemento que est� abierto actualmente es �span�
Error in if (proj == "longlat") { : missing value where TRUE/FALSE needed
In addition: Warning messages:
1: package ‘gWidgetsRGtk2’ was built under R version 3.4.3
2: package ‘RGtk2’ was built under R version 3.4.3
3: package ‘gWidgets’ was built under R version 3.4.3

By the way, I am using a PC under windows 7.

Best regards
