### 2018-02-06 lbusett/MODIStsp#109 @lbusett
Hi. Sorry for the trouble. Could you tell us if you are using the cran version or the github one?
