### 2018-02-07 #184 @jeffreyhanson
Sorry for the late reply @lbusett, I just installed the latest GitHub version and all the tests pass now. Thank for you fixing this so quickly.
