### 2018-02-07 #184 @lbusett
Good. In the end, both problems were due to a breaking change in  xml2. Since the cran version uses xml, I was not notified about it.

By the way: a user signalled us a bug in the gui that I probably introduced while refactoring the code for ropensci submission. I am going to commit a fix for it ASAP.
