### 2018-02-07 lbusett/MODIStsp#109 @lbusett
Hi. I was able to replicate this on the github version. It appears to happen if you specify a "user defined" projection. At next execution, the GUI fails to open. Could you confirm this?

I'll try to fix it ASAP. In the meantime, you could probably have it working again by deleting
the "MODIStsp_Previous.json" file in the "ExtData/Previous" folder within the MODIStsp folder of your "R" library. If it does not work, you should probably reinstall the package.
