### 2018-02-07 #184 @jeffreyhanson
## Package Review

- [:white_check_mark:] As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

#### Documentation

- [:white_check_mark:] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
- [:white_check_mark:] **Installation instructions:** for the development version of package and any non-standard dependencies in README
- [:white_check_mark:] **Vignette(s)** demonstrating major functionality that runs successfully locally
- [:white_check_mark:] **Function Documentation:** for all exported functions in R help
- [:white_check_mark:] **Examples** for all exported functions in R Help that run successfully locally
- [:white_check_mark:] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

#### Functionality

- [:white_check_mark:] **Installation:** Installation succeeds as documented.
- [:white_check_mark:] **Functionality:** Any functional claims of the software been confirmed.
- [:white_check_mark:] **Performance:** Any performance claims of the software been confirmed.
- [:white_check_mark:] **Automated tests:** Unit tests cover essential functions of the package
- [:x:] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines

#### Final approval (post-review)

Estimated hours spent reviewing:

---

### Review Comments

The _MODIStsp_ package is absolutely brilliant. In my experience, downloading and pre-processing raw MODIS data can be extremely painful. The _MODIStsp_ package automates this; meaning that users can spend less time wrangling data and more time analyzing data! This package also provides a graphical user interface too! The package documentation is comprehensive and well written. It contains several vignettes that cover installation, interactive and non-interactive usage, and frequently asked questions. The package website is stunning. I think this R package will excite both new and advanced R users. All in all, _MODIStsp_ is a wonderful piece of open source software.

#### Major issues

1. A lot of the examples and vignettes require the user to input their own path to store the data. I would find it really useful if these examples just defaulted to using the temporary directory (i.e. `base::tempdir()`) so I could run them by copying and pasting code.

#### Minor issues

1. Running `devtools::document()` yields the following warnings:
```
1: In tools::parse_Rd(path) :
  /home/jeff/GitHub/MODIStsp/man/install_MODIStsp_launcher.Rd:20: unknown macro '\Progra'
2: In tools::parse_Rd(path) :
  /home/jeff/GitHub/MODIStsp/man/install_MODIStsp_launcher.Rd:20: unknown macro '\bin'
```
2. It is my understanding that recent changes to the [CRAN policies](https://cran.r-project.org/web/packages/policies.html) indicate that example code should write to the default temporary directory. If my understanding is correct, then the `MODIStsp_extract` example should use `base::tempdir()` instead of writing to `base::system.file("testdata/", package = "MODIStsp")`. I have copied the relevant text from the CRAN policies below:
```
Packages should not write in the user's home filespace (including clipboards), nor anywhere else on the
file system apart from the R session’s temporary directory (or during installation in the location pointed to
by TMPDIR: and such usage should be cleaned up). Installing into the system’s R installation (e.g.,
scripts to its bin directory) is not allowed.
```
3. The references in _vignettes/MODIStsp.pdf_ are not shown correctly. They contain text like "[-@Hengl2010]" instead of the in-text reference.

#### ROpensci packaging guidelines

1. [This package uses _RCurl_](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#-recommended-scaffolding). From what I can tell, the only usage of _RCurl_ is `RCurl::getURL (line 132 in _R/MODIStsp_download.R_) so you might be able to drop this dependency without too much hassle.
2. [This package uses `base::print` instead of `base::message` in a couple of places.](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#-console-messages) This only happens in two places (lines 236 & 334 in _R/MODIStsp_extract.R_) so you might be able replace this with `base::message` fairly easily.
