### 2018-02-07 lbusett/MODIStsp#109 @domras65
Hi, and I confirm your guess, since I tried to download some staff using a
UTM projection suitable for the Iberian Peninsula (zone UTM30).

Again, wonderful tool.

Best regards

Domingo Rasilla
Department of Geography.
University of Cantabria
39005 Santander Spain

<http://www.avg.com/email-signature?utm_medium=email&utm_source=link&utm_campaign=sig-email&utm_content=webmail>
Libre
de virus. www.avg.com
<http://www.avg.com/email-signature?utm_medium=email&utm_source=link&utm_campaign=sig-email&utm_content=webmail>
<#DAB4FAD8-2DD7-40BB-A1B8-4E2AA1F9FDF2>

2018-02-07 10:11 GMT+01:00 Lorenzo Busetto <notifications@github.com>:

> Hi. I was able to replicate this on the github version. It appears to
> happen if you specify a "user defined" projection. At next execution, the
> GUI fails to open. Could you confirm this?
>
> I'll try to fix it ASAP. In the meantime, you could probably have it
> working again by deleting
> the "MODIStsp_Previous.json" file in the "ExtData/Previous" folder within
> the MODIStsp folder of your "R" library. If it does not work, you should
> probably reinstall the package.
>
