### 2018-02-07 1e792c4
fix bug on gui on user defined proj
