### 2018-02-07 lbusett/MODIStsp#109 @lbusett
Should be fixed now. Install the latest github version for a working version.
