### 2018-02-07 #184 @lbusett
@jeffreyhanson Thank for your very positive review! All your comments make sense and can be implemented fairly easily. I'll wait however also for the review from @lwasser and then address all comments in one go.

@lwasser if you are still working on this, please note that I just committed a new version to github, solving a recently discovered bug which broke the GUI when a user defined projection was selected (https://github.com/lbusett/MODIStsp/issues/109). Apparently, I introduced that bug while refactoiring the code for ropensci. Unfortunately, it is almost impossible to develop unit tests for the GUI functions, making it difficult to debug them...
