### 2018-02-08 lbusett/MODIStsp#109 @domras65
Dear friend, thank you very much. I will do it.

Nevertheless, I have found another problem. When I try to visualize one of
the previous files downloaded by your tool, I am not able to do it. The
syntax used and the response are the following:

tif_file<-"MOD10_A2_MAX_SNW_2000_055.tif"
plot(tif_file)
Error in plot.window(...) : se necesitan valores finitos de 'ylim'
In addition: Warning messages:
1: In xy.coords(x, y, xlabel, ylabel, log) : NAs introduced by coercion
2: In min(x) : no non-missing arguments to min; returning Inf
3: In max(x) : no non-missing arguments to max; returning -Inf

I guess that the problem is related with the projection, and something is
missing, but I do not know what since I am not very familiar with
geographic projections; for Northern Spain, which is the area of interest,
the UTM zone is 30. So, could you take a look to both the tiff file and the
json file in order to help me?

The *.json parameters are the following (except password)
{
  "sel_prod": "Snow_Cov_8-Day_500m (M*D10_A2)",
  "sensor": "Terra",
  "prod_version": "6",
  "start_date": "2000-01-01",
  "end_date": "2018-02-07",
  "bandsel": [1, 0],
  "indexes_bandsel": 0,
  "quality_bandsel": [0, 0, 0, 0, 0, 0, 0, 0],
  "start_x": 17,
  "end_x": 17,
  "start_y": 4,
  "end_y": 4,
  "use_aria": false,
  "download_server": "http",
  "download_range": "full",
  "proj": "User Defined",
  "user_proj4": "+proj=utm +zone=30 +datum=WGS84 +units=m +no_defs
+ellps=WGS84 +towgs84=0,0,0",
  "out_res_sel": "Native",
  "out_res": "native",
  "full_ext": "Resized",
  "resampling": "near",
  "out_format": "GTiff",
  "ts_format": "ENVI Meta Files",
  "rts": "No",
  "compress": "None",
  "nodata_change": "No",
  "scale_val": "No",
  "delete_hdf": "No",
  "reprocess": "No",
  "bbox": [300000, 4720000, 420000, 4830000],
  "out_folder": "C:\\Users\\Domingo\\Documents",
  "out_folder_mod": "C:\\Users\\Domingo\\Documents",
  "MODIStspVersion": "1.3.3.9000",
  "custom_indexes": []
}

<http://www.avg.com/email-signature?utm_medium=email&utm_source=link&utm_campaign=sig-email&utm_content=webmail>
Libre
de virus. www.avg.com
<http://www.avg.com/email-signature?utm_medium=email&utm_source=link&utm_campaign=sig-email&utm_content=webmail>
<#DAB4FAD8-2DD7-40BB-A1B8-4E2AA1F9FDF2>
