### 2018-02-08 lbusett/MODIStsp#109 @lbusett
From the syntax you reported, It seems that you are not really opening the file, so you are trying to plot a string.

You need at least something like:

```
library(raster)
tif_file<-"MOD10_A2_MAX_SNW_2000_055.tif"
r <- raster(tif_file)
plot(r)
```

additionally, the string referencing to the file name needs to be a full path (unless the file is in your working dir).
