### 2018-02-09 #184 @lwasser
Thank you @lbusett  i'm still working on the review. the install set me back this week - so sorry!
I hope to finish it today but i have to try to get it working on my laptop (i was on a desktop tuesday) for that to happen so I **may** have to submit this monday - worse case scenario. It's on my radar tho and the package itself is VERY VERY VERY COOL!
