### 2018-02-12 #184 @lwasser
All - so sorry this is a bit late. I hope it's helpful.

## Package Review

- [x ] As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

#### Documentation

- [x ] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
- [x ] **Installation instructions:** for the development version of package and any non-standard dependencies in README
- [x ] **Vignette(s)** demonstrating major functionality that runs successfully locally
- [ x] **Function Documentation:** for all exported functions in R help
- [x ] **Examples** for all exported functions in R Help that run successfully locally
- [ x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

#### Functionality

- [ x] **Functionality:** Any functional claims of the software been confirmed.
- [x ] **Performance:** Any performance claims of the software been confirmed.
   and a reasonable range of inputs and conditions. All tests pass on the local machine. I DIDN"T RUN UNIT TESTS OR LOOK AT THIS ELEMENT CLOSELY
- [ ] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines SEE OTHER REVIEW!

#### Final approval (post-review)

- [x ] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**

Estimated hours spent reviewing: 5

---

### Review Comments

Overall I think this is a **fantastic** package. Anyone who has downloaded MODIS data knows that it can be a challenge. This allows for a fully reproducible workflow to grab the data. It’s also surprisingly fast! And I didn’t need my login. Hat tip to Lorenzo for creating such a useful package!

My notes below could be in some cases specific to my computer. I began my review my installing, and playing with the interface. I didn’t read the interface overview because I wanted to think about the intuitiveness of the interface itself FIRST. Then read documentation second!

## Installation

I ran into significant issues installing this on  my mac. I still don’t have it working on my laptop but did get it working on my desktop. It would be very nice to have one fully functional set of instructions for mac that is well tested! Here are a few notes:

A few notes:
Quartz needs to be installed first before you install cairo
Initially I ran into an issue with GTK+ not being installed.

This GIST was helpful
https://gist.github.com/sebkopf/9405675
Once I had GTK+ installed RGtk2 installed nicely

THIS Part was confusing to me.

Next, edit the configure options for GTK to require x11 rather than Quartz:
brew edit gtk+
in the def install section, remove the reference to quartz and switch to:
--with-gdktarget=x11,
--enable-x11-backend

Finally once I was able to get RGtk2 installed I got this message:

```
library(RGtk2)
R session is headless; GTK+ not initialized.
```

Once I installed these two (in installed from source just to be sure) that message seemed to go away

GDAL -- I didn’t have gdal setup on my desktop (only my laptop) so I had to install that too. Just a note that the instructions specify H4 support in GDAL and have command line instructions to check for this. However just note that sometimes gdal2 doesn’t play nicely on a mac!

See: [GDAL2 link on a mac](https://www.karambelkar.info/2016/10/gdal-2-on-mac-with-homebrew/ )if gdal isnt linked properly the command line functions don’t work. Just a note! This is tricky open source GIS mac space :)

## Interface

Here I will be a bit nitpicky fully recognizing that this package is AWESOME. As a MAC user, I really don’t like quartz driven apps. It feels like the MAC interface is an afterthought and is clunky and the graphics are not crisp and are hard to read. If you’ve ever used ENVI on a mac you understand. It would be AWESOME to have a BETTER mac specific interface. Probably a ton of work but just better. My two cents.

With that said, the interface itself was overall very intuitive for me to work with. I was able to quickly get to some data without reading any documentation. However - if you want this package to be even more broadly used, a few more help buttons would be nice to include.  More specifics are below:

It’s very nice to have all of this information in one spot vs the NASA website. However I wonder if we could make the interface more intuitive with a few other features.

**Processing layers:** I think what this does is allow me to select whatever products went into the derived product that I select. However, could you add a little help button to clarify that? Also when I select a few layers, there is no indication that I have selected anything in the gui. I can tell I selected a few things by hitting the “Click to Select” button again but could that maybe change in some way to suggest i’ve made a selection?

The **product information** page link is super awesome. I think that is too buried! I’m wondering if it could could go below the product box OR if the product box could have a link NEXT to it or something? It’s too nice of a feature to be so buried!! Also i’d make it a different color or button type since it’s a link rather than a selection. Maybe just text that is underlined in blue or something?
I’m confused by the add custom indexes button here as well. I will read more in the documentation but I wonder if this belongs here or if it could be added somewhere with some help as well to guide what you can add here. The hint is nice but I don’t think it’s enough for a user to really know what they can add / request here.

Now because i’m using quartz for this there is a weird quirk. When I have several boxes open (the select processing layers box and the insert new spectral index box) PLUS the original tsp interface... I have to figure out what order they are in so I can close them. Some users may think the entire thing is locked when really it’s just too many popups. This is a bit clunky for me to have 3 boxes and have to figure out the proper order! And it locked my computer up a bunch. I’m sure this is quartz specific but any way to control it?

**Download Method**
Next I come to the download method part of the interface. The help icons are very nice!
Spatial Extent
This part of the gui could confuse a user who doesn’t understand what a spatial extent is. In fact this is an area I helped my student with.

**Show Tiles Map:** This isn’t working properly for me. It spun and spun and did finally load but then I couldn’t click on it. It’s thus hard for me to understand how I select the extent UNLESS I have a numeric bounding box. Is there a way to streamline this aspect of the interface so you can create a box like you can on the website? And perhaps add more help buttons

**Reprojection:** A little help popup would be nice here that stats the default proj of MODIS and maybe a big FLAG associated with potential reprojection issues (my personal bias). Reprojection is something that overall so many people don’t understand. And so warning them it changes the data and introduces some error could be nice or just a link to a discussion about this.

**Processing Options**
I wonder if this should be called output file format or something like that. It may just be clearer to the user if you have output format options.

The Create raster stacks option is confusing to me. Does this create an R rasterstack object?
Finally i’m also a bit confused by the HDF file download section. I have 2 format options: ENVI and GTIFF. Yet below I have a save HDF file option. Am I able to select h5 somewhere? Or if this a general folder where the data will be saved? A few help buttons would be very helpful here. If this provides both Hdf and gtiff outputs that would be good to know or turn off if the user only wants gtiff.

**Output**
The output -- boy it was very fast grabbing the data !! I LOVE that I didn’t have to log in :)
What was returned however surprised me. I got what I think are asc files (.dat) with header files and then some hdf files. However I requested geotiff. Then there are R data files. I presume these contain the raster stacks.

I know this is a lot of feedback but overall i really think this is an excellent package! I’d just love for it a bit a bit more user friendly to us mac users :)

Non GUI Interface!

Boy is this nice. Fully automated and reproducible downloads. Here is a bit more nitpicking that you can take or leave! The documentation website is AWESOME!  Thank you.

http://lbusett.github.io/MODIStsp/articles/noninteractive_execution.html

**Vignettes**
https://cran.r-project.org/web/packages/MODIStsp/vignettes/MODIStsp.pdf
Just a note that H4 and H5 can store raster data but are not inherently raster formats

WELL DONE! I hope there is slightly better mac support in a future iteration but overall i was impressed by how well the package worked!
