### 2018-02-12 #184 @lbusett
@lwasser Thanks for your review! I think that your idea of going through the interface before reading the documentation was great: now I have a really useful analysis of a first-time user concerning the GUI implementation!.

I find most of your comments regarding the GUI quite useful. I believe I can implement most of them reasonably easily (though I do not know if I can manage to do something regarding the xquartz interface on mac... sorry... ;-), while some others could be tricky and I have to think about them a bit more.

I will be out of office for a couple of days: when I am back, I'll dig into this a bit more, and probably reply with a list of what of your suggestions I can / plan to implement and which would be a bit of a mess...

Thanks again for a very useful review!

Lorenzo

PS: If you requested GEOTiff and got back .dat (which are ENVI files) there is definitely something strange happening. I'll check to see if I have introduced (another) bug while refactoring!
