### 2018-02-17 #184 @lbusett
Hi all,

  sorry for the late reply, but it was a very busy week.

Thanks again for your favourable and useful reviews. I just wanted to share some quick comments on the reviews before I start implementing changes:

## Review 1 - @jeffreyhanson

> Major issues
> A lot of the examples and vignettes require the user to input their own path to store the data. I would find it really useful if these examples just defaulted to using the temporary directory (i.e. base::tempdir()) so I could run them by copying and pasting code.

Agreed. My "reason" for asking users to explicitly set an output folder was that "novice" users do not know about the existence of the  `R` tempdir, so I feared that having to look for files in tempdir would confuse them. I can however explain how to do that in the examples / vignettes.

> It is my understanding that recent changes to the CRAN policies indicate that example code should write to the default temporary directory. If my understanding is correct, then the MODIStsp_extract example should use base::tempdir() instead of writing to base::system.file("testdata/", package = "MODIStsp"). I have copied the relevant text from the CRAN policies below:

The `MODIStsp_extract` example (and test) use the `base::system.file("testdata/", package = "MODIStsp"` folder only to **read** some preprocessed data from that folder so that it is not needed to download anything from NASA servers to run the example. I know it could be misleading: I will add some comments in the code.

Concerning the new CRAN policy, however, what **I will have to change for sure** is the "automatic" saving of previous processing options in `ExtData/Previous/MODIStsp_previous.json".  For that, I will probably ask permission to the user at first execution.

> The references in vignettes/MODIStsp.pdf are not shown correctly. They contain text like "[-@hengl2010]" instead of the in-text reference.

Thanks for signalling this. The pdf vignette has been a bit neglected since I moved most information on the website. I will see that it is up-to-scratch (and also probably move it to a html version).

> ROpensci packaging guidelines
> This package uses RCurl. From what I can tell, the only usage of RCurl is `RCurl::getURL (line 132 in R/MODIStsp_download.R) so you might be able to drop this dependency without too much hassle.
> This package uses base::print instead of base::message in a couple of places. This only happens in two places (lines 236 & 334 in R/MODIStsp_extract.R) so you might be able replace this with base::message fairly easily.`

Agreed. I do not remember if using RCurl on that call has a reason, or it is simply a leftover of previous versions.

## Review 2 - @lwasser

> Installation
I ran into significant issues installing this on my mac. I still don’t have it working on my laptop but did get it working on my desktop. It would be very nice to have one fully functional set of instructions for mac that is well tested!

Agreed, and sorry for the trouble. I will try to **streamline the installation instructions** for mac (and to check them more thoroghly) before submitting a new version. Concerning GDAL installation, I plan to **add specific instructions for mac  here**: http://lbusett.github.io/MODIStsp/articles/installation.html#installing-r-and-gdal

Can I ask what is the "usual" way to install GDAL with hdf4 support on a mac ? Using `homebrew` ? Do you think that running this would be sufficient?

> brew install hdf4
> brew link --overwrite hdf4
> brew install gdal --complete --enable-unsupported --with-hdf4

## Comments on Interface

Thanks for this very valuable and thorough feedback. Without reporting all your comments in full:

### This is what I am currently planning to do:

- **Improve the Product / Layers selection interface**, by adding the Product Information button also in the main GUI, and finding a way to show info about the currently selected products in the main GUI ( I have a couple of ideas in mind on that : I could share some screenshots for your comments when I implement it)

- **Add help suggestions on the "Add new index" functionality**. I will however not move it in the main interface, because the Add Index functionality is product-specific and you need to be able to see which indexes are already available before adding a new one. Therefore I believe it needs to stay in the "Select Layers" area.

- **See what I can do about the "multiple boxes" issue**. Consider however that that those need to be "modal" windows, so they actively block each other. Only thing I may be able to do is to have the "select layers" window disappear when the "add new window" box is opened.

- **Improve the "spatial extent" selection interface** by adding helper buttons and maybe reshape a bit the interface (i.e., make the possibility to load a spatial file to specify the bounding box more preminent).

- **Modify the "reprojection selection"**. I think that I will leave only two options in the selection dropbox: "Native" and "User Selected" (as it happens for the resolution). This will make clear that if I use "Native" no reprojection will be done. I could also have a "warning" appear if the user specifies a different projection if you think it is needed. Consider however that MODIStsp only allows for nearest neighbour or modal resampling, so the issue concerning modifying data is reduced, IMO.

- **Improve the "processing options" (It will become "Output Format", thanks for the suggestion) and "output folder" selection sections**.  and provide more help. On this:

  - Concerning the **_saving hdfs_**: it allows users to specify a path (different form the "main output" one) to store the **original** MODIS hdf files. This is useful because:
 _i)_ if one needs to process different areas within the same tile(s) he can download and store the HDF files only once, and use those for different processings,
_ii)_ It allows to create an "archive" of HDFs on a remote storage, and get only the processed results locally (for example, we keep all original HDFs on a NAS, while I can get the MODIStsp processed files on my machine),and
_iii)_ it allows users who already have an archive of MODIS HDFs available to process them using MODIStsp without the need of downloading again.
I will **add some help** on this, and maybe **move the selection widget BELOW that of the main output folder.**

  - Concerning the "**_rasterstacks"_**: Yes, you can create rasterStacks files automatically. This is explained in the documentation, but would probably benefit by a dedicated help button. The other option ("Create Virtual Rasters") allows instead to create virtual multitemporal files (i.e., text files pointing to the single-band images processed by MODIStsp) accessible easily from QGIS or ENVI or other image processing software. Again, I'll add some help as suggested.

   - Concerning the **_".dat" files_**: Those are ENVI binary files. I do not remembere WHY I called them ".dat". Probably because back-in-the-day ARCGIS used ".dat" for ENVI files and I was still using it (shame on me... ;-) ) . I could change the extension to ".envi" to make things clearer. What do you think?
 As a sidenote, I tried it today, and **if I select "GeoTiff" outputs, I get geotiff files in the output folder**. Could you check this? Where you could have gotten ".dat" files is in the "Time_series/ENVI_META" subfolder if the "create virtual rasters" selector was set on "ENVI", or "ENVI and GDAL".

  - Concerning in general the **_"output conventions"_**: The "conventions" for output storage are explained in the documentation (in particular, http://lbusett.github.io/MODIStsp/articles/output.html). Is that clear enough or do you think that additional details should be provided?  (e.g., I am thinking about screenshots of a dummy folder structure, though they are ugly... ;-))

### This is what I have to think a bit about:

- Add the possibility to **interactively selecting tiles on a map**. On this:

1) Consider that the current "Show Tiles Map" button does only that. That is, it shows a map to allow seeing which tiles the user should select with the sliders to get data for an area. Sorry if you wasted time "clicking" on it: nothing could have really happened ;-) . I am however surprised that `...It spun and spun and did finally load...`. For me, it loads immediately. Can I ask you to tell me what happens if you run:
```
    library(MODIStp)
   dev.new()
   raster::plot(raster::raster(system.file("ExtData/MODIS_Tiles.gif", package = "MODIStsp")))
```
That is probably a "xquartz" issue that may have a workaround.

2) Allowing to "interactively" select tiles is a good suggestion. (The `MODIS` package does this with the `getTile` function). However, I am a bit reluctant to implement it because:

   - It would require using functionalities provided by `mapedit`, bringing in a lot of dependencies (e.g., sf, shiny, leaflet, dplyr, mapview...)
   - It would require to open the selection "tool" in a browser or within RStudio. I personally think that one nice thing about  the MODIStsp GUI is that it "packs" all the functionality in a compact,  all-in-one interface.

For these reasons, I do not know if I will do something on this besides improving the GUI / help as described above. Let me know what do you think.

### This is what I will not probably be able to do anything about (at least for now):

- Improve the GUI interface for mac / XQuartz.

The `MODIStsp_GUI` function is very complex (and was a mess to develop...). "Reshaping" it for mac would require a lot of time and would be difficlult for me since I do not work on a mac. I am sorry if it is "clunky", but for the time being it will probably need to stay like this.

**However**, I may try to improve it a bit by tweaking default fonts etcetera. Would you mind sharing a couple of screenshots of the interface so that I can seem how/how much is it different from what I see on Linux or Windows (that you can look at here: http://lbusett.github.io/MODIStsp/articles/interactive_execution.html) ?

As an additional note: an obvious way to improve (?) the interface would be to go all the way down the "shiny" path. However, I am not fluent on Shiny and it would require a lot of work. Maybe on next iterations I can consider this.

## Comments on Non GUI Interface / vignettes

> Boy is this nice. Fully automated and reproducible downloads. Here is a bit more nitpicking that you can take or leave!

Glad you like it! Just a double check: I do not find any additional "nitpicking" here ;-) . Is something missing, or you did not have any additional comments?

> Vignettes
https://cran.r-project.org/web/packages/MODIStsp/vignettes/MODIStsp.pdf
Just a note that H4 and H5 can store raster data but are not inherently raster formats

Noted.

===============

Ok, that's it. Longer than I expected but it was also useful for me to clarify the roadmap! Feel free to comment/ask for clarifications.

@karthik (when you come back)

1)  I will probably need 2-3 weeks to finalize this, since I will have to do it mostly off work hours. Is that ok for you?

2) Can I add the "under review" badge on github?

Thanks again for all the feedback. I will keep you posted on advancements on implementation of the changes.

Lorenzo
