### 2018-02-17 lbusett/MODIStsp#112 @lbusett
ropensci review
Roadmap for ropensci review:

## for jeffreyhanson review:

- [x] make so that all examples in vignettes and documentation save in tempdir, and that at first GUI execution the output folders are set to $tempdir If $tempdir$ is found as folder selection the outputs go to tempdir/MODStsp and tempdir/MODIStsp/HDFs

Done in 233df7f86a9825a362dcc8ab9b26ae792725a9a6; 8d612b0ec56f6eea426b1686c771fd5afaa26109; 6567949205b228ed4dbcba8385018ba272cfd09f

- [x] make so that at first MODIStsp execution the user is asked permission to save the options file in /Extdata/Previous

Done in 233df7f86a9825a362dcc8ab9b26ae792725a9a6

- [x] Remove Rcurl dependency

Done in fd687db2082c524e71b24cf44e8db167ed5d27ff. We now use `httr::HEAD` to check file size over ftp

- [x] Change "print" to "message" in MODIStsp_extract

- [x] fix cran warnings about "C:\Progra" etcetera

Done in b2860372164730c6b0fd4f0bca06229ae5a1da1a

- [x] Fix the references in the PDF vignette and website

## for lwasser review:

- [x] Improve installation instructions for mac - package and GDAL

Done in ce72a7cefd6a3a0d5a3a199d35488e1c8c8b5df6 and others. Mainly uses instructions found in
https://zhiyzuo.github.io/installation-rattle/

- [x] Improve the Product / Layers selection interface, by adding the Product Information button also in the main GUI, and finding a way to show info about the currently selected products in the main GUI

Done in f893e4f71d2e7bad482594b5200b9c03ccd35f85 and others. Selected layers are immediately visible in the Main Window + help buttons were added to explain what the Quality Indicators and Spectral Indices are.

- [x] Add help suggestions on the "Add new index" functionality.

Finalized in f893e4f71d2e7bad482594b5200b9c03ccd35f85

- [x]  Make the "select layers" window disappear when the "add new window" box is opened.

Stripped. I see no easy way to do this. I do not believe this is a major issue, though. I can reevaluate if this is considered mandatory.

- [x] Improve the "spatial extent" selection interface  - consider "interactive selection" functionality

Finalized in 50442120dd232f9c31a3daba77e16629d8e59aae - Now it is possible to interactively select the MODIS tiles or custom extent using the "Select from Map" button, and see the currently selected extent in a browser using the "Show Current Extent" button.

- [x] Modify the "reprojection selection" dropbox: remove all default choices besides "Native" and "user defined

Finalized in 50442120dd232f9c31a3daba77e16629d8e59aae.

* Now only "Native" and "User Defined" are possible;
* Input field for custom projection now also allows inserting EPSG codes or UTM codes;
* Warning messages are sent in case a User Projection is selected;
* Changing projection now re-sets the custom extent to avoid problems due to sequential reprojections
of the extent (i.e., continuously enlarging bbox to avoid "losing" area).

- [x] Move the hdf folder selector below the main output folder one and add help buttons on both output folder selectors

Done in f893e4f71d2e7bad482594b5200b9c03ccd35f85 and others

- [x] Modify inputs for virtual files: Use a checkbox group allowing to specify rasterStack, ENVI META or
       GDAL vrt do that we have only one selector

Done in f893e4f71d2e7bad482594b5200b9c03ccd35f85 and others. Now all "virtual files" (i.e., ENVI, R and GDAL) are treated "equally" and can be selected using a single checkbox group.

- [x] add help regarding virtual files outputs

Done in f893e4f71d2e7bad482594b5200b9c03ccd35f85 and others

- [x] change default output format to "GeoTiff" and virtual raster files to "rasterStack"

- [x] ? change extension for envi files to ".envi" + allow for double pattern (.dat and .envi) in the files search to avoid breaking "already made" time series for old users

I decided to strip this to avoid problems with already existing users' time series

- [x]  improve in general the help boxes; add some formatting

Done in f893e4f71d2e7bad482594b5200b9c03ccd35f85 and others
