### 2018-02-18 233df7f
Use tempdir as default for tests and and examples + ask permission
- Tests and examples default to tempdir
- At first execution, ask permission to write to the prev_options.json file
- At first GUI execution, output folders are set to $tempdir
- Modified all test files so that they explicitly point to $tempdir or $testdata
- Modified examples and tests for MODIstp_extract to avoid writing virtual files
outside tempdir
-added "ask_permission" and "load_opts" helpers to shorten main function
-started work on rewording examples (to be checked!)
