### 2018-02-18 8d612b0
corrected paths for tests
