+++ tests/testthat/test_MODIStsp.R
@@ -1,7 +1,7 @@
 context("MODIStsp Test 0: Gracefully fail on input problems")
 testthat::test_that(
   "Tests on MODIStsp", {
-
+    
     library(testthat)
     # no options file
     expect_error(MODIStsp(gui = FALSE),
@@ -14,21 +14,21 @@ testthat::test_that(
       options_file = system.file("ExtData", "MODIStsp_ProdOpts.xml",
                                  package = "MODIStsp"),
       gui = FALSE), "Unable to read the provided JSON")
-
+    
     # Credentials for earthdata login for http download are wrong
     expect_error(MODIStsp(
       options_file = system.file("testdata/test05_wrong_pwd.json",
                                  package = "MODIStsp"),
       gui = FALSE, n_retries = 2), "Username and/or password are not valid")
-
+    
     # Try to access via ftp a product only available over http fails
     # gracefully
     expect_error(MODIStsp(
       options_file = system.file("testdata/test04_ftp.json",
                                  package = "MODIStsp"),
       gui = FALSE, n_retries = 2), "Please switch to http download")
-
-
+    
+    
   })
 
 
@@ -36,32 +36,32 @@ context("MODIStsp Test 1: Basic processing on bands and quality
         indicators")
 testthat::test_that(
   "Tests on MODIStsp", {
-
+    
     library(testthat)
-
+    
     # skip("Skip tests - since they rely on download they are only run locally")
     # skip_on_cran()
     # skip_on_travis()
-
+    
     ### Test 1: test of the basic operations of MODIStsp.                   ####
     #   The test process two bands and extracts one quality indicator from a
     #   single local hdf file for MOD11A2 product  without any
     #   additional preprocessing operations. Output files are in GeoTiff format.
-
+    
     MODIStsp(test = 1)
-    out_files  <- list.files(file.path(tempdir(),
+    out_files  <- list.files(file.path(tempdir(), "MODIStsp", 
                                        "Surf_Temp_8Days_GridSin_v6"),
                              pattern = ".tif$",
                              recursive = TRUE, full.names = TRUE)
     file_sizes <- file.info(out_files)$size
-
+    
     # check that size of files file resulting from test run are equal to those
     #  obtained with a "working" MODIStsp version
     expect_equal(file_sizes, c(80670, 80670, 40916, 40916))
-
+    
     # check that median value of files file resulting from test run are
     # equal to those obtained with a "working" MODIStsp version
-
+    
     means <- unlist(
       lapply(out_files,
              FUN = function(x) {
@@ -82,12 +82,12 @@ testthat::test_that(
   "Tests on MODIStsp", {
     # skip_on_cran()
     # skip_on_travis()
-
+    
     MODIStsp(test = 2)
     out_files_dat  <- list.files(file.path(
-      tempdir(), "Surf_Temp_8Days_GridSin_v6"),
+      tempdir(), "MODIStsp","Surf_Temp_8Days_GridSin_v6"),
       pattern = ".dat$", recursive = TRUE, full.names = TRUE)
-
+    
     # same checks as before on file size and raster stats
     file_sizes_dat <- file.info(out_files_dat)$size
     expect_equal(file_sizes_dat, c(52000, 26000))
@@ -99,7 +99,7 @@ testthat::test_that(
     )
     expect_equal(means, c(13447.650685, 1.757238),
[…]
