### 2018-02-20 6567949
correct typo on options file (possibly) crashing tests on travis
