+++ R/MODIStsp.R
@@ -61,16 +61,16 @@
 #'
 #' # Running the tool using the settings previously saved in a specific options
 #' # file
-#' 
+#'
 #' # **NOTE** Output files of examples will be saved to file.path(tempdir(), "MODIStsp").
 #' # You can run the example with `gui = TRUE` to set a different output folder!
-#' 
+#'
 #' \dontrun{
-#' 
-#' # Here we use a test json file saved in MODIStsp installation folder - 
+#'
+#' # Here we use a test json file saved in MODIStsp installation folder -
 #' # Downloads and processed 3 MOD13A2 images over the Como Lake (Lombardy, Italy)
 #' # and retrieves NDVI and EVI data, plus the Usefulness Index Quality Indicator.
-#'  
+#'
 #' options_file <- system.file("testdata/MOD13A2.json", package = "MODIStsp")
 #' MODIStsp(gui = FALSE, options_file = options_file)}
 #'
@@ -89,19 +89,19 @@
 #' # spatial file (e.g., to perform the same processing on several extents)
 #'
 #' \dontrun{
-#' extent_list  <- c(system.file("testdata/gardalake.shp", package = "MODIStsp"), 
+#' extent_list  <- c(system.file("testdata/gardalake.shp", package = "MODIStsp"),
 #'                   system.file("testdata/iseolake.shp", package = "MODIStsp"))
-#' 
-#' # Note that you can also put all your extent files in a specific folder and 
-#' # create the extent list using 
+#'
+#' # Note that you can also put all your extent files in a specific folder and
+#' # create the extent list using
 #' # extent_list = list.files("X:/path/containing/some/shapefiles/", "\\.shp$")
-#' 
+#'
 #' options_file <- system.file("testdata/MOD13A2.json", package = "MODIStsp")
 #' for (single_shape in extent_list) {
 #'   MODIStsp(gui = FALSE, options_file = options_file,
 #'            spatial_file_path = single_shape )
 #' }
-#' 
+#'
 #' }
 #'
 #' # See also http://lbusett.github.io/MODIStsp/articles/noninteractive_execution.html)
@@ -113,20 +113,20 @@ MODIStsp <- function(gui               = TRUE,
                      test              = -1,
                      n_retries         = 20,
                      verbose           = TRUE) {
-  
+
   options("guiToolkit" = "RGtk2")
   # Make so that "raster" functions does not automatically add extensions on
   # output files. This is automatically reset to TRUE at the end of the session
   raster::rasterOptions(setfileext = FALSE)
-  
+
   #   __________________________________________________________________________
   #   Initialize processing                                                 ####
-  
+
   #   __________________________________________________________________________
   #   If test mode is selected, select the options file for testing         ####
   #   and set other parameters
   if (test > -1) {
-    
+
     gui <- FALSE
     message("MODIStsp is running in test mode.")
     # read names of available json test files
@@ -154,19 +154,19 @@ MODIStsp <- function(gui               = TRUE,
               unzip   = "internal")
       }
     }
-    
+
     # Assign the selected test Option File
     options_file <- test_files[test]
     start <- TRUE
   }
-  
+
   #   __________________________________________________________________________
   #   On interactive execution, ensure that gWidgetsRGtk2 is avauilable     ####
-  
+
   if (gui) {
     #nocov start
     if (!pacman::p_exists("gWidgetsRGtk2", local = TRUE)) {
-      
+
       message(strwrap("Library 'gWidgetsRGtk2' is not installed. It is required
                       to run MODIStsp in interactive mode!\n\n",
                       "Do you want to install it now?"),
@@ -178,15 +178,15 @@ MODIStsp <- function(gui               = TRUE,
[…]
