### 2018-02-20 fd687db
Removed RCurl dependency
- Changed to to httr::HEAD()
- Also suppressed warnings related to httr traversing the ftp structure
