+++ NAMESPACE
@@ -5,7 +5,6 @@ export(MODIStsp_addindex)
 export(MODIStsp_extract)
 export(MODIStsp_resetindexes)
 export(install_MODIStsp_launcher)
-importFrom(RCurl,getURL)
 importFrom(bitops,bitAnd)
 importFrom(bitops,bitShiftR)
 importFrom(data.table,data.table)
@@ -47,8 +46,10 @@ importFrom(gdalUtils,gdalsrsinfo)
 importFrom(gdalUtils,gdalwarp)
 importFrom(grDevices,dev.new)
 importFrom(httr,GET)
+importFrom(httr,HEAD)
 importFrom(httr,RETRY)
 importFrom(httr,authenticate)
+importFrom(httr,config)
 importFrom(httr,content)
 importFrom(httr,progress)
 importFrom(httr,timeout)
