### 2018-02-20 lbusett/MODIStsp#113 @cobanton
MODIStsp_extract function does not account for poor quality pixels
I am trying to use MODIStsp_extract to get Snow Cover averages for a grid Study Area Shapefile. I used the processing GUI to download MODIS Snow Cover files. What I am seeing is that the "mean" function treats values above 100 as part of the calculation which are not true snow cover values but refer to quality assessments like "missing data", "cloud", "no decision", etc. I have tried manually coercing all values above 100 to NA's so that the na.rm =T argument would cut them out, but the extract function fails if I try to do this. Is there a way to change the >100 values to something to make the extraction work or are there fixes that can be put into the function to account for this?

Thank you
Colby
