### 2018-02-20 lbusett/MODIStsp#113 @lbusett
Hi,

   two quick things:

1) could you please tell us which specific MODIS product you are using ? Does it have multiple fill-values?

2) forcing values to NA should work. Would you mind sharing the code you are using ?
