### 2018-02-20 lbusett/MODIStsp#113 @cobanton
Thanks for the quick reply!
I am using the MODIS/Terra Snow Cover Daily L3 Global 500m Grid V5. I am unsure about the multiple fill values, still pretty new to working with raster stacks.
Here is what I have been doing:

#Loads entire raststack for all 4 years
load("I:/MODIS/time series/Snow_Cov_Daily_500m_v5/Time_Series/RData/MOD10_A1_SC_352_2013_90_2017_RData.RData")
#Change all values above 100 to NA
raster_ts[raster_ts>100]<- NA
#values(raster_ts)[values(raster_ts)>100]=NA

I've tried the 2 different options above to change NA values, both give similar outputs. It seems like it is changing the "Z" attribute as well, ie date.

Thanks for the help!
