### 2018-02-22 82a2264
WIP on GUI improvements
partial work on showing info on selected layers
