+++ R/load_prodopts.R
@@ -1,11 +1,11 @@
-#' @title Load characteristics of the different MODIS products 
+#' @title Load characteristics of the different MODIS products
 #' @description FUNCTION_DESCRIPTION
 #' @param gui `logical` if TRUE, the function was called from an interactive
-#'  MODIStsp session. 
+#'  MODIStsp session.
 #' @return OUTPUT_DESCRIPTION
 #' @details Load characteristics of the different MODIS products from `prodopts_file`
 #   or load them from the XML options file and create the `prodopts_file` RData
-#   file if the `prodopts_file` got lost or is outdated (this because reading 
+#   file if the `prodopts_file` got lost or is outdated (this because reading
 #   from RData is much faster, but the XML allows for easier maintenance and
 #   update of the MODIS products descriptions)
 #' @rdname load_prodopts
@@ -22,6 +22,7 @@ load_prodopts <- function(gui) {
   # XML file describing MODIS products
   xml_file <- system.file("ExtData", "MODIStsp_ProdOpts.xml",
                           package = "MODIStsp")
+
   if (file.exists(prodopts_file)) {
     prod_opt_list <- get(load(prodopts_file))
     if (is.null(attr(prod_opt_list, "MODIStspVersion"))) {
@@ -42,18 +43,18 @@ load_prodopts <- function(gui) {
       mess     <- gWidgets::gwindow(title  = "Please wait...",
                                     width  = 400,
                                     height = 40)
-      
+
       mess_lab <- gWidgets::glabel(text      = mess_text,
                                    editable  = FALSE,
                                    container = mess)
       Sys.sleep(0.05)
       #nocov end
     }
-    
+
     MODIStsp_read_xml(prodopts_file = prodopts_file,
                       xml_file      = xml_file)
     prod_opt_list <- get(load(prodopts_file))
-    
+
     if (gui) {
       #nocov start
       # dispose message window
