+++ R/MODIStsp_process.R
@@ -47,7 +47,7 @@
 #' @param full_ext `characrter ["Full_Ext" | "Resized"]` If \"Full_Ext\", process
 #'   the entire extent of the selected tiles. Otherwise, crop the output to
 #'   output bbox.
-#' @param bbox `numeric(4)` Output bounding box (xmin, xmax, ymin, ymax) in
+#' @param bbox `numeric(4)` Output bounding box (xmin, ymin, xmax, ymax) in
 #'   out_proj coordinate system.
 #' @param out_format `character ["ENVI" | "GTiff"]` Desired output format.
 #' @param compress `character ["None" | "PACKBITS" | "LZW" | "DEFLATE"]`
@@ -166,12 +166,12 @@ MODIStsp_process <- function(sel_prod, start_date, end_date, out_folder,
                              use_aria       = TRUE,
                              download_range = "full",
                              gui            = TRUE,
-                             n_retries, 
+                             n_retries,
                              verbose) {
-  
+
   mess_text <- "MODIStsp --> Starting processing"
   # initialize processing messages in case of interactive execution ----
-  
+
   if (gui) {
     #nocov start
     mess     <- gWidgets::gwindow(title = "Processing Status",
@@ -187,10 +187,10 @@ MODIStsp_process <- function(sel_prod, start_date, end_date, out_folder,
   }
 
   process_message(mess_text, gui, mess_lab, verbose)
-  
+
   #   __________________________________________________________________________
   #   Intialize processing variables                                        ####
-  
+
   # as.integer(NoData) cause NoData ranges (e.g. 249-255) to be
   # suppressed. So, in this cases NoData values will not be recognised. This
   # problem will be solved in future with a cycle on NoData ranges.
@@ -199,40 +199,40 @@ MODIStsp_process <- function(sel_prod, start_date, end_date, out_folder,
   if (any(is.na(as.numeric(nodata_in)))) {
     nodata_in[is.na(as.numeric(nodata_in))] <- "None"
   }
-  
+
   if (any(is.na(as.numeric(quality_nodata_in)))) {
     quality_nodata_in[is.na(as.numeric(quality_nodata_in))] <- "None"
   }
-  
+
   # if NoData change set to no, set out_nodata to nodata_in
   if (nodata_change == "No") {
     nodata_out <- nodata_in
   }
-  
+
   # set-up processing folders ----
-  
+
   # Folder for HDF storage
   dir.create(out_folder_mod, recursive = TRUE, showWarnings = FALSE)
-  
+
   # main output folder --> subfolder of "out_folder" named after the selected
   # MODIS product
   out_prod_folder <- file.path(out_folder, main_out_folder)
   dir.create(out_prod_folder, showWarnings = FALSE, recursive = TRUE)
-  
+
   # get start/end years from start_date/end_date
   start_year <- unlist(strsplit(start_date, "[.]"))[1]
   end_year   <- unlist(strsplit(end_date, "[.]"))[1]
-  
+
   # workaround to avoid generating error if no indexes and/or quality bands are
   # present/computable for the selected product
-  
+
   if (length(indexes_bandnames) == 0) indexes_bandsel <- integer(0)
   if (length(quality_bandnames) == 0) quality_bandsel <- integer(0)
-  
+
   # Save original choice of bands in bandsel_orig_choice (bandsel is later
   # modified to set to 1 all bands needed for indexes and quality
   bandsel_orig_choice <- bandsel
-  
+
   #  ___________________________________________________________________________
   #  Build a matrix which associates each SI or QI available for the selected
   #  product with the original layers required to compute it
@@ -249,53 +249,53 @@ MODIStsp_process <- function(sel_prod, start_date, end_date, out_folder,
[…]
