+++ R/MODIStsp_GUI.R
@@ -60,8 +60,8 @@ MODIStsp_GUI <- function(general_opts,
   main_win <- gbasicdialog(
     title      = paste0("MODIStsp - v. ", utils::packageVersion("MODIStsp")),
     parent     = NULL,
-    do.buttons = FALSE
-  )
+    do.buttons = FALSE,
+    use.scroll_window = scroll_window)
 
   # frame1 and 2 with expand=FALSE grant that widgets are not "too much
   # expanded", neither horizontally nor vertically
@@ -73,7 +73,7 @@ MODIStsp_GUI <- function(general_opts,
 
   main_frame2 <- ggroup(container  = main_frame1,
                         horizontal = FALSE,
-                        expand     = TRUE,
+                        expand     = FALSE,
                         use.scroll_window = scroll_window)
 
   #   __________________________________________________________________________
@@ -81,7 +81,7 @@ MODIStsp_GUI <- function(general_opts,
 
   main_group  <- ggroup(container  = main_frame2,
                         horizontal = FALSE,
-                        expand     = TRUE)
+                        expand     = FALSE)
 
   if (scroll_window) {
     getToolkitWidget(main_win)$maximize()
@@ -103,13 +103,18 @@ MODIStsp_GUI <- function(general_opts,
   sel_cat       <- mod_prod_cat$cat[match(sel_prod, mod_prod_list)]
   sel_prodopts  <- prod_opt_list[[sel_prod]]
   # set defaults for projection names and corresponding proj4strings
-  out_proj_names <- c("Sinusoidal", "UTM 32N", "Latlon WGS84", "Latlon MODIS",
+  out_proj_names <- c("Native",
                       "User Defined")
+  # browser()
   out_proj_list  <- list(
-    "Sinusoidal"   = "+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs", #nolint
-    "UTM 32N"      = "+init=epsg:32632 +proj=utm +zone=32 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0", #nolint
-    "Latlon WGS84" = "+init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0", #nolint
-    "Latlon MODIS" = "+init=epsg:4008 +proj=longlat +ellps=clrk66 +no_defs",
+    "Native"   = ifelse(
+      prod_opt_list[[sel_prod]][[general_opts$prod_version]][["tiled"]] == "1",
+      "+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs",
+      "+init=epsg:4008 +proj=longlat +ellps=clrk66 +no_defs"),
+    # #nolint
+    # "UTM 32N"      = "+init=epsg:32632 +proj=utm +zone=32 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0", #nolint
+    # "Latlon WGS84" = "+init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0", #nolint
+    # "Latlon MODIS" = "+init=epsg:4008 +proj=longlat +ellps=clrk66 +no_defs",
     "User Defined" = ""
   )
   mod_proj_str <- "+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs" #nolint
@@ -135,14 +140,14 @@ MODIStsp_GUI <- function(general_opts,
   #   __________________________________________________________________________
   #   Initialize Widgets for category selection                             ####
 
[…]
