### 2018-02-22 lbusett/MODIStsp#113 @cobanton
Hey Lorenzo,

Unfortunately, that did not work. The only thing that changed is sometimes it gave an "NA" instead of a value of 250. I tried a slight work around using the code I pasted above where I changed all values above 100 to "NA" and then brought the z attribute to get the extract function to work. But even if I put na.rm = T in the argument, it seems that any polygon in my shapefile that overlaps with  a NA pixel comes out as NA as well.

Am I interpreting the na.rm = T argument correctly?

Thanks for you help
Colby
