### 2018-02-23 lbusett/MODIStsp#113 @lbusett
Strange. I thought that should have fixed the issue. Are you able to share a test image and the shapefile you are using for extraction so that I can check?

Lorenzo
