### 2018-02-23 lbusett/MODIStsp#113 @cobanton
Attached is the shapefile and raster stack in a zipped folder. What would you want a test image of? The output or the code I'm using?

Colby
[For github.zip](https://github.com/lbusett/MODIStsp/files/1752666/For.github.zip)
