### 2018-02-23 lbusett/MODIStsp#113 @lbusett
I'd need a geotiff of one date, if possible. Or you can save your MODIStsp options from the gui and send me that, so that I can replicate your processing.
