### 2018-02-23 lbusett/MODIStsp#113 @cobanton
Enclosed in that zip is a .tiff for one day of snow cover data.
[For github.zip](https://github.com/lbusett/MODIStsp/files/1753066/For.github.zip)
