### 2018-02-24 lbusett/MODIStsp#113 @lbusett
Thanks. I will have a look ASAP (probably on monday).
