### 2018-02-25 f893e4f
Various improvements in the GUI
Interactive selection of extent + help messages + move long event handlers in gui_helpers + updates in timeseries format selection + various small tweaks
