+++ R/MODIStsp_GUI.R
@@ -25,6 +25,7 @@
 #' @importFrom jsonlite fromJSON write_json
 #' @importFrom sp CRS
 #' @importFrom grDevices dev.new
+#' @importFrom data.table rbindlist   
 #' @importFrom utils packageVersion browseURL head tail
 #' @importFrom gWidgets svalue gconfirm gmessage gbasicdialog ggroup
 #'  getToolkitWidget gframe gdroplist enabled size addSpring glabel
@@ -37,12 +38,14 @@ MODIStsp_GUI <- function(general_opts,
                          opts_jsfile,
                          prodopts_file,
                          scroll_window){
+  
+  help_messages <- create_help_messages()
   #   __________________________________________________________________________
   #   NOTE: The function is excluded from coverage reports since it must be ####
   #   used interactively
-
+  
   # nocov start
-
+  
   # create a new env to facilitate values-passing between widgets + set the
   # environment of helpers to that of the function to avoid having to create
   # globals
@@ -50,43 +53,51 @@ MODIStsp_GUI <- function(general_opts,
   gui_env$start <- FALSE
   environment(gui_load_options) <- environment()
   environment(gui_save_options) <- environment()
-  bbox_out      <- NULL
+  environment(gh_childs)        <- environment()
+  environment(gh_load_extent)   <- environment()
+  environment(gh_selectmap)     <- environment()
+  environment(gh_view_extent)   <- environment()
+  environment(gh_selcat)        <- environment()
+  environment(gh_selprod)       <- environment()
+  environment(gh_tiles_from_bbox) <- environment()
+  
+  bbox_out  <- NULL
   #   __________________________________________________________________________
   #   Start Building the GUI                                                ####
   wids <- list()
   # workaround to avoid NOTE
-  compress_dict <- mod_proj_str <- NULL
-
+  compress_dict <- mod_proj_str <- size <- NULL
+  
   main_win <- gbasicdialog(
     title      = paste0("MODIStsp - v. ", utils::packageVersion("MODIStsp")),
     parent     = NULL,
     do.buttons = FALSE,
     use.scroll_window = scroll_window)
-
+  
   # frame1 and 2 with expand=FALSE grant that widgets are not "too much
   # expanded", neither horizontally nor vertically
-
+  
   main_frame1 <- ggroup(container  = main_win,
                         horizontal = TRUE,
                         expand     = FALSE,
                         use.scroll_window = scroll_window)
-
+  
   main_frame2 <- ggroup(container  = main_frame1,
                         horizontal = FALSE,
                         expand     = FALSE,
                         use.scroll_window = scroll_window)
-
+  
   #   __________________________________________________________________________
   #   Initialize Main container: holds all widgets                          ####
-
+  
   main_group  <- ggroup(container  = main_frame2,
                         horizontal = FALSE,
                         expand     = FALSE)
-
+  
   if (scroll_window) {
     getToolkitWidget(main_win)$maximize()
   }
-
+  
   mod_prod_cat <- as.data.frame(
     t(vapply(prod_opt_list, function(x){
       c(x[[1]]$cat01, x[[1]]$cat02)
@@ -95,7 +106,7 @@ MODIStsp_GUI <- function(general_opts,
   )
   names(mod_prod_cat) <- c("cat01", "cat02")
   mod_prod_cat$cat    <- apply(mod_prod_cat, 1, paste, collapse = " - ")
-
+  
   # get the product name selected in the previous options file and find the
   # corresponding category
   mod_prod_list <- names(prod_opt_list)
[…]
