### 2018-02-26 0b7be5b
Fix regex patterns
