### 2018-02-26 698c774
Manage multiple NA values
Allow MODIStsp to manage multiple NA values passed by xml ProdOpts file: If "Change nodata values" is "Yes", consider all nodata values and convert in the standard output nodata value; otherwise, consider only the first nodata value as nodata.
