### 2018-02-26 cbda884
Take the last nodata value instead than the first
