### 2018-02-26 0d7edf7
Test that multiple nodata were correctly changed
