+++ NAMESPACE
@@ -71,6 +71,7 @@ importFrom(raster,plot)
 importFrom(raster,raster)
 importFrom(raster,rasterOptions)
 importFrom(raster,rasterize)
+importFrom(raster,reclassify)
 importFrom(raster,res)
 importFrom(raster,setZ)
 importFrom(raster,stack)
