+++ man/split_nodata_values.Rd
@@ -70,7 +70,7 @@ Notice that the last NODATA value is the only one which is considered if
 \examples{
 MODIStsp:::split_nodata_values(c("255","250,254:255"))
 MODIStsp:::split_nodata_values(c("255","250,254:255"), take_all = FALSE)
-create_nodata_rcl(c("255","255,101:254"), c("255","255"))
+MODIStsp:::create_nodata_rcl(c("255","250,254:255"), c("255","255"))
 }
 \author{
 Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
