### 2018-02-26 b6929a0
Update products with nodata ranges
Change "x-y" to "x:y"
