+++ R/split_nodata_values.R
@@ -41,7 +41,7 @@
 #'  This function interprets these strings and convert them in vectors with 
 #'  single values.
 #'  Notice that the last NODATA value is the only one which is considered if 
-#'  "Change NoData values" was set to "No". 
+#'  'Change NoData values' was set to 'No'. 
 #' @author Luigi Ranghetti, phD (2018) \email{ranghetti.l@@irea.cnr.it}
 #' @examples 
 #' MODIStsp:::split_nodata_values(c("255","250,254:255"))
