### 2018-02-27 5217acf
updates to make the tests pass
