+++ R/split_nodata_values.R
@@ -1,83 +1,82 @@
 #' @title Split NODATA values or create matrix for reclassification
 #' @name split_nodata_values
-#' @description Internal functions: 
+#' @description Internal functions:
 #'  [split_nodata_values] splits the ranges of NODATA saved
 #'  in the xml product file in readable vector;
-#'  [create_nodata_rcl] creates the matrix for the reclassification of NODATA 
+#'  [create_nodata_rcl] creates the matrix for the reclassification of NODATA
 #'  values to be used with [raster::reclassify] function.
-#' @param nodata_in Character vector corresponding to input NODATA values 
+#' @param nodata_in Character vector corresponding to input NODATA values
 #'  as saved in the xml product file (one or more values per band).
-#' @param nodata_out Character vector corresponding to output NODATA values 
+#' @param nodata_out Character vector corresponding to output NODATA values
 #'  as saved in the xml product file (one single value per band).
 #' @param take_all Logical: if TRUE (default), all the NODATA values are considered;
-#'  if FALSE, only the last one is taken. 
+#'  if FALSE, only the last one is taken.
 #'  See "details" for the meaning of this parameter.
-#' @return [split_nodata_values]  returns a list with the same length 
-#'  of `nodata_in` vector, in which each element 
+#' @return [split_nodata_values]  returns a list with the same length
+#'  of `nodata_in` vector, in which each element
 #'  is a vector with all the NODATA values.
 #' @details MODIS products can have more than one NODATA values (sometimes
 #'  with different meanings, e.g. 255 = "fill" and 254 = "detector saturated"
-#'  in [MOD09A1](https://nsidc.org/data/mod10a1) product). By setting 
-#'  "Change NoData values" to "Yes" in the GUI, all the NODATA values are 
-#'  coherced to one single new NODATA value; conversely, setting it to "No"
+#'  in [MOD09A1](https://nsidc.org/data/mod10a1) product). By setting
+#'  "Change NoData values" to "Yes" in the GUI, all the NODATA values are
+#'  coerced to one single new NODATA value; conversely, setting it to "No"
 #'  only one value is assumed to be NODATA.
-#'  The parameter `take_all` is assumed to be used in this way, by using this 
+#'  The parameter `take_all` is assumed to be used in this way, by using this
 #'  function with `take_all = TRUE` with "Change NoData values" = "Yes" and
 #'  `take_all = FALSE` with "Change NoData values" = "No".
-#'  
+#'
 #'  In the xml product file, NODATA ranges are set as:
 #'  - `x` for products with single NODATA values;
 #'  - `x,y,z` for products with a vector of NODATA values;
 #'  - `x:y` for products with a range of NODATA values;
 #'  - `x:y,z` for a combination of NODATA ranges and/or values.
-#'  
-#'  In [split_nodata_values] *NODATA values are assumed to be integer*: 
-#'  this means that intervals are splitted in integer values 
+#'
+#'  In [split_nodata_values] *NODATA values are assumed to be integer*:
+#'  this means that intervals are splitted in integer values
 #'  (e.g. "250:255" becomes "250 251 252 253 254 255").
 #'  Conversely, function [create_nodata_rcl] creates intervals, so it can also
-#'  manage float values (in practice, this should not make difference within 
+#'  manage float values (in practice, this should not make difference within
 #'  MODIS products, since NODATA values are always integer values).
-#'  
[…]
