### 2018-02-27 lbusett/MODIStsp#113 @lbusett
Hi,

  we should have fixed the issue. Could you please test by reinstalling using:

```
devtools::install_github("lbusett/MODIStsp", ref = "feature/multiple_na")
```

and tell us if it solves your problem?
