### 2018-02-27 lbusett/MODIStsp#113 @cobanton
Hey Lorenzo

That did not work for me. It is still counting all of the 250 valued pixels (which corresponds to clouds blocking an accurate snow cover estimate) into an average for the grid in my shape.
