### 2018-02-27 lbusett/MODIStsp#113 @lbusett
Oh, sorry, I forgot one thing: be sure to set the "Change NoData" option to "Yes" in the GUI before trying again.
