### 2018-02-27 lbusett/MODIStsp#113 @lbusett
And another thing: You will have to re-process the rasters. So, delete the old MODIStsp ouput folder.
