### 2018-02-28 082361b
Update the selector of the output projection
Only "Native" (sinusoidal or gographic) and "User defined" are allowed. By selecting "User defined", the user cad add both a proj4string, an EPSG code or an UTM zone number. If the input is not recognised, the selection is switched to the previously working one (native or defined).
