### 2018-02-28 lbusett/MODIStsp#113 @lbusett
Elaborating a bit more: instead of changing the extract function, we modified the processing so that for products with multiple fill-values, such as yours, all fill-values are automatically converted to NA in the output rasters if you set change nodata to yes.
