### 2018-02-28 3fef08e
Accept UTM zones as string (e.g. "32N")
