### 2018-02-28 lbusett/MODIStsp#113 @cobanton
Thanks Lorenzo. I tried re processing the initial files but I am getting an error that I did not receive before.
Any advice on how to get past this?
Processing and Reprojecting Terra SC files for date: 2013_12_10
Error in rgdal::getRasterData(con, offset = offs, region.dim = reg, band = object@data@band) :
  Failure during raster IO
