### 2018-02-28 lbusett/MODIStsp#113 @cobanton
I also get these warnings after I run MODIStsp()
GDAL version in use: 2.2.0

(rsession.exe:4420): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 79: Element 'markup' was closed, but the currently open element is 'span'

(rsession.exe:4420): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 79: Element 'markup' was closed, but the currently open element is 'span'

(rsession.exe:4420): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 73: Element 'markup' was closed, but the currently open element is 'span'

(rsession.exe:4420): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 73: Element 'markup' was closed, but the currently open element is 'span'

(rsession.exe:4420): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 79: Element 'markup' was closed, but the currently open element is 'span'

(rsession.exe:4420): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 79: Element 'markup' was closed, but the currently open element is 'span'

(rsession.exe:4420): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 74: Element 'markup' was closed, but the currently open element is 'span'

(rsession.exe:4420): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 74: Element 'markup' was closed, but the currently open element is 'span'

(rsession.exe:4420): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 74: Element 'markup' was closed, but the currently open element is 'span'

(rsession.exe:4420): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 74: Element 'markup' was closed, but the currently open element is 'span'

(rsession.exe:4420): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 74: Element 'markup' was closed, but the currently open element is 'span'

(rsession.exe:4420): Gtk-WARNING **: Failed to set text from markup due to error parsing markup: Error on line 1 char 74: Element 'markup' was closed, but the currently open element is 'span'
