### 2018-02-28 lbusett/MODIStsp#113 @lbusett
Ok.

    forget about the Gtk-warning for the moment.
Concerning the rgdal error, I never saw anything similar.... do you have it "immediately" on the first date you are trying to process?

Also, could you share your session info:

`devtools::session_info()`

and also save the JSON options from the GUI and send it to me?

Lorenzo
