### 2018-02-28 lbusett/MODIStsp#113 @cobanton
Yes it comes up on the first date. I can't seem to find anything similar in any forums thus far.

Here is the session info:
setting  value
 version  R version 3.4.2 (2017-09-28)
 system   x86_64, mingw32
 ui       RStudio (1.1.383)
 language (EN)
 collate  English_United States.1252
 tz       America/Los_Angeles
 date     2018-02-28

Packages -------------------------------
 package       * version    date
 base          * 3.4.2      2017-09-28
 bitops          1.0-6      2013-08-17
 cairoDevice   * 2.24       2017-01-06
 codetools       0.2-15     2016-10-05
 colorspace      1.3-2      2016-12-14
 compiler        3.4.2      2017-09-28
 curl            3.0        2017-10-06
 data.table      1.10.4-3   2017-10-27
 datasets      * 3.4.2      2017-09-28
 devtools        1.13.3     2017-08-02
 digest          0.6.12     2017-01-27
 foreach         1.4.3      2015-10-13
 foreign         0.8-69     2017-06-22
 gdalUtils       2.0.1.7    2015-10-10
 ggplot2       * 2.2.1      2016-12-30
 graphics      * 3.4.2      2017-09-28
 grDevices     * 3.4.2      2017-09-28
 grid            3.4.2      2017-09-28
 gtable          0.2.0      2016-02-26
 gWidgets      * 0.0-54     2014-08-28
 gWidgetsRGtk2 * 0.0-85     2018-01-04
 hexbin          1.27.2     2018-01-15
 httr            1.3.1      2017-08-20
 iterators       1.0.8      2015-10-13
 jsonlite        1.5        2017-06-01
 lattice       * 0.20-35    2017-03-25
 latticeExtra  * 0.6-28     2016-02-09
 lazyeval        0.2.0      2016-06-12
 magrittr        1.5        2014-11-22
 maptools      * 0.9-2      2017-03-25
 memoise         1.1.0      2017-04-21
 methods       * 3.4.2      2017-09-28
 MODIStsp      * 1.3.3.9000 2018-02-27
 munsell         0.4.3      2016-02-13
 pacman          0.4.6      2017-05-14
 parallel        3.4.2      2017-09-28
 plyr            1.8.4      2016-06-08
 R.methodsS3     1.7.1      2016-02-16
 R.oo            1.21.0     2016-11-01
 R.utils         2.5.0      2016-11-07
 R6              2.2.2      2017-06-17
 raster        * 2.6-7      2017-11-13
 rasterVis     * 0.43       2018-01-10
 RColorBrewer  * 1.1-2      2014-12-07
 Rcpp            0.12.13    2017-09-28
 RCurl           1.95-4.10  2018-01-04
 rgdal         * 1.2-16     2017-11-21
 rgeos           0.3-26     2017-10-31
 RGtk2         * 2.20.33    2017-05-10
 rlang           0.1.2      2017-08-09
 scales          0.5.0      2017-08-24
 sp            * 1.2-7      2018-01-19
 stats         * 3.4.2      2017-09-28
 stringi         1.1.6      2017-11-17
 stringr         1.3.0      2018-02-19
 tibble          1.3.4      2017-08-22
 tools           3.4.2      2017-09-28
 utils         * 3.4.2      2017-09-28
 viridisLite     0.3.0      2018-02-01
 withr           2.0.0      2017-07-28
 xml2            1.2.0      2018-01-24
 xts             0.10-1     2017-12-20
 zoo             1.8-0      2017-04-12
 source
 local
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 CRAN (R 3.4.2)
 CRAN (R 3.4.3)
 local
 CRAN (R 3.4.2)
 CRAN (R 3.4.3)
 local
 CRAN (R 3.4.2)
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 CRAN (R 3.4.2)
 CRAN (R 3.4.0)
 CRAN (R 3.4.3)
 local
 local
 local
 CRAN (R 3.4.3)
 CRAN (R 3.4.0)
 CRAN (R 3.4.3)
 CRAN (R 3.4.3)
 CRAN (R 3.4.2)
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 CRAN (R 3.4.2)
 CRAN (R 3.4.3)
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 local
 Github (lbusett/MODIStsp@5217acf)
 CRAN (R 3.4.3)
 CRAN (R 3.4.0)
 local
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 CRAN (R 3.4.2)
 CRAN (R 3.4.3)
 CRAN (R 3.4.3)
 CRAN (R 3.4.1)
 CRAN (R 3.4.2)
 CRAN (R 3.4.3)
 CRAN (R 3.4.3)
 CRAN (R 3.4.3)
 CRAN (R 3.4.1)
 CRAN (R 3.4.2)
 CRAN (R 3.4.3)
 CRAN (R 3.4.3)
 local
 CRAN (R 3.4.2)
 CRAN (R 3.4.3)
 CRAN (R 3.4.2)
 local
 local
 CRAN (R 3.4.3)
 CRAN (R 3.4.2)
 CRAN (R 3.4.3)
 CRAN (R 3.4.3)
 CRAN (R 3.4.0)

I cannot get the GUI to save a json file so i attached a screenshot of my GUI.
