### 2018-02-28 lbusett/MODIStsp#113 @lbusett
Ok. I'll look into it. One thing : could you try to remove the package completely before reinstalling from github? (use remove.packages("MODIStsp")

Also, can you please tell me what problem you have while attempting to save the Json?
