### 2018-02-28 lbusett/MODIStsp#113 @cobanton
I removed and reinstalled, but getting a different error now when do any function:
Error in fetch(key) :
  lazy-load database 'C:/Program Files/R/R-3.4.2/library/MODIStsp/help/MODIStsp.rdb' is corrupt

For JSON file, when I press the "Save options" button I can pick a folder and a name (with a .json extension) and press save. But then the file is not there. I've tried different folders and different names and not putting the .json extension, but it does not seem to save it ever.
