### 2018-02-28 lbusett/MODIStsp#113 @lbusett
Exit and reenter rstudio
