### 2018-02-28 lbusett/MODIStsp#113 @cobanton
The save problem gives an error now:
Error: lexical error: invalid char in json text.
                                       I:\MODIS\gui.json
                     (right here) ------^
