### 2018-02-28 lbusett/MODIStsp#113 @lbusett
Oks. working on this. You were a bit unlucky: we are in the process of a major release, and apparently I did a mess with the github version.... Sorry for the trouble.

I'll keep you posted.
