### 2018-02-28 9df6ecc
Various updates in functions and testdata
All to accomodate the changes in the GUI, which led to changes in the values of some "selector variables" (e.g., ts_format, compression, etc.). Seem also to solve the GTk problems
