### 2018-02-28 46e0d84
workaround to have multiple na working on windows
had to resort to saving to tiff
