+++ R/MODIStsp_process_bands.R
@@ -219,6 +219,13 @@ MODIStsp_process_bands <- function(out_folder_mod, modislist,
       )
     }
 
+    # Ugly workaround to avoid crash on Windows - I do not understand why, but
+    # using vrts fails in this case so we have to save a temporary tif
+    if (Sys.info()['sysname'] == "Windows") {
+      temptif <- tempfile(fileext = ".tif")
+      gdal_translate(outfile_vrt_prev, temptif, verbose = F)
+      outfile_vrt_prev <- temptif
+    }
     recl_in <- stack(outfile_vrt_prev)
     recl_out <- raster::reclassify(
       recl_in, 
