### 2018-03-01 7a69bec
Correct bug on extract function + small tweaks
