+++ R/gui_helpers.R
@@ -13,7 +13,7 @@ gui_update_bboxlabels <- function(bbox_out,
                                   units,
                                   wids) {
   #nocov start
-  
+
   digits <- ifelse(units == "dec. degrees", 4, 1)
   gWidgets::svalue(wids$output_ul_east)  <- formatC(bbox_out[1, 1],
                                                     digits = digits,
@@ -42,12 +42,12 @@ gui_update_tiles <- function(bbox_out,
                              modis_grid,
                              wids) {
   #nocov start
-  
+
   bbox_mod  <- reproj_bbox(bbox_out,
                            gWidgets::svalue(wids$output_proj4),
                            mod_proj_str,
                            enlarge = TRUE)
-  
+
   d_bbox_mod_tiled    <- raster::crop(modis_grid, raster::extent(bbox_mod))
   gWidgets::svalue(wids$start_x) <- min(d_bbox_mod_tiled$H)
   gWidgets::svalue(wids$end_x)   <- max(d_bbox_mod_tiled$H)
@@ -103,20 +103,20 @@ gui_load_options <- function(opts_jsfile,
                              wids,
                              prod_opt_list) {
   #nocov start
-  
+
   # load file and reset all widgets to values found in the loaded file
   general_opts <- try(jsonlite::fromJSON(opts_jsfile), silent = TRUE)
-  
+
   # stop on error
   if (class(general_opts) == "try-error") {
     stop("Unable to read the provided JSON options file. Please check your ",
          "inputs!")
   }
-  
+
   sel_prod      <- general_opts$sel_prod
   sel_prodopts  <- prod_opt_list[[sel_prod]]
   cur_prodopts <- sel_prodopts[[general_opts$prod_version]]
-  
+
   gWidgets::svalue(wids$cat) <- paste(
     prod_opt_list[[general_opts$sel_prod]][[general_opts$prod_version]]$cat01,
     prod_opt_list[[general_opts$sel_prod]][[general_opts$prod_version]]$cat02,
@@ -125,52 +125,52 @@ gui_load_options <- function(opts_jsfile,
   gWidgets::svalue(wids$prod) <- general_opts$sel_prod
   gWidgets::svalue(wids$vers) <- general_opts$prod_version
   gWidgets::svalue(wids$sens) <- general_opts$sensor
-  
+
   # set dummy variables holding the initial values of selected bands
   gui_env$temp_wid_bands         <- general_opts$bandsel
   gui_env$temp_wid_bands_indexes <- general_opts$indexes_bandsel
   gui_env$temp_wid_bands_quality <- general_opts$quality_bandsel
-  
-  
+
+
   curr_sel_layers <- paste(
     cur_prodopts[["band_fullnames"]][which(general_opts[["bandsel"]] == 1)],
     collapse = "; ")
   gWidgets::svalue(wids$sel_layers) <- curr_sel_layers
-  
+
   curr_sel_qual <- paste(
     cur_prodopts[["quality_fullnames"]][which(general_opts[["quality_bandsel"]] == 1)],
     collapse = "; ")
   gWidgets::svalue(wids$sel_qi) <- curr_sel_qual
-  
+
   check_names_indexes <- c(
     cur_prodopts[["indexes_fullnames"]],
     as.list(general_opts$custom_indexes[[sel_prod]]
             [[gWidgets::svalue(wids$vers)]])$indexes_fullnames
   )
-  
+
   curr_sel_si <- paste(
     check_names_indexes[which(general_opts[["indexes_bandsel"]] == 1)],
     collapse = "; ")
   gWidgets::svalue(wids$sel_si) <- curr_sel_si
-  
+
   gWidgets::svalue(wids$server)   <- general_opts$download_server
[…]
