+++ NAMESPACE
@@ -8,6 +8,7 @@ export(MODIStsp)
 export(MODIStsp_addindex)
 export(MODIStsp_extract)
 export(MODIStsp_resetindexes)
+export(check_proj4string)
 export(install_MODIStsp_launcher)
 importFrom(bitops,bitAnd)
 importFrom(bitops,bitShiftR)
