+++ DESCRIPTION
@@ -49,11 +49,11 @@ Suggests:
     png,
     grid,
     httptest,
-    mapedit (>= 0.3.2), 
-    mapview (>= 2.3.0), 
-    shiny, 
+    sf(>= 0.6-1),
     leaflet,
-    sf(>= 0.6-1)
+    shiny, 
+    mapview (>= 2.3.0), 
+    mapedit (>= 0.3.2)
 SystemRequirements: Cairo >= 1.0.0, ATK (>= 1.10.0), Pango (>= 1.10.0), GTK+ (>=
     2.8.0), GLib (>= 2.8.0), Curl, GDAL (>= 1.6.3), PROJ.4 (>= 4.4.9)
 URL: https://github.com/lbusett/MODIStsp
