### 2018-03-01 03ce4c2
Reproject extent on call to view current extent
so that it is shown correctly
