### 2018-03-01 2d1fa69
Added multiple maps in view extent
