### 2018-03-01 lbusett/MODIStsp#113 @lbusett
Hi. This time, it should really be fixed! The rgdal problem we had yesterday was a Windows-specific issue that is now solved.

Can you try:

1) removing MODIStsp again, from a "clean" R session
2) exit and re-enter RStudio;
3) reinstall using:

```
devtools::install_github("lbusett/MODIStsp", ref = "feature/test_navalues")
```

I tested this on my machine and your data, and it should solve your problem.

You will also be our first tester for some new features related to the GUI : We tried to improve the interface so that it is easier to use. For example, if you install also "mapedit" you will be able to select the extent interactively from a map. Any feedback is appreciated!

Lorenzo

PS: Sorry for the mess in these days. I hope that now the issue is really solved.
