### 2018-03-01 lbusett/MODIStsp#113 @cobanton
Hey Lorenzo

That new map function is very cool! I'm having an issue downloading though. Its telling me No available data for year: 2013 for Sensor Terra in selected dates. Its saying this for all 4 years I'm looking for even though they downloaded previously. I tried ftp and http, both did not work. The ftp error said the files were not available for ftp download.  Do I need to change any of the GUI options?
