### 2018-03-01 lbusett/MODIStsp#113 @lbusett
strange, I can download MOD10A1 no problem from http. They are indeed not available over ftp.

Can you try running:

`  MODIStsp(test = 4)`

and also running processing without changing projection?

BTW: Now the JSON saving option should work. Can you send me your JSon?
