### 2018-03-01 lbusett/MODIStsp#113 @cobanton
I ran the MODIStsp(test=4) and it worked properly with no errors. But now
when I try to do the MODIStsp() the GUI does not show up and I get this:
MODIStsp()
GDAL version in use: 2.2.0

(rsession.exe:5352): Gtk-WARNING **: Failed to set text from markup due to
error parsing markup: Error on line 1 char 87: Element 'markup' was closed,
but the currently open element is 'span'

(rsession.exe:5352): Gtk-WARNING **: Failed to set text from markup due to
error parsing markup: Error on line 1 char 87: Element 'markup' was closed,
but the currently open element is 'span'

(rsession.exe:5352): Gtk-CRITICAL **: gtk_combo_box_set_active: assertion
`index_ >= -1' failed

(rsession.exe:5352): Gtk-CRITICAL **: gtk_text_buffer_set_text: assertion
`text != NULL' failed
Error in if (gWidgets::svalue(wids$proj) == "User Defined") { :
  argument is of length zero

The GUI window does not show up at all anymore.

On Thu, Mar 1, 2018 at 11:08 AM, Lorenzo Busetto <notifications@github.com>
wrote:

> strange, I can download MOD10A1 no problem from http. They are indeed not
> available over ftp.
>
> Can you try running:
>
> MODIStsp(test = 4)
>
> and also running processing without changing projection?
>
> BTW: Now the JSON saving option should work. Can you send me your JSon?
>
