### 2018-03-01 lbusett/MODIStsp#113 @lbusett
sh*t. I get the same after running the test....

     Must have an error on saving the previous options...
