### 2018-03-01 lbusett/MODIStsp#113 @lbusett
Ok . If you run this:

`file.remove(system.file("ExtData/Previous/MODIStsp_Previous.json", package = "MODIStsp"))`

it should work again. I had some "old" settings in the test files. So do not run other "tests" until I fix so.

However, you told me that the previous test finished correctly, right ? The HDFs were downloaded and all?
