### 2018-03-01 lbusett/MODIStsp#113 @cobanton
Am I doing the bounding box coordinates wrong? When I put in the
easting/northing utm's and then change the projection below, the
coordinates change. Should I leave the newly changed values or change them
back?
