### 2018-03-01 lbusett/MODIStsp#113 @lbusett
The files should be in "R" tempdir, in subfolder MODIStsp.

`file.path(tempdir(), "MODIStsp")`

Ah, maybe I get it ! Try FIRST changing the projection and THEN inserting the coordinates!
I should document that aspect better.... If you first insert the coordinates, and then change the projection MODIStsp interprets it as if you inserted the coordinates in the original MODIS sinusoidal, and then try to convert them (in your case) in UTM12
