### 2018-03-01 lbusett/MODIStsp#113 @cobanton
That makes perfect sense. Its running now perfectly! It will take some time
to download all the files, but I will run the Extract function as soon as
files are available and let you know.

Many thanks for all of your help!

On Thu, Mar 1, 2018 at 12:40 PM, Lorenzo Busetto <notifications@github.com>
wrote:

> The files should be in "R" tempdir, in subfolder MODIStsp.
>
> file.path(tempdir(), "MODIStsp")
>
> Ah, maybe I get it ! Try FIRST changing the projection and THEN inserting
> the coordinates!
> I should document that aspect better.... If you first insert the
> coordinates, and then change the projection MODIStsp interprets it as if
> you inserted the coordinates in the original MODIS sinusoidal, and then try
> to convert them (in your case) in UTM12
>
