### 2018-03-01 lbusett/MODIStsp#113 @cobanton
Just tried that out, looks like it worked.  THANK YOU!

On Thu, Mar 1, 2018 at 1:13 PM, Lorenzo Busetto <notifications@github.com>
wrote:

> Well, it makes sense when you know it.... ;-)
>
> Glad we hopefully solved it. If you want to try if now MODIStsp_extract
> works, try opening one of the TIFF files in QGIS, and see if the NODATA
> areas are shown "transparent". If it is so, then also the extract function
> should skip those areas.
>
