### 2018-03-01 e27aea0
WIP on trying to have something useful on reprojections
Proably the only sensinble thing will be to automatically remove the current selected area whenever the user changes projection
