+++ R/gui_helpers.R
@@ -14,19 +14,24 @@ gui_update_bboxlabels <- function(bbox_out,
                                   wids) {
   #nocov start
 
-  digits <- ifelse(units == "dec. degrees", 4, 1)
-  gWidgets::svalue(wids$output_ul_east)  <- formatC(bbox_out[1, 1],
+  digits <- ifelse(units == "dec.degrees", 4, 1)
+  
+  gWidgets::svalue(wids$output_xmin)  <- formatC(bbox_out[1, 1],
                                                     digits = digits,
                                                     format = "f")
-  gWidgets::svalue(wids$output_ul_north) <- formatC(bbox_out[2, 2],
+  
+  gWidgets::svalue(wids$output_ymin) <- formatC(bbox_out[2, 1],
+                                                digits = digits,
+                                                format = "f")
+  gWidgets::svalue(wids$output_xmax)  <- formatC(bbox_out[1, 2],
+                                                 digits = digits,
+                                                 format = "f")
+  
+  gWidgets::svalue(wids$output_ymax) <- formatC(bbox_out[2, 2],
                                                     digits = digits,
                                                     format = "f")
-  gWidgets::svalue(wids$output_lr_east)  <- formatC(bbox_out[1, 2],
-                                                    digits = digits,
-                                                    format = "f")
-  gWidgets::svalue(wids$output_lr_north) <- formatC(bbox_out[2, 1],
-                                                    digits = digits,
-                                                    format = "f")
+
+
   #nocov end
 }
 
@@ -177,10 +182,10 @@ gui_load_options <- function(opts_jsfile,
   gWidgets::svalue(wids$output_res)      <- general_opts$out_res
   gWidgets::svalue(wids$output_resmeth)  <- general_opts$resampling
   gWidgets::svalue(wids$output_ext)      <- general_opts$full_ext
-  gWidgets::svalue(wids$output_ul_east)  <- general_opts$bbox[1]
-  gWidgets::svalue(wids$output_lr_east)  <- general_opts$bbox[3]
-  gWidgets::svalue(wids$output_lr_north) <- general_opts$bbox[2]
-  gWidgets::svalue(wids$output_ul_north) <- general_opts$bbox[4]
+  gWidgets::svalue(wids$output_xmin)  <- general_opts$bbox[1]
+  gWidgets::svalue(wids$output_xmax)  <- general_opts$bbox[3]
+  gWidgets::svalue(wids$output_ymin) <- general_opts$bbox[2]
+  gWidgets::svalue(wids$output_ymax) <- general_opts$bbox[4]
   gWidgets::svalue(wids$reprocess)       <- general_opts$reprocess
   gWidgets::svalue(wids$delete)          <- general_opts$delete_hdf
   gWidgets::svalue(wids$nodata)          <- general_opts$nodata_change
@@ -277,10 +282,10 @@ gui_save_options <- function(general_opts,
   general_opts$out_res     <- gWidgets::svalue(wids$output_res)
   general_opts$resampling  <- gWidgets::svalue(wids$output_resmeth)
   general_opts$full_ext    <- gWidgets::svalue(wids$output_ext)
-  general_opts$bbox        <- c(gWidgets::svalue(wids$output_ul_east),
-                                gWidgets::svalue(wids$output_lr_north),
-                                gWidgets::svalue(wids$output_lr_east),
-                                gWidgets::svalue(wids$output_ul_north))
+  general_opts$bbox        <- c(gWidgets::svalue(wids$output_xmin),
[…]
