### 2018-03-02 lbusett/MODIStsp#113 @lbusett
Looking into this. Would be useful if you could share your options file.
