### 2018-03-02 acce4d8
WIP on dealing with reprojection
