+++ R/gui_helpers.R
@@ -14,22 +14,22 @@ gui_update_bboxlabels <- function(bbox_out,
                                   wids) {
   #nocov start
 
-  digits <- ifelse(units == "dec.degrees", 4, 1)
-  
+  digits <- ifelse(units == "dec.degrees", 5, 1)
+
   gWidgets::svalue(wids$output_xmin)  <- formatC(bbox_out[1, 1],
-                                                    digits = digits,
-                                                    format = "f")
-  
+                                                 digits = digits,
+                                                 format = "f")
+
   gWidgets::svalue(wids$output_ymin) <- formatC(bbox_out[2, 1],
                                                 digits = digits,
                                                 format = "f")
   gWidgets::svalue(wids$output_xmax)  <- formatC(bbox_out[1, 2],
                                                  digits = digits,
                                                  format = "f")
-  
+
   gWidgets::svalue(wids$output_ymax) <- formatC(bbox_out[2, 2],
-                                                    digits = digits,
-                                                    format = "f")
+                                                digits = digits,
+                                                format = "f")
 
 
   #nocov end
@@ -42,14 +42,14 @@ gui_update_bboxlabels <- function(bbox_out,
 #' @importFrom gWidgets svalue
 #' @noRd
 #'
-gui_update_tiles <- function(bbox_out,
+gui_update_tiles <- function(bbox_out, curr_proj,
                              mod_proj_str,
                              modis_grid,
                              wids) {
   #nocov start
 
   bbox_mod  <- reproj_bbox(bbox_out,
-                           gWidgets::svalue(wids$output_proj4),
+                           curr_proj,
                            mod_proj_str,
                            enlarge = TRUE)
 
@@ -66,10 +66,10 @@ gui_update_tiles <- function(bbox_out,
 #' @importFrom utils head tail
 #' @noRd
 #'
-gui_get_proj <- function(sel_output_proj) {
+gui_get_proj <- function(curr_proj) {
   #nocov start
   utils::head(strsplit(utils::tail(
-    strsplit(sel_output_proj@projargs, "+proj=")[[1]], 1), " +")[[1]], 1)
+    strsplit(curr_proj@projargs, "+proj=")[[1]], 1), " +")[[1]], 1)
   #nocov end
 }
 
@@ -79,16 +79,16 @@ gui_get_proj <- function(sel_output_proj) {
 #' @importFrom utils head tail
 #' @noRd
 #'
-gui_get_units <- function(sel_output_proj,
+gui_get_units <- function(curr_proj,
                           proj) {
   #nocov start
   if (proj == "longlat") {
     units <- "dec.degrees"
   } else {
     units <- ifelse(
-      length(strsplit(sel_output_proj@projargs, "+units=")[[1]]) > 1,
+      length(strsplit(curr_proj@projargs, "+units=")[[1]]) > 1,
       utils::head(strsplit(utils::tail(strsplit(
-        sel_output_proj@projargs, "+units=")[[1]], 1), " +")[[1]], 1),
+        curr_proj@projargs, "+units=")[[1]], 1), " +")[[1]], 1),
       "Unknown"
     )
   }
@@ -120,7 +120,7 @@ gui_load_options <- function(opts_jsfile,
 
   sel_prod      <- general_opts$sel_prod
   sel_prodopts  <- prod_opt_list[[sel_prod]]
[…]
