+++ R/load_opts.R
@@ -16,24 +16,24 @@
 #' @importFrom jsonlite fromJSON write_json
 #' @importFrom utils packageVersion
 load_opts <- function(opts_jsfile) {
-  
+
   if (file.exists(opts_jsfile)) {
-    
+
     general_opts <- try(jsonlite::fromJSON(opts_jsfile), silent = TRUE)
-    
+
     # stop on errors
     if (class(general_opts) == "try-error" ) {
       stop(strwrap("Unable to read the provided JSON options file. Please check
                    your inputs!"))
     }
-    
+
     if (!("MODIStspVersion" %in% names(general_opts))) {
       stop(
         strwrap("The specified json file does not appear to be a valid MODIStsp
            options file. Please check your inputs!")
       )
     }
-    
+
     if (is.null(general_opts$MODIStspVersion) |
         general_opts$MODIStspVersion < "1.3.3.9000") {
       stop("The option file in use (", opts_jsfile, ") was created with a ",
@@ -56,7 +56,7 @@ load_opts <- function(opts_jsfile) {
     # "general_opts" structure used to  communicate with the GUI, set default
     # values and save it as a JSON file
 
-    
+
     #nocov start (this is only executed at first ever execution)
     general_opts <- list(
       sel_prod        = "Surf_Ref_8Days_500m (M*D09A1)",
@@ -79,7 +79,7 @@ load_opts <- function(opts_jsfile) {
       download_server = "http",
       download_range  = "full",
       proj            = "Native",
-      user_proj4      = "+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs", #nolint
+      output_proj4    = "+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs", #nolint
       out_res_sel     = "Native",
       out_res         = "",
       full_ext        = "Select MODIS Tiles",
