### 2018-03-02 acb611b
simplified GUI reproj. handlers
also modifie test files according to new var names (output_proj4)
