### 2018-03-04 5044212
Finalized work on improved extent selection
- Using editMap and selectFeatures to allow interactive selection + selectMap to show the current extent
- Removed possibility to reproject an already selected extent. Now, if a extent was selected and the user asks to change projection the previous selection is removed (a message is given!). This avoid problems related to the need to "expand" the bbox each time a reprojection is asked + avoids problems if the user specifies coordinates in the wronf CRS
- Various improvements in GUI event handlers and code refactoring
