+++ R/MODIStsp_GUI.R
@@ -38,17 +38,16 @@ MODIStsp_GUI <- function(general_opts,
                          opts_jsfile,
                          prodopts_file,
                          scroll_window){
-
-  help_messages <- create_help_messages()
+  # nocov start
   #   __________________________________________________________________________
   #   NOTE: The function is excluded from coverage reports since it must be ####
-  #   used interactively
-
-  # nocov start
+  #   used interactively. The same goes for all helper function and event 
+  #   handlers!
+  help_messages <- create_help_messages()  
 
   # create a new env to facilitate values-passing between widgets + set the
   # environment of helpers to that of the function to avoid having to create
-  # globals
+  # globals TODO consider remeoving and passing everything explicitly!
   gui_env       <- new.env()
   gui_env$start <- FALSE
   environment(gui_load_options) <- environment()
@@ -59,45 +58,47 @@ MODIStsp_GUI <- function(general_opts,
   environment(gh_view_extent)   <- environment()
   environment(gh_selcat)        <- environment()
   environment(gh_selprod)       <- environment()
+  environment(gh_changeproj)    <- environment()
   environment(gh_tiles_from_bbox) <- environment()
-
+  
+  
+  
   bbox_out  <- NULL
   #   __________________________________________________________________________
   #   Start Building the GUI                                                ####
   wids <- list()
   # workaround to avoid NOTE
   compress_dict <- mod_proj_str <- size <- NULL
-
+  
   main_win <- gbasicdialog(
     title      = paste0("MODIStsp - v. ", utils::packageVersion("MODIStsp")),
     parent     = NULL,
-    do.buttons = FALSE,
-    use.scroll_window = scroll_window)
-
+    do.buttons = FALSE)
+  
   # frame1 and 2 with expand=FALSE grant that widgets are not "too much
   # expanded", neither horizontally nor vertically
-
+  
   main_frame1 <- ggroup(container  = main_win,
                         horizontal = TRUE,
                         expand     = FALSE,
                         use.scroll_window = scroll_window)
-
+  
   main_frame2 <- ggroup(container  = main_frame1,
                         horizontal = FALSE,
                         expand     = FALSE,
                         use.scroll_window = scroll_window)
-
+  
   #   __________________________________________________________________________
   #   Initialize Main container: holds all widgets                          ####
-
+  
   main_group  <- ggroup(container  = main_frame2,
                         horizontal = FALSE,
-                        expand     = FALSE)
-
+                        expand     = TRUE)
+  
   if (scroll_window) {
     getToolkitWidget(main_win)$maximize()
   }
-
+  
   mod_prod_cat <- as.data.frame(
     t(vapply(prod_opt_list, function(x){
       c(x[[1]]$cat01, x[[1]]$cat02)
@@ -106,7 +107,7 @@ MODIStsp_GUI <- function(general_opts,
   )
   names(mod_prod_cat) <- c("cat01", "cat02")
   mod_prod_cat$cat    <- apply(mod_prod_cat, 1, paste, collapse = " - ")
-
+  
   # get the product name selected in the previous options file and find the
   # corresponding category
   mod_prod_list <- names(prod_opt_list)
@@ -119,19 +120,15 @@ MODIStsp_GUI <- function(general_opts,
   out_proj_list  <- list(
     "Native"   = ifelse(
       prod_opt_list[[sel_prod]][[general_opts$prod_version]][["tiled"]] == "1",
-      "+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs",
[…]
