### 2018-03-04 #184 @lbusett
Hi all,

 a quick update: we are **almost done** with implementing the required modifications (You can find a commented list of work advancement here: https://github.com/lbusett/MODIStsp/issues/112).

Before starting fine tuning the release (e.g., update documentation, website, etcetera) **I wanted to ask (if possible) some comments on the changes we implemented on the GUI** to address comments by @lwasser.

You can install the current development version using:

`devtools::install_github("lbusett/MODIStsp", ref = "ropensci-review")`

(Note that you will need to have the current `mapedit` CRAN version (0.4.1) installed for this to work - for now I added it as an import, but I plan to put it as a "Suggest" to keep the dependency tree as small as possible).

The new GUI implements most of your comments (You can see a screenshot below). In particular:

- adds functionality for **interactive selection of processing extent** (In the end, I decided that it was a too good idea to skip it, and `mapedit` allowed me to do it relatively easily)
- Improves selectors for time series formats
- streamlined selection of output projection (now it is also possible to select a projection by simply specifiying an epsg code) and adds warning messages
- adds new/improved help buttons

I still need to fine-tune it, but I wanted to ask if you think this goes in the right direction and / or if you have further comments/suggestions.

![image](https://user-images.githubusercontent.com/4480397/36944425-1bd88ba8-1f9c-11e8-9dbb-1e31d7965b6b.png)

 Thanks in advance for any feedback,

Lorenzo
