### 2018-03-05 lbusett/MODIStsp#113 @lbusett
Hi Colby. Are you still having trouble with this? Otherwise, I will close the issue.
