### 2018-03-05 lbusett/MODIStsp#113 @cobanton
Hey Lorenzo,

I tried it a couple times and just did each winter separately.  The Extract
function seems to be working properly now as well. Thanks for all your help!

Colby

On Mon, Mar 5, 2018 at 8:40 AM, Lorenzo Busetto <notifications@github.com>
wrote:

> Hi Colby. Are you still having trouble with this? Otherwise, I will close
> the issue.
>
