### 2018-03-06 lbusett/MODIStsp#113 @ranghetti
Dear Colby,
I have noticed just now that you already posted the screenshot of the GUI with the parameters. I will have a look as soon as possible.
