### 2018-03-06 c8069e6
Fix seasonal interval (#113)
The problem was related to seasonal winter intervals
