### 2018-03-06 3d97370
Fix seasonal interval (#113)
The problem was related to seasonal winter intervals
