### 2018-03-06 83b10c6
Fix seasonal interval (#113)
The problem was related to seasonal winter intervals
