### 2018-03-06 lbusett/MODIStsp#113 @ranghetti
Dear Colby,
now this issue should have been fixed; you can try to reinstall the package from GitHub, and you should be able to process the whole period without need to separate each winter.
If you will do it, it could be useful to us to have your feedback.
