+++ vignettes/MODIStsp.Rmd
@@ -2,6 +2,7 @@
 title: 'MODIStsp: A Tool for Automatic Preprocessing of MODIS Time Series - v1.3.3'
 author: "Lorenzo Busetto ([busetto.l@irea.cnr.it](mailto:lbusett@gmail.com)), Luigi
   Ranghetti ([ranghetti.l@irea.cnr.it](mailto:ranghetti.l@irea.cnr.it))"
+bibliography: MODIStsp.bib
 date: '`r Sys.Date()`'
 output:
   rmarkdown::pdf_document:
@@ -260,7 +261,7 @@ The  GUI allows selecting all processing options required for the creation of th
   library(png)
   library(grid)
   library(knitr)
-  img <- readPNG("Figure_1_Vignette.png")
+  img <- readPNG("MAIN_GUI.png")
   grid.raster(img)
 ```
 
@@ -278,7 +279,7 @@ After selecting the product and version, clicking the **"Click to Select Desired
 ```{r proc_layers, echo=FALSE, fig.align="center", message=FALSE, warning=FALSE, paged.print=FALSE}
   library(png)
   library(grid)
-  img <- readPNG('Figure_3.png')
+  img <- readPNG('SELECT_GUI.png')
   grid.raster(img)
 ```
 
@@ -307,7 +308,7 @@ You can however **specify other SIs to be computed without modifying MODIStsp so
 ```{r indexfig, echo=FALSE, fig.height=2.3, fig.width=6, message=FALSE, warning=FALSE, fig.align="center"}
   library(png)
   library(grid)
-  img <- readPNG('figure_4.png')
+  img <- readPNG('NEWIND_GUI.png')
   grid.raster(img)
 ```
 
@@ -341,14 +342,25 @@ The **Period** drop-down menu allows to choose between two options:
 
 Allows to define the area of interest for the processing. Two main options are possible: 
 
-1.  **Full Tiles Extent**: specify which MODIS tiles need to be processed using the "Start" and "End" horizontal and vertical sliders in the _Required MODIS Tiles_ frame.  During processing, data from the different tiles is mosaiced, and a single file covering the total area is produced for each acquisition date (Note:  pressing the "show map" button, a representation of the MODIS tiles grid is shown to facilitate the selection).
+1.  **Select MODIS Tiles**: specify which MODIS tiles need to be processed either by: 
 
-2.  **Resized**: specify a custom spatial extent for the desired outputs either by: 
+    a. Using the "Start" and "End" horizontal and vertical sliders in the _Required MODIS Tiles_ frame.  
+    b. pressing the __"Select on Map"__ button. A map will open in a browser window, allowing interactive selection of the required tiles
+    
+During processing, data from the different tiles is mosaiced, and a single file covering the total area is produced for each acquisition date
+
+2.  **Define Custom Area**: specify a custom spatial extent for the desired outputs either by: 
     
     a. Manually inserting the coordinates of the Upper Left and Lower Right corners of the area of interest in the __Bounding Box__ frame. _Coordinates of the corners must be provided in the coordinate system of the selected  output projection_; 
     
     b. pressing the __"Load Extent from a Spatial File" and selecting a raster or vector spatial file__. In this case, the bounding box of the selected file is retrieved, converted in the selected output projection, and shown in the "Bounding Box" frame. Required input MODIS tiles are also automatically retrieved from the output extent, and the tiles selection sliders modified accordingly.
     
+    c. pressing the __"Select on Map"__ button. A map will open in a browser window, allowing interactive selection of the spatial extent using the tools on the left
+    
+    
+**Note: ** pressing the "show current extent" will open a browser window highlighting
+the currently selected spatial extent.
+ 
[…]
