+++ vignettes/interactive_execution.Rmd
@@ -1,5 +1,6 @@
 ---
 title: "Running the tool in Interactive Mode: the MODIStsp GUI"
+bibliography: MODIStsp.bib
 ---
 
 ```{r setup, include=FALSE}
@@ -7,23 +8,26 @@ knitr::opts_chunk$set(echo = TRUE)
 library(DT)
 ```
 
-The easiest way to use `MODIStsp` is to use its powerful GUI (Graphical User Interface) for 
-selection of processing options, and then run the processing. 
+The easiest way to use `MODIStsp` is to use its powerful GUI (Graphical User Interface) 
+for selection of processing options, and then run the processing. 
 
 To open the GUI, load the package and launch the MODIStsp function, with no parameters:
 ```{r, eval=FALSE, message=FALSE, warning=FALSE, include=TRUE, caption=FALSE}
 library(MODIStsp)
 MODIStsp()
 ```
-This **opens a GUI** from which processing options can be specified and eventually saved (or loaded) (Notes: [1](#note-1), [2](#note-2)). 
+This **opens a GUI** from which processing options can be specified and eventually 
+saved (or loaded) (Notes: [1](#note-1), [2](#note-2)). 
 
-The  GUI allows selecting all processing options required for the creation of the desired MODIS time series. The main available processing options are described in detail in the following.
+The  GUI allows selecting all processing options required for the creation of the 
+desired MODIS time series. The main available processing options are described in 
+detail in the following.
 
 ```{r GUIfig, echo=FALSE, fig.align="center", fig.heigth=7.5, message=FALSE, warning=FALSE}
   library(png)
   library(grid)
   library(knitr)
-  img <- readPNG("Figure_1_Vignette.png")
+  img <- readPNG("MAIN_GUI.png")
   grid.raster(img)
 ```
 
@@ -38,21 +42,30 @@ ________________________________________________________________________________
 
 The top-most menus allow to specify details of the desired output time series:
 
-1. **"Category"** and **"Product"**: Selects the MODIS product of interest  
-2. **MODIS platform(s)**: Selects if only TERRA, only AQUA or Both MODIS platforms should be considered for download and creation of the time series
-3. **version**: Selects whether processing version 5 or 6 (when available) of MODIS products has to be processed
+1. **"Category"** and **"Product"**: Selects the MODIS product of interest;
+2. **MODIS platform(s)**: Selects if only TERRA, only AQUA or Both MODIS platforms 
+should be considered for download and creation of the time series;
+3. **version**: Selects whether processing version 5 or 6 (when available) of MODIS
+products has to be processed
 
-After selecting the product and version, clicking the **"Click to Select Desired Layers"** button opens the *_Select Processing Layers_* GUI panel, from which the user **must** select which MODIS original layers and/or derived Quality Indexes (QI) and Spectral Indexes (SI) layers should be processed: 
+After selecting the product and version, clicking the **"Change Selection"** 
+button opens the *_Select Processing Layers_* GUI panel, from which the user **must** 
+select which MODIS original layers and/or derived Quality Indexes (QI) and Spectral
+Indexes (SI) layers should be processed: 
 ```{r proc_layers, echo=FALSE, fig.align="center", message=FALSE, warning=FALSE, paged.print=FALSE}
   library(png)
   library(grid)
-  img <- readPNG('Figure_3.png')
+  img <- readPNG('SELECT_GUI.png')
   grid.raster(img)
 ```
 
-- The left-hand frame allows to select which _original MODIS layers_ should be processed.
-- The central frame allows to select which _Quality Indicators should be extracted_ from the original MODIS Quality Assurance layers. 
[…]

[…]
