+++ vignettes/installation.Rmd
@@ -162,80 +162,41 @@ Save the modified file using `ctrl+x ctrl+c`, followed by `y` to quit emacs.
 
 ```
 ~$ export PKG_CONFIG_PATH=/usr/local/lib/pkgconfig:/usr/local/lib/pkgconfig/gtk+-2.0.pc:/opt/X11/lib/pkgconfig
-
-
 ```
 
-5. Now install the `gtk+` library with x11 support. To do so, you first have to 
-.
+8. Download the newest source file for RGtk2 from https://cran.r-project.org/web/packages/RGtk2/index.html.
 
-For installation on MAC OSX sierra, there are three main issues:
-- As outlined here in the comment by **tobybot11** (https://gist.github.com/sebkopf/9405675), 
-RGtk requires the x11 libraries/headers (gdk/gdkx.h specifically) and does not
-work with the quartz libraries/headers which now are the default for GTK
-- When installing the dependencies gWidgetsRGtk2 and cairoDevice from CRAN you
-need to choose the version "OS X Mavericks binaries" and not "Package source"
-- You have to be sure that gdal is installed with HDF4 support.
+9. Assuming that the path to this file is ~/Downloads. Run the following in
+terminal (change the path if you did not download in ~/Downloads):
 
-1 - Update to R > 3.4  if needed, then update all packages
+```
+~$ cd ~/Downloads
+~/Downloads$ R CMD INSTALL RGtk2_2.20.33.tar.gz
+```
+Note that versions may vary depending on when you download the source file. 
+Use the exact filename that sits in the directory (Tip: use tab to help autocomplete).
 
-```{r eval = FALSE}
-# update packages
-update.packages()
-# install the development version of devtools:
-install.packages(c("devtools"))
-devtools::install_github("hadley/devtools")
-```
-
-2 - Now, install RGtk2 using Homebrew (https://gist.github.com/sebkopf/9405675).
-First, ensure you have cairo installed with "--with-x11". Open a terminal and 
-run:
-
-```bash
-brew uninstall cairo --ignore-dependencies
-brew install --with-x11 cairo 
-```
-next, edit the configure options for GTK to require x11 rather than Quartz:
-```bash
-brew edit gtk+
-```
-in the def install section, remove the reference to quartz and switch to: 
-```
---with-gdktarget=x11,
---enable-x11-backend
-```
-Now install: 
-```bash
-brew install --build-from-source --verbose gtk+
-export PKG_CONFIG_PATH=/usr/local/lib/pkgconfig:/usr/local/lib/pkgconfig/gtk+-2.0.pc:/opt/X11/lib/pkgconfig
-```
-
-3 - next, install `RGtk2`. Since install.packages("Rgtk2") is not going to work, go [here](https://cran.r-project.org/web/packages/RGtk2/index.html), download 
-RGtk2_2.20.33.tar.gz and, from a terminal run:
-```bash
-R CMD INSTALL RGtk2_2.20.33.tar.gz
-```
 Now, open R and run: 
 
 ```{r eval = FALSE}
 library(RGtk2)
 ```
-hopefully, `RGtk2` will load without errors !
+hopefully, `RGtk2` will load without errors! If so, you should be ready to go, 
+and you can:
 
-4 - Install packages `gWidgetsRGtk2` and `cairoDevice`
-
-**Very important !!!!** from CRAN, download the **"OS X Mavericks binaries"** for 
-[gWidgetsRGtk2](https://cran.r-project.org/web/packages/gWidgetsRGtk2/index.html) and [cairoDevice](https://cran.r-project.org/web/packages/cairoDevice/index.html)(
[…]
