### 2018-03-08 ce72a7c
continue work on documentation + website update + new test
