+++ R/MODIStsp.R
@@ -113,20 +113,20 @@ MODIStsp <- function(gui               = TRUE,
                      test              = -1,
                      n_retries         = 20,
                      verbose           = TRUE) {
-  
+# browser()
   options("guiToolkit" = "RGtk2")
   # Make so that "raster" functions does not automatically add extensions on
   # output files. This is automatically reset to TRUE at the end of the session
   raster::rasterOptions(setfileext = FALSE)
-  
+
   #   __________________________________________________________________________
   #   Initialize processing                                                 ####
-  
+
   #   __________________________________________________________________________
   #   If test mode is selected, select the options file for testing         ####
   #   and set other parameters
   if (test > -1) {
-    
+
     gui <- FALSE
     message("MODIStsp is running in test mode.")
     # read names of available json test files
@@ -154,19 +154,19 @@ MODIStsp <- function(gui               = TRUE,
               unzip   = "internal")
       }
     }
-    
+
     # Assign the selected test Option File
     options_file <- test_files[test]
     start <- TRUE
   }
-  
+
   #   __________________________________________________________________________
   #   On interactive execution, ensure that gWidgetsRGtk2 is avauilable     ####
-  
+
   if (gui) {
     #nocov start
     if (!pacman::p_exists("gWidgetsRGtk2", local = TRUE)) {
-      
+
       message(strwrap("Library 'gWidgetsRGtk2' is not installed. It is required
                       to run MODIStsp in interactive mode!\n\n",
                       "Do you want to install it now?"),
@@ -178,15 +178,15 @@ MODIStsp <- function(gui               = TRUE,
         stop(strwrap("MODIStsp can not work in Interactive mode without
                      gWidgetsRGtk2! Aborting!"))
       }
-      
+
     }
     options("guiToolkit" = "RGtk2")
     #nocov end
   }
-  
+
   # Check GDAL version ----
   if (is.null(getOption("gdalUtils_gdalPath"))) {
-    
+
     welcome_text <- strwrap("Welcome to MODIStsp!\n\nWe will now search for a
                             valid GDAL installation - please wait! (this will
                             happen only once)", width = 60)
@@ -194,7 +194,7 @@ MODIStsp <- function(gui               = TRUE,
       #nocov start
       welcome_win       <- gWidgets::gwindow(title  = "Welcome", width = 400,
                                              height = 100)
-      
+
       welcome_lab       <- gWidgets::glabel(text      = welcome_text,
                                             container = welcome_win,
                                             editable  = FALSE)
@@ -206,53 +206,61 @@ MODIStsp <- function(gui               = TRUE,
     } else {
       message("[", date(), "] ", welcome_text)
     }
-    
+
     gdalUtils::gdal_setInstallation(ignore.full_scan = TRUE)
   }
   gdal_version <- package_version(
     gsub("^GDAL ([0-9.]*)[0-9A-Za-z/., ]*", "\\1",
          rgdal::getGDALVersionInfo(str = "--version"))
   )
-  
+
   # GDAL version used as minimum required version
   gdal_minversion  <- package_version("1.11.1")
   gdal_hdf_support <- length(grep("HDF4",
                                   gdalUtils::gdalinfo(formats = TRUE))) > 0
-  
+
   if (gdal_version < gdal_minversion) {
     stop(paste0("GDAL version must be at least ",
                 gdal_minversion,
                 ". Please update it."))
   }
-  
+
   if (!gdal_hdf_support) {
     stop("Your local GDAL installation does not support HDF4 format.\n",
          "Please install HDF4 support and recompile GDAL. See:\n",
          strwrap("http://lbusett.github.io/MODIStsp/articles/installation.html#
[…]
