+++ R/MODIStsp_GUI.R
@@ -41,10 +41,10 @@ MODIStsp_GUI <- function(general_opts,
   # nocov start
   #   __________________________________________________________________________
   #   NOTE: The function is excluded from coverage reports since it must be ####
-  #   used interactively. The same goes for all helper function and event 
+  #   used interactively. The same goes for all helper function and event
   #   handlers!
-  help_messages <- create_help_messages()  
-
+  help_messages <- create_help_messages()
+# browser()
   # create a new env to facilitate values-passing between widgets + set the
   # environment of helpers to that of the function to avoid having to create
   # globals TODO consider remeoving and passing everything explicitly!
@@ -60,45 +60,45 @@ MODIStsp_GUI <- function(general_opts,
   environment(gh_selprod)       <- environment()
   environment(gh_changeproj)    <- environment()
   environment(gh_tiles_from_bbox) <- environment()
-  
-  
-  
+
+
+
   bbox_out  <- NULL
   #   __________________________________________________________________________
   #   Start Building the GUI                                                ####
   wids <- list()
   # workaround to avoid NOTE
   compress_dict <- mod_proj_str <- size <- NULL
-  
+
   main_win <- gbasicdialog(
     title      = paste0("MODIStsp - v. ", utils::packageVersion("MODIStsp")),
     parent     = NULL,
     do.buttons = FALSE)
-  
+
   # frame1 and 2 with expand=FALSE grant that widgets are not "too much
   # expanded", neither horizontally nor vertically
-  
+
   main_frame1 <- ggroup(container  = main_win,
                         horizontal = TRUE,
                         expand     = FALSE,
                         use.scroll_window = scroll_window)
-  
+
   main_frame2 <- ggroup(container  = main_frame1,
                         horizontal = FALSE,
                         expand     = FALSE,
                         use.scroll_window = scroll_window)
-  
+
   #   __________________________________________________________________________
   #   Initialize Main container: holds all widgets                          ####
-  
+
   main_group  <- ggroup(container  = main_frame2,
                         horizontal = FALSE,
                         expand     = TRUE)
-  
+
   if (scroll_window) {
     getToolkitWidget(main_win)$maximize()
   }
-  
+
   mod_prod_cat <- as.data.frame(
     t(vapply(prod_opt_list, function(x){
       c(x[[1]]$cat01, x[[1]]$cat02)
@@ -107,7 +107,7 @@ MODIStsp_GUI <- function(general_opts,
   )
   names(mod_prod_cat) <- c("cat01", "cat02")
   mod_prod_cat$cat    <- apply(mod_prod_cat, 1, paste, collapse = " - ")
-  
+
   # get the product name selected in the previous options file and find the
   # corresponding category
   mod_prod_list <- names(prod_opt_list)
@@ -125,10 +125,10 @@ MODIStsp_GUI <- function(general_opts,
     "User Defined" = ""
   )
   mod_proj_str <- "+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs" #nolint
-  
+
   #   __________________________________________________________________________
   #   Initialize Widgets for product selection and bands selection          ####
-  
+
   satprod_frame <- gframe(
     text       = strwrap("<span foreground='red' size='x-large'>
                          MODIS Product and Platform </span>"),
@@ -137,34 +137,34 @@ MODIStsp_GUI <- function(general_opts,
     container  = main_group,
     spacing    = 5
   )
-  
+
   # set dummy global variables holding the initial values of selected bands
   gui_env$temp_wid_bands         <- general_opts$bandsel
   gui_env$temp_wid_bands_indexes <- general_opts$indexes_bandsel
[…]

[…]
