### 2018-03-11 08e5fc5
Improved gui_loadopts behavior when loading jsons w/user proj
+ moved all event handlers to dedicated .R files
