### 2018-03-11 0725455
Moved event handlers to dedicated files
also, explicitly added required args and removed "..." from arguments
