+++ R/MODIStsp_GUI.R
@@ -43,16 +43,20 @@ MODIStsp_GUI <- function(general_opts,
   #   NOTE: The function is excluded from coverage reports since it must be ####
   #   used interactively. The same goes for all helper function and event
   #   handlers!
+  
+  compress_dict <- mod_proj_str <- NULL
+  bbox_out  <- NULL
   help_messages <- create_help_messages()
-# browser()
+  
   # create a new env to facilitate values-passing between widgets + set the
   # environment of helpers to that of the function to avoid having to create
   # globals TODO consider remeoving and passing everything explicitly!
+  
   gui_env       <- new.env()
   gui_env$start <- FALSE
   environment(gui_load_options) <- environment()
   environment(gui_save_options) <- environment()
-  environment(gh_childs)        <- environment()
+  environment(gh_sellayers)        <- environment()
   environment(gh_load_extent)   <- environment()
   environment(gh_selectmap)     <- environment()
   environment(gh_view_extent)   <- environment()
@@ -60,45 +64,42 @@ MODIStsp_GUI <- function(general_opts,
   environment(gh_selprod)       <- environment()
   environment(gh_changeproj)    <- environment()
   environment(gh_tiles_from_bbox) <- environment()
-
-
-
-  bbox_out  <- NULL
+  
   #   __________________________________________________________________________
   #   Start Building the GUI                                                ####
   wids <- list()
   # workaround to avoid NOTE
-  compress_dict <- mod_proj_str <- size <- NULL
-
+  
+  
   main_win <- gbasicdialog(
     title      = paste0("MODIStsp - v. ", utils::packageVersion("MODIStsp")),
     parent     = NULL,
     do.buttons = FALSE)
-
+  gWidgets::size(main_win) <- list(width = 820)
   # frame1 and 2 with expand=FALSE grant that widgets are not "too much
   # expanded", neither horizontally nor vertically
-
+  
   main_frame1 <- ggroup(container  = main_win,
                         horizontal = TRUE,
-                        expand     = FALSE,
+                        expand     = TRUE,
                         use.scroll_window = scroll_window)
-
+  
   main_frame2 <- ggroup(container  = main_frame1,
                         horizontal = FALSE,
-                        expand     = FALSE,
+                        expand     = TRUE,
                         use.scroll_window = scroll_window)
-
+  
   #   __________________________________________________________________________
   #   Initialize Main container: holds all widgets                          ####
-
+  
   main_group  <- ggroup(container  = main_frame2,
                         horizontal = FALSE,
                         expand     = TRUE)
-
+  
   if (scroll_window) {
     getToolkitWidget(main_win)$maximize()
   }
-
+  
   mod_prod_cat <- as.data.frame(
     t(vapply(prod_opt_list, function(x){
       c(x[[1]]$cat01, x[[1]]$cat02)
@@ -107,7 +108,7 @@ MODIStsp_GUI <- function(general_opts,
   )
   names(mod_prod_cat) <- c("cat01", "cat02")
   mod_prod_cat$cat    <- apply(mod_prod_cat, 1, paste, collapse = " - ")
-
+  
   # get the product name selected in the previous options file and find the
   # corresponding category
   mod_prod_list <- names(prod_opt_list)
@@ -125,10 +126,10 @@ MODIStsp_GUI <- function(general_opts,
     "User Defined" = ""
   )
   mod_proj_str <- "+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs" #nolint
-
+  
   #   __________________________________________________________________________
[…]
