### 2018-03-11 6aead80
various small tweaks on documentation
