+++ R/MODIStsp.R
@@ -57,52 +57,68 @@
 #' # last used settings, in interactive mode (i.e., with gui = TRUE).
 #'
 #' \dontrun{
-#' MODIStsp()}
-#'
+#' MODIStsp()
+#' }
 #' # Running the tool using the settings previously saved in a specific options
 #' # file
 #'
-#' # **NOTE** Output files of examples will be saved to file.path(tempdir(), "MODIStsp").
-#' # You can run the example with `gui = TRUE` to set a different output folder!
+#' # **NOTE** Output files of examples are saved to file.path(tempdir(), "MODIStsp").
+#' # You can run the examples with `gui = TRUE` to set a different output folder!
 #'
 #' \dontrun{
 #'
-#' # Here we use a test json file saved in MODIStsp installation folder -
+#' # Here we use a test json file saved in MODIStsp installation folder which
 #' # Downloads and processed 3 MOD13A2 images over the Como Lake (Lombardy, Italy)
 #' # and retrieves NDVI and EVI data, plus the Usefulness Index Quality Indicator.
 #'
-#' options_file <- system.file("testdata/MOD13A2.json", package = "MODIStsp")
-#' MODIStsp(gui = FALSE, options_file = options_file)}
+#' options_file <- system.file("testdata/test_MOD13A2.json", package = "MODIStsp")
+#' MODIStsp(gui = FALSE, options_file = options_file, verbose = FALSE)
+#' }
 #'
+#' \dontrun{
 #' # Running the tool using the settings previously saved in a specific option file
 #' # and specifying the extent from a spatial file allows to re-use the same
 #' # processing settings to perform download and reprocessing on a different area
+#'  
+#' options_file <- system.file("testdata/test_MOD13A2.json", package = "MODIStsp")
+#' spatial_file <- system.file("testdata/lakeshapes/garda_lake.shp", package = "MODIStsp")
+#' MODIStsp(gui = FALSE, options_file = options_file,
+#'   spatial_file_path = spatial_file, verbose = FALSE)
+#' }
 #'
 #' \dontrun{
-#' options_file <- system.file("testdata/MOD13A2.json", package = "MODIStsp")
-#' spatial_file <- system.file("testdata/gardalake.shp")
-#' MODIStsp(gui = FALSE, options_file = options_file,
-#'   spatial_file_path = spatial_file)}
-#'
 #' # Running the tool using the settings previously saved in a
 #' # specific options file and specifying each time the extent from a different
 #' # spatial file (e.g., to perform the same processing on several extents)
 #'
-#' \dontrun{
-#' extent_list  <- c(system.file("testdata/gardalake.shp", package = "MODIStsp"),
-#'                   system.file("testdata/iseolake.shp", package = "MODIStsp"))
-#'
+#' extent_list  <- c(system.file("testdata/lakeshapes/garda_lake.shp", 
+#'                               package = "MODIStsp"),
+#'                   system.file("testdata/lakeshapes/iseo_lake.shp", 
+#'                               package = "MODIStsp"))
+#' extent_list
+#' 
 #' # Note that you can also put all your extent files in a specific folder and
-#' # create the extent list using
-#' # extent_list = list.files("X:/path/containing/some/shapefiles/", "\\.shp$")
+#' # create the extent list using for example.
+#' # extent_list = list.files(system.file("testdata/lakeshapes/", package = "MODIStsp"), 
+#' #                          full.names = TRUE, "\\.shp$")
 #'
-#' options_file <- system.file("testdata/MOD13A2.json", package = "MODIStsp")
+#' options_file <- system.file("testdata/test_MOD13A2.json", package = "MODIStsp")
 #' for (single_shape in extent_list) {
[…]
