### 2018-03-11 390cd57
moved mapedit and friends to Suggests
