+++ DESCRIPTION
@@ -42,18 +42,18 @@ Imports:
     stringr (>= 1.0.0),
     xts (>= 0.9-7),
     xml2 (>= 1.2.0),
-    sf(>= 0.6-1),
-    leaflet,
-    shiny, 
-    mapview (>= 2.3.0), 
-    mapedit (>= 0.4.1)
 Suggests:
     spelling,
     knitr,
     rmarkdown,
     png,
     grid,
-    httptest
+    httptest,
+    sf(>= 0.6),
+    leaflet,
+    shiny, 
+    mapview (>= 2.3.0), 
+    mapedit (>= 0.4.1)
 SystemRequirements: Cairo >= 1.0.0, ATK (>= 1.10.0), Pango (>= 1.10.0), GTK+ (>=
     2.8.0), GLib (>= 2.8.0), Curl, GDAL (>= 1.6.3), PROJ.4 (>= 4.4.9)
 URL: https://github.com/lbusett/MODIStsp
