+++ vignettes/MODIStsp.Rmd
@@ -5,16 +5,16 @@ author: "Lorenzo Busetto ([busetto.l@irea.cnr.it](mailto:lbusett@gmail.com)), Lu
 bibliography: MODIStsp.bib
 date: '`r Sys.Date()`'
 output:
-  rmarkdown::pdf_document:
-    fig_caption: yes
-    number_sections: yes
-    toc: yes
-    toc_depth: 2
   rmarkdown::html_document:
     fig_caption: yes
     number_section: yes
     toc: no
     toc_depth: 1
+  rmarkdown::pdf_document:
+    fig_caption: yes
+    number_sections: yes
+    toc: yes
+    toc_depth: 2
   rmarkdown::html_vignette:
     fig_caption: yes
     number_section: yes
@@ -30,21 +30,42 @@ vignette: >
 
 # Introduction
 
-MODIStsp is a novel "R" package allowing to automatize the creation of time series of rasters derived from MODIS Land Products data. It allows to perform several preprocessing steps on MODIS data available within a given time period. 
+MODIStsp is a novel "R" package allowing to automatize the creation of time series
+of rasters derived from MODIS Land Products data. It allows to perform several 
+preprocessing steps on MODIS data available within a given time period. 
 
-Development of MODIStsp started from modifications of the ModisDownload "R" script by Thomas Hengl [-@Hengl2010], and successive adaptations by Babak Naimi [-@Naimi2014]. The basic functionalities for download and preprocessing of MODIS datasets provided by these scripts were gradually incremented with the aim of: 
+Development of MODIStsp started from modifications of the ModisDownload "R" script
+by Thomas Hengl [-@Hengl2010], and successive adaptations by Babak Naimi [-@Naimi2014]. 
+The basic functionalities for download and preprocessing of MODIS datasets provided 
+by these scripts were gradually incremented with the aim of: 
 
-* developing a stand-alone application allowing to perform several preprocessing steps (e.g., download, mosaicing, reprojection and resize) on all available MODIS land products by exploiting  a powerful and user-friendly GUI front-end;
-* allowing the creation of time series of both MODIS original layers and additional Quality Indicators (e.g., data acquisition quality, cloud/snow presence, algorithm used for data production, etc. ) extracted from the aggregated bit-field QA layers;
-* allowing the automatic calculation and creation of time series of several additional Spectral Indexes starting form MODIS surface reflectance products.
+* developing a stand-alone application allowing to perform several preprocessing
+steps (e.g., download, mosaicing, reprojection and resize) on all available MODIS 
+land products by exploiting  a powerful and user-friendly GUI front-end;
+* allowing the creation of time series of both MODIS original layers and additional 
+Quality Indicators (e.g., data acquisition quality, cloud/snow presence, algorithm 
+used for data production, etc. ) extracted from the aggregated bit-field QA layers;
+* allowing the automatic calculation and creation of time series of several additional
+Spectral Indexes starting form MODIS surface reflectance products.
 
-All processing parameters can be easily set with a user-friendly GUI, although non-interactive execution exploiting a previously created Options File is possible. Stand-alone execution outside an "R" environment is also possible, allowing to use scheduled execution of MODIStsp to automatically update time series related to a MODIS product and extent whenever a new image is available. 
+All processing parameters can be easily set with a user-friendly GUI, although
[…]
