+++ .travis.yml
@@ -36,7 +36,7 @@ matrix:
             - libgtk2.0-dev
             - libglib2.0-dev
             - libcurl4-openssl-dev
-            - libudunits2-0
+            - libudunits2-dev
       r_binary_packages:
         - cairoDevice
         - RGtk2
