### 2018-03-12 eb62a2e
final changes for ropensci resubmission
linting and website
