+++ vignettes/MODIStsp.Rmd
@@ -548,41 +548,116 @@ output folder, as explained in detail [HERE](articles/output.html).
 # Non-Interactive Execution from within R
 
 ## Specifying a saved "Options file"
-`MODIStsp` can be launched in non-interactive mode within an `R` session by
-setting the optional `GUI` parameter to FALSE, and the `Options_File` parameter
-to the path of a previously saved JSON Options file. This allows to exploit
-`MODIStsp` functionalities within generic "R" processing scripts
+`MODIStsp` can be launched in non-interactive mode within an `R` session by setting 
+the optional `GUI` parameter to FALSE, and the `Options_File` parameter to the path 
+of a previously saved JSON Options file. 
+This allows to exploit `MODIStsp` functionalities within generic "R" processing 
+scripts: 
 
-```{r, eval=FALSE}
+```{r eval=FALSE, message=FALSE, warning=FALSE}
 library(MODIStsp) 
+
+# **NOTE** Output files of examples are saved to file.path(tempdir(), "MODIStsp").
+# You can run the examples with `gui = TRUE` to set a different output folder!
+
 # --> Specify the path to a valid options file saved in advance from MODIStsp GUI 
-options_file <- "X:/yourpath/youroptions.json" 
-  
+# Here we use a test json file saved in MODIStsp installation folder which
+# downloads and processed 3 MOD13A2 images over the Como Lake (Lombardy, Italy)
+# and retrieves NDVI and EVI data, plus the Usefulness Index Quality Indicator.
+
+options_file <- system.file("testdata/test_MOD13A2.json", package = "MODIStsp")
+
 # --> Launch the processing
-MODIStsp(gui = FALSE, options_file = options_file)
+MODIStsp(gui = FALSE, options_file = options_file, verbose = FALSE)
+
+# Outputs are in this case in subfolder "MODIStsp/VI_16Days_1Km_v6" of `base::tempdir()`: 
+
+out_fold <- file.path(tempdir(), "MODIStsp/VI_16Days_1Km_v6") 
+list.files(out_fold)
+#> [1] "EVI"         "NDVI"        "QA_usef"     "Time_Series"
+
+list.files(file.path(out_fold ,"EVI"))
+#> [1] "MOD13A2_EVI_2016_161.tif" "MOD13A2_EVI_2016_177.tif"
 ```
+
+## Looping over different Options files
+
+If you need to process different MODIS products, you can prepare beforehand different
+MODIStsp options files by using the GUI, and then loop over them like this:
+
+```{r eval=FALSE, message=FALSE, warning=FALSE}
+
+opts_files <- c(system.file("testdata/test_MOD13A2.json", package = "MODIStsp"), 
+                system.file("testdata/test_MOD10A2.json", package = "MODIStsp"))
+
+for (opts_file in opts_files) {
+  MODIStsp(gui = FALSE, options_file = opts_file)
+}
+
+# MOD13A2 ouputs
+out_fold <- file.path(tempdir(), "MODIStsp/VI_16Days_1Km_v6") 
+list.files(out_fold)
+#> [1] "EVI"         "NDVI"        "QA_usef"     "Time_Series"
+list.files(file.path(out_fold ,"EVI"))
+#> [1] "MOD13A2_EVI_2016_161.tif" "MOD13A2_EVI_2016_177.tif"
+
+# MOD10A2 ouputs
+out_fold <- file.path(tempdir(), "MODIStsp/Surf_Temp_8Days_1Km_v6") 
+list.files(out_fold)
[…]
