+++ vignettes/installation.Rmd
@@ -141,16 +141,26 @@ export PKG_CONFIG_PATH=/usr/local/lib/pkgconfig:/usr/local/lib/pkgconfig/gtk+-2.
 ```
 <br>
 
-__2. Install the `cairo` library with x11 support__. To do so, you first have to 
-change the way homebrew wants to install gtk+. In an editor, write: 
-
+__2. Install the `cairo` library with x11 support__. Enter the following into your terminal:
+ 
+```
+~$ brew uninstall --force cairo --ignore-dependencies
+~$ brew cask install xquartz
+~$ brew install --with-x11 cairo
+```
+<br>
+ 
+__3. Install the `gtk+` library__:
+ 
+To do so, you first have to change the way homebrew wants to install gtk+. In an editor, write: 
+ 
 ```
 ~$ brew edit gtk+
 ```
-
+ 
 A text editor will open. Look in the file, and find a section that begins with
-"def install". Substitute the current `args` with the following text: 
-
+"def install". Substitute the current `args` section with the following text: 
+ 
 ```
 def install
  args = [
@@ -165,12 +175,10 @@ def install
          "--enable-x11-backend"
         ]
 ```     
-
+ 
 Save the modified file using `ctrl+x ctrl+c`, followed by `y` to quit emacs.
-<br>
-
-__3. Install the `gtk+` library__ by issuing this command:
-
+Now install the library using: 
+ 
 ```
 ~$ brew install --build-from-source --verbose gtk+
 ```
