### 2018-03-12 #184 @lbusett
at last, we managed to wrap-up an improved MODIStsp version addressing reviewers' suggestion. The current version can be installed with:

    devtools::install_github("lbusett/MODIStsp")

In short, without going in detail through reviewers comments as in the last posts), here is a quick list of the implemented changes:

## for @jeffreyhanson review:

- [x] All examples in vignettes and documentation now save by default in tempdir (also, at first GUI execution the output folders are set to $tempdir).  If $tempdir$ is found as folder selection the outputs go to **tempdir/MODStsp** and **tempdir/MODIStsp/HDFs**

- [x] To avoid problems with new CRAN rules, at first MODIStsp execution the user is now asked permission to save MODIStsp previous options file in /Extdata/Previous

- [x] Rcurl dependency was removed. We now use `httr::HEAD` to check file size over ftp

- [x] We switched from "print" to "message" in MODIStsp_extract

- [x] We fixed cran warnings about "C:\Progra" etc.

-  [x] We fixed the references in the PDF vignette

## for @lwasser review:

- [x] We improved installation instructions for mac - fro both MODistsp package and GDAL

Improved instructions can be found at http://lbusett.github.io/MODIStsp/articles/installation.html#installing-on-mac and http://lbusett.github.io/MODIStsp/articles/installation.html#macos

Instructions were mainly derived from https://zhiyzuo.github.io/installation-rattle/. @lwasser : May you have a look and see if you think these are "good enough"?

### GUI Improvements

We made various improvements in the GUI to address your comments (You can see a screenshot of the current GUI in my last post)

- [x] Improved the Product / Layers selection interface

* The Product Information button is now also available in the main GUI
* Selected layers are immediately visible in the Main Window. Help buttons were added to explain what the Quality Indicators and Spectral Indices are.

- [x] Added help suggestions on the "Add new index" functionality.

- [x] Strongly improved the "spatial extent" selection interface
* Now it is possible to interactively select the MODIS tiles or custom spatial extent using the "Select from Map" button, and see the currently selected extent in a browser using the "Show Current Extent" button.

- [x] Improved / streamlined the "reprojection selection" section

* Now only "Native" and "User Defined" can be chosen in the dropdown menu (befor, we also had some "random" projections in the dropdown because we use them often, but I think that this is cleaner)
* When specifying a custom projection, in the Input field it is now possible to specify a projection simply by specifying an EPSG code or a UTM codes, instead of writing the full proj4 string;
* Warning messages are now sent in case a User Projection is selected, to warn about possible problems introduced by reprojection;
* Changing projection now re-sets the custom extent to avoid problems due to sequential reprojections
of the extent (i.e., continuously enlarging bbox to avoid "losing" area).

- [x] Moved the hdf folder selector below the main output folder one and added help buttons on both output folder selectors to help the user

- [x] Modified the interface for selection of "formats" of virtual time series files.
* Now we use a checkbox group allowing specifying which "time series formats" are desired, among R rasterStack, ENVI META or GDAL vrts (i.e.,  ENVI, R and GDAL timeseries are now "treated equally" and are selected using a single menu). We also added help buttons to facilitate the user.

- [x]  Improved in general the help buttons and added some text formatting.

- [x] Changed the default output format to "GeoTiff", the default virtual raster files format to "rasterStack" and the default compression to "None"

These changes address most of the points raised by @lwasser , and I think that they in general improved user-friendliness of the interface (so, thanks for the suggestions!). The only point we were not able to address were those regarding:

> When I have several boxes open (the select processing layers box and the insert new spectral index box) PLUS the original tsp interface... I have to figure out what order they are in so I can close them. Some users may think the entire thing is locked when really it’s just too many popups. This is a bit clunky for me to have 3 boxes and have to figure out the proper order! And it locked my computer up a bunch. I’m sure this is quartz specific but any way to control it?

Unfortunately, I see no easy way to do this because the different windows are "modal". I do not believe however this is a major issue. I can reevaluate if this is considered mandatory.

## Other changes:

* We improved functionality for dealing with NoData for products with multiple fill-values  and fixed a couple of bugs affecting the "Seasonal" time series download (see https://github.com/lbusett/MODIStsp/issues/113)

* We made some changes in documentation. In particular, we improved some of the pkgdown articles to better help the users (see for example http://lbusett.github.io/MODIStsp/articles/noninteractive_execution.html and file:///D:/Documents/Source/git/MODIStsp/docs/articles/output.html)

* To allow the new "interactive" spatial extent selection, we added "sf", "leaflet", "shiny", "mapview" and "mapedit" as "Suggests" in the DESCRIPTION file.  If the user clicks on "Select from Map" and `mapedit` is not installed, he is prompted to install it. In this way, we avoid bringing in "heavy" dependencies. @karthik: do you think that is ok, or would it be better to put them in "Imports"?

OK, that's it. We hope you will like the new version and that we addressed all points raised in the review. Please let us know if you see any other issues or you need further changes/clarifications.
