### 2018-03-14 #184 @lwasser
this all looks great @lbusett  :) excited to see this package available on ropensci . it's a great contribution!!
