### 2018-03-15 7926c16
corrected GUI size and updated spelling
