+++ R/MODIStsp_GUI.R
@@ -75,7 +75,7 @@ MODIStsp_GUI <- function(general_opts,
     title      = paste0("MODIStsp - v. ", utils::packageVersion("MODIStsp")),
     parent     = NULL,
     do.buttons = FALSE)
-  gWidgets::size(main_win) <- list(width = 820)
+  gWidgets::size(main_win) <- list(width = 900)
   # frame1 and 2 with expand=FALSE grant that widgets are not "too much
   # expanded", neither horizontally nor vertically
 
@@ -262,7 +262,7 @@ MODIStsp_GUI <- function(general_opts,
   wids$sel_layers <- gtext(text = ifelse(curr_sel_layers == "",
                                          "- None Selected - ", curr_sel_layers),
                            container = layer_group,
-                           width     = 400,
+                           width     = 600,
                            height    = 15,
                            editable  = FALSE,
                            expand    = TRUE)
@@ -291,7 +291,7 @@ MODIStsp_GUI <- function(general_opts,
   wids$sel_qi   <- gtext(text = ifelse(curr_sel_qual == "",
                                        "- None Selected - ", curr_sel_qual),
                          container = qual_group,
-                         width     = 400,
+                         width     = 600,
                          height    = 15,
                          editable  = FALSE,
                          expand    = TRUE)
@@ -326,7 +326,7 @@ MODIStsp_GUI <- function(general_opts,
   wids$sel_si  <- gtext(text = ifelse(curr_sel_si == "",
                                       "- None Selected - ", curr_sel_si),
                         container = si_group,
-                        width     = 540,
+                        width     = 600,
                         height    = 15,
                         editable  = FALSE,
                         expand    = TRUE)
