### 2018-03-24 73286b4
Fixes bug on GUI if using Combined products
reported in https://github.com/lbusett/MODIStsp/issues/118

+

changes how the compress option is saved in the json file
