### 2018-03-24 bac80b8
fix regression on options saving introduced by last commit
