### 2018-03-24 #184 @lbusett
Hi @jeffreyhanson and @karthik ,

    just wanted to ask if by any chance you had time to have a look at the implemented changes.

Asking because I am planning a CRAN release to introduce the new functionality (and - more importantly - the associated bug fixes...), but I'd rather do that at the end of the review process if possible.
