### 2018-03-31 0927542
correction in load_opts + moved argument check at beginning of MODIStsp
