+++ R/load_opts.R
@@ -85,7 +85,7 @@ load_opts <- function(opts_jsfile) {
       full_ext        = "Select MODIS Tiles",
       resampling      = "near",
       out_format      = "GTiff",
-      ts_format       = "rasterStack",
+      ts_format       = "R rasterStack",
       compress        = "None",
       nodata_change   = "No",
       scale_val       = "No",
