+++ .travis.yml
@@ -40,7 +40,6 @@ matrix:
       r_binary_packages:
         - cairoDevice
         - RGtk2
-      r_packages:
         - udunits2
         - units
         - covr
@@ -49,6 +48,5 @@ matrix:
         - leaflet
         - mapview
         - mapedit
-
 after_success:
   - Rscript -e 'library(covr);coveralls();codecov()'
