+++ .travis.yml
@@ -37,6 +37,7 @@ matrix:
             - libglib2.0-dev
             - libcurl4-openssl-dev
             - libudunits2-dev
+            - libgdal1h 
       r_binary_packages:
         - cairoDevice
         - RGtk2
