### 2018-04-06 de759f8
WIP on solving bugs on extent selection for 0.05 degrees products
Partially solves #122 by setting automatically the extent for 0.05 deg. products to -180 180 -90 90 and disabling tiles selection. Also "forces" reset of the extent on product/category changes
TODO: See if possible to avoid the GTK-CRITICAL messages (or at least hide them) .
