+++ R/gh_selprod.R
@@ -6,10 +6,11 @@
 #'
 gh_selprod <- function(h, wids, prod_opt_list, general_opts) {
   #nocov start
+
   sel_prod   <- ifelse(!is.null(gWidgets::svalue(wids$prod)),
                        gWidgets::svalue(wids$prod),
                        sel_prod)
-  
+
   sel_prodopts <- prod_opt_list[[sel_prod]]
   # Select the last version (it assumes that versions in xml file are in
   # increasing order)
@@ -26,15 +27,33 @@ gh_selprod <- function(h, wids, prod_opt_list, general_opts) {
     gWidgets::svalue(wids$sens)  <- general_opts$sensor
   }
   # On product change, automatically modify the default projection - latlon
-  # for tiled, Sinu for nontiled
-  
+  # for non-tiled, Sinu tiled
+
   if (sel_prodopts[[gWidgets::svalue(wids$vers)]]$tiled == 0) {
+
     gWidgets::enabled(tiles_group) <- FALSE
-    gWidgets::svalue(wids$proj_choice)    <- "Native"
+    gWidgets::enabled(bbox_group)  <- TRUE
+    gWidgets::svalue(wids$output_ext)   <- "Define Custom Area"
+    gWidgets::svalue(wids$proj_choice)  <- "Native"
     gWidgets::svalue(wids$output_proj4) <-
       "+init=epsg:4008 +proj=longlat +ellps=clrk66 +no_defs"
+    gWidgets::svalue(wids$output_xmin) <- -180
+    gWidgets::svalue(wids$output_xmax) <-  180
+    gWidgets::svalue(wids$output_ymin) <- -90
+    gWidgets::svalue(wids$output_ymax) <-  90
+
   } else {
-    gWidgets::enabled(tiles_group) <- TRUE
+    gWidgets::svalue(wids$output_xmin) <- ""
+    gWidgets::svalue(wids$output_xmax) <- ""
+    gWidgets::svalue(wids$output_ymin) <- ""
+    gWidgets::svalue(wids$output_ymax) <- ""
+    if (gWidgets::svalue(wids$output_ext) == "Define Custom Area") {
+      gWidgets::enabled(bbox_group)   <- TRUE
+      gWidgets::enabled(tiles_group)  <- FALSE
+    } else {
+      gWidgets::enabled(bbox_group)   <- FALSE
+      gWidgets::enabled(tiles_group)  <- TRUE
+    }
     gWidgets::svalue(wids$proj_choice)    <- "Native"
     gWidgets::svalue(wids$output_proj4) <-
       "+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs" #nolint
@@ -44,9 +63,9 @@ gh_selprod <- function(h, wids, prod_opt_list, general_opts) {
   gui_env$temp_wid_bands         <- 0
   gui_env$temp_wid_bands_indexes <- 0
   gui_env$temp_wid_bands_quality <- 0
-  
+
   cur_prodopts <- sel_prodopts[[gWidgets::svalue(wids$vers)]]
-  
+
   gWidgets::svalue(wids$sel_layers) <- "- None selected -"
   gWidgets::svalue(wids$sel_qi)     <- ifelse(
     length(cur_prodopts[["quality_bandnames"]]) == 0,
@@ -58,5 +77,6 @@ gh_selprod <- function(h, wids, prod_opt_list, general_opts) {
     "- Spectral Indexes can not be computed on this product - ",
     " - None selected - "
   )
+
   #nocov end
 }
