### 2018-04-07 07d5aa4
Finishh fix on mcd products extent
plus, unlink some temporary tif files (#123)
plus, fix travis.yml
