+++ R/gui_helpers.R
@@ -16,21 +16,21 @@ gui_update_bboxlabels <- function(bbox_out,
                                   wids,
                                   reset = FALSE) {
   #nocov start
-  
+
   if (!reset) {
     digits <- ifelse(units == "dec.degrees", 5, 1)
-    
+
     gWidgets::svalue(wids$output_xmin)  <- formatC(bbox_out[1, 1],
                                                    digits = digits,
                                                    format = "f")
-    
+
     gWidgets::svalue(wids$output_ymin) <- formatC(bbox_out[2, 1],
                                                   digits = digits,
                                                   format = "f")
     gWidgets::svalue(wids$output_xmax)  <- formatC(bbox_out[1, 2],
                                                    digits = digits,
                                                    format = "f")
-    
+
     gWidgets::svalue(wids$output_ymax) <- formatC(bbox_out[2, 2],
                                                   digits = digits,
                                                   format = "f")
@@ -41,7 +41,7 @@ gui_update_bboxlabels <- function(bbox_out,
     gWidgets::svalue(wids$output_xmax)  <- ""
     gWidgets::svalue(wids$output_ymax)  <- ""
   }
-  
+
   #nocov end
 }
 
@@ -59,12 +59,12 @@ gui_update_tiles <- function(bbox_out,
                              modis_grid,
                              wids) {
   #nocov start
-  
+
   bbox_mod  <- reproj_bbox(bbox_out,
                            curr_proj,
                            mod_proj_str,
                            enlarge = TRUE)
-  
+
   d_bbox_mod_tiled    <- raster::crop(modis_grid, raster::extent(bbox_mod))
   gWidgets::svalue(wids$start_x) <- min(d_bbox_mod_tiled$H)
   gWidgets::svalue(wids$end_x)   <- max(d_bbox_mod_tiled$H)
@@ -125,20 +125,20 @@ gui_load_options <- function(opts_jsfile,
                              prod_opt_list,
                              compress_dict) {
   #nocov start
-  
+
   # load file and reset all widgets to values found in the loaded file
   general_opts <- try(jsonlite::fromJSON(opts_jsfile), silent = TRUE)
-  
+
   # stop on error
   if (class(general_opts) == "try-error") {
     stop("Unable to read the provided JSON options file. Please check your ",
          "inputs!")
   }
-  
+
   sel_prod      <- general_opts$sel_prod
   sel_prodopts  <- prod_opt_list[[sel_prod]]
   cur_prodopts  <- sel_prodopts[[general_opts$prod_version]]
-  
+
   gWidgets::svalue(wids$cat) <- paste(
     prod_opt_list[[general_opts$sel_prod]][[general_opts$prod_version]]$cat01,
     prod_opt_list[[general_opts$sel_prod]][[general_opts$prod_version]]$cat02,
@@ -147,50 +147,50 @@ gui_load_options <- function(opts_jsfile,
   gWidgets::svalue(wids$prod) <- general_opts$sel_prod
   gWidgets::svalue(wids$vers) <- general_opts$prod_version
   gWidgets::svalue(wids$sens) <- general_opts$sensor
-  
+
   # set dummy variables holding the initial values of selected bands
   gui_env$temp_wid_bands         <- general_opts$bandsel
   gui_env$temp_wid_bands_indexes <- general_opts$indexes_bandsel
   gui_env$temp_wid_bands_quality <- general_opts$quality_bandsel
-  
-  
+
+
   curr_sel_layers <- paste(
[…]
