+++ R/MODIStsp_GUI.R
@@ -534,21 +534,33 @@ MODIStsp_GUI <- function(general_opts,
                        c("Select MODIS Tiles", "Define Custom Area")),
     handler  = function(h, ...) {
       current_sel <- gWidgets::svalue(wids$output_ext)
-      # if (sel_prodopts[[gWidgets::svalue(wids$vers)]]$tiled == 1) {
-      # On "Select MODIS Tiles" selection, disable the bbox fields
-      if (current_sel == "Select MODIS Tiles") {
-        gWidgets::enabled(bbox_group)      <- FALSE
-        gWidgets::enabled(tiles_from_bbox) <- FALSE
-        gWidgets::enabled(bbox_from_file)  <- FALSE
-        gWidgets::enabled(tiles_group)     <- TRUE
+      sel_prod   <- ifelse(!is.null(gWidgets::svalue(wids$prod)),
+                           gWidgets::svalue(wids$prod),
+                           sel_prod)
+      sel_prodopts <- prod_opt_list[[sel_prod]]
+      if (sel_prodopts[[gWidgets::svalue(wids$vers)]]$tiled == 1) {
+        # On "Select MODIS Tiles" selection, disable the bbox fields
+        if (current_sel == "Select MODIS Tiles") {
+          gWidgets::enabled(bbox_group)      <- FALSE
+          gWidgets::enabled(tiles_from_bbox) <- FALSE
+          gWidgets::enabled(bbox_from_file)  <- FALSE
+          gWidgets::enabled(tiles_group)     <- TRUE
+        } else {
+          # On "Custom  Area" selection, enable the bbox fields
+          gWidgets::enabled(bbox_group)      <- TRUE
+          gWidgets::enabled(tiles_from_bbox) <- TRUE
+          gWidgets::enabled(bbox_from_file)  <- TRUE
+          gWidgets::enabled(tiles_group)     <- FALSE
+        }
       } else {
-        # On "Custom  Area" selection, enable the bbox fields
-        gWidgets::enabled(bbox_group)      <- TRUE
-        gWidgets::enabled(tiles_from_bbox) <- TRUE
-        gWidgets::enabled(bbox_from_file)  <- TRUE
-        gWidgets::enabled(tiles_group)     <- FALSE
+        if (current_sel == "Select MODIS Tiles") {
+          gWidgets::svalue(wids$output_ext)  <- "Define Custom Area"
+          gWidgets::enabled(bbox_group)      <- TRUE
+          gWidgets::enabled(tiles_from_bbox) <- TRUE
+          gWidgets::enabled(bbox_from_file)  <- TRUE
+          gWidgets::enabled(tiles_group)     <- FALSE
+        }
       }
-      # }
     }
   )
 
