+++ R/MODIStsp.R
@@ -79,7 +79,7 @@
 #' # Running the tool using the settings previously saved in a specific option file
 #' # and specifying the extent from a spatial file allows to re-use the same
 #' # processing settings to perform download and reprocessing on a different area
-#'  
+#'
 #' options_file <- system.file("testdata/test_MOD13A2.json", package = "MODIStsp")
 #' spatial_file <- system.file("testdata/lakeshapes/garda_lake.shp", package = "MODIStsp")
 #' MODIStsp(gui = FALSE, options_file = options_file,
@@ -91,15 +91,15 @@
 #' # specific options file and specifying each time the extent from a different
 #' # spatial file (e.g., to perform the same processing on several extents)
 #'
-#' extent_list  <- c(system.file("testdata/lakeshapes/garda_lake.shp", 
+#' extent_list  <- c(system.file("testdata/lakeshapes/garda_lake.shp",
 #'                               package = "MODIStsp"),
-#'                   system.file("testdata/lakeshapes/iseo_lake.shp", 
+#'                   system.file("testdata/lakeshapes/iseo_lake.shp",
 #'                               package = "MODIStsp"))
 #' extent_list
-#' 
+#'
 #' # Note that you can also put all your extent files in a specific folder and
 #' # create the extent list using for example.
-#' # extent_list = list.files(system.file("testdata/lakeshapes/", package = "MODIStsp"), 
+#' # extent_list = list.files(system.file("testdata/lakeshapes/", package = "MODIStsp"),
 #' #                          full.names = TRUE, "\\.shp$")
 #'
 #' options_file <- system.file("testdata/test_MOD13A2.json", package = "MODIStsp")
@@ -108,12 +108,12 @@
 #'            spatial_file_path = single_shape, verbose = FALSE )
 #' }
 #'
-#' # output files are places in separate folders: 
+#' # output files are places in separate folders:
 #' outfiles_garda <- list.files(file.path(tempdir(), "MODIStsp/garda_lake/VI_16Days_1Km_v6/EVI"),
 #'            full.names = TRUE)
-#' library(raster)            
+#' library(raster)
 #' plot(raster(outfiles_garda[1]))
-#' 
+#'
 #' outfiles_iseo <- list.files(file.path(tempdir(), "MODIStsp/iseo_lake/VI_16Days_1Km_v6/EVI"),
 #'            full.names = TRUE)
 #' plot(raster(outfiles_iseo[1]))
@@ -132,43 +132,43 @@ MODIStsp <- function(gui               = TRUE,
   # Make so that "raster" functions does not automatically add extensions on
   # output files. This is automatically reset to TRUE at the end of the session
   raster::rasterOptions(setfileext = FALSE)
-  
+
   #   _________________________________________________________________________
   #   check arguments                                                       ####
-  
+
   if (is.null(options_file) & gui == FALSE) {
     stop("You need to provide a valid `.json` options file to run MODIStsp",
          " in non-interactive mode. \n",
          "Please provide a valid \"options_file\" path or run ",
          "with gui=TRUE to create and save one.")
   }
-  
+
   if (!is.null(options_file)) {
     if (!file.exists(options_file))
       stop("The specified `.json` options file was not found. \n",
            "Please provide a valid \"options_file\" path or run ",
            "without specifying one to create and save one.")
   }
-  
-  
+
+
   # On first execution (or if the file is not found), ask the user permission
   # for saving a options file in "your-R-library/MODIStsp/ExtData/Previous"
-  
+
   if (interactive()) {
     permission <- ask_permission()
   } else {
     permission <- FALSE
   }
-  
-  
+
+
   #   __________________________________________________________________________
   #   Initialize processing                                                 ####
-  
+
[…]
