### 2018-04-11 4be12e6
Updated NEWS on decommissioning of v005
