+++ README.Rmd
@@ -55,6 +55,14 @@ L. Busetto, L. Ranghetti (2016) MODIStsp: An R package for automatic preprocessi
 
 ## Important News ! 
 
+- 11/04/2018 - Due to new NASA Policies the MODIS FTP servers were shut down 
+starting, April 2, 2018. **FTP download is therefore no longer working** and will
+be removed in the next MODIStsp version!
+
+- 11/04/2018 - [**Decommissioning of MODIS Version 5 Land Data Products**]( https://lpdaac.usgs.gov/about/news_archive/decommissioning_modis_version_5_land_data_products_april_9_2018_second_notice). As per NASA notice above, MODIS v005 products are going to be 
+decommisioned, and will soon be no longer available for download. Support for those
+products will be removed in the next MODIStsp version!.
+
 - 11/08/2017 - MODIStp 1.3.3 was released today. It provides improvements in
 processing speed, as well as the usual bug fixes
 (thanks to all the users that signalled problems !). Check the [Release Notes](https://github.com/lbusett/MODIStsp/releases/tag/v1.3.3) for further details !
