### 2018-04-11 c222a68
improve cleanup of temporary files
- creates a tmpdir/mstp_temp folder where temporaries are stores
- unlinks the folder on exits
(fixes #123)
