+++ R/get_mod_dirs.R
@@ -1,19 +1,16 @@
-#' @title Get list of MODIS data folders from http/ftp server
+#' @title Get list of MODIS data folders from http server
 #' @description Accessory function to get the full list of directories on the
 #'  lpdaac http site containing data included in the time range selected for
 #'  processing (modified after Barry Rowlingson function):
 #' @param http `character` http site on lpdaac corresponding to the selected MODIS
 #'   product
-#' @param ftp `character` ftp site corresponding to the selected MODIS product
-#' @param download_server `character ["http" | "ftp" | "offline"]` download service
-#'   to be used; if NA, the script tries to download with http, using ftp if the
-#'   download fails (Used in non-interactive execution to allow trying to use
-#'   ftp if http fails)
+#' @param download_server `character ["http" | "offline"]` download service
+#'   to be used; if NA, the script tries to download with http.
 #' @param user `character` username for earthdata http server
 #' @param password `character` password for earthdata http server
 #' @param yy `character` Year for which the folder containing HDF images are to
 #'  be identified
-#' @param n_retries `numeric` number of times the access to the http/ftp server
+#' @param n_retries `numeric` number of times the access to the http server
 #'   should be retried in case of error before quitting, Default: 20
 #' @param gui `logical`` indicates if processing was called from the GUI
 #'   environment or not. If not, processing messages are sent to a log file
@@ -21,7 +18,7 @@
 #' @param out_folder_mod  `character` output folder for MODIS HDF storage
 #' @param .Platform `character` os platform (from call to .Platform)
 #' @return `character arraty` listing all available folders (a.k.a. dates) for
-#'   the requested MODIS product on lpdaac http or ftp archive, for the years
+#'   the requested MODIS product on lpdaac http archive, for the years
 #'   included in the time range selected for processing.
 #' @author Original code by Babak Naimi (\code{.getModisList}, in
 #' \href{http://r-gis.net/?q=ModisDownload}{ModisDownload.R})
@@ -34,7 +31,8 @@
 #' @importFrom httr GET authenticate timeout content
 #' @importFrom stringr str_extract str_split str_split_fixed str_sub
 
-get_mod_dirs <- function(http, ftp, download_server,
+get_mod_dirs <- function(http,
+                         download_server,
                          user, password,
                          yy,
                          n_retries,
@@ -67,7 +65,7 @@ get_mod_dirs <- function(http, ftp, download_server,
       if (gui) {
         #nocov start
         switch <- gWidgets::gconfirm(
-          "http server seems to be down! Do you want to switch to ftp?",
+          "http server seems to be down! Do you want to retry?",
           icon = "question")
         if (!switch) {
           stop("You selected to abort processing. Goodbye!")
@@ -101,48 +99,6 @@ get_mod_dirs <- function(http, ftp, download_server,
   }
   success <- FALSE
   # retrieve processign dates in case of "ftp" download ----
-  if (download_server == "ftp") {
-
-    while (!success) {
-      # send request to server
-      year_ftp <- paste0(ftp, yy, "/")
-      response <- try(suppressWarnings(httr::RETRY("GET",
-                                  year_ftp,
-                                  times = n_retries,
-                                  pause_base = 0.1,
-                                  pause_cap = 3,
-                                  quiet = FALSE)))
-      # On interactive execution, after n_retries attempt ask if quit or ----
-      # retry
-      if (class(response) == "try-error") {
-        if (gui) {
-          #nocov start
-          confirm <- gWidgets::gconfirm(
-            "ftp server seems to be down! Do you want to retry?",
-            icon = "question")
-          if (!confirm) stop("You selected to abort processing. Goodbye!")
-          #nocov end
-        } else {
[…]
