### 2018-05-02 4af5328
Finalized ftp removal
