+++ R/MODIStsp_extract.R
@@ -83,7 +83,7 @@
 #' @importFrom tools file_path_sans_ext
 #' @importFrom gdalUtils gdal_rasterize
 #' @examples
-#' \dontrun {
+#' \dontrun{
 #' # Extract average and standard deviation values from a rts object created by
 #' # MODIStsp for each polygon of a shapefile, for each date in the period
 #' # between 2001-01-01 and 2014-12-31
