### 2018-05-07 1093b7d
removed covr from binary installs
