+++ .travis.yml
@@ -40,6 +40,5 @@ matrix:
       r_binary_packages:
         - cairoDevice
         - RGtk2
-        - covr
 after_success:
   - Rscript -e 'library(covr);coveralls();codecov()'
