### 2018-05-07 979bc76
WIP on new products in list
Added MOD44W, corrections on MOD14A2, added MCD12C1 v6, MCD12Q1v6, MCD18A1 v6, MCD18A2 v6.

ALL TO BE CHECKED! (strange mcd12c1 format, floating point...)
