### 2018-05-07 1d49b09
removed ftp links from xml
added logo in vignettes
