+++ vignettes/installation.Rmd
@@ -99,7 +99,7 @@ ________________________________________________________________________________
 ## <i class="fa fa-apple" aria-hidden="true"></i> Installing on Mac
 
 **NOTE**: The following  installation notes should be valid for MODIStsp installation 
-on R 3.4.0 and above with Mac OSX Sierra. They were mainly taken (i.e., blatantly copied...) from: https://zhiyzuo.github.io/installation-rattle/**. Thanks to Zhiya Zuo for providing this!
+on R 3.4.0 and above with Mac OSX Sierra. They were mainly taken (i.e., blatantly copied...) from: https://zhiyzuo.github.io/installation-rattle/. Thanks to Zhiya Zuo for providing this!
 
 To properly install `MODIStsp` you will need to first install package `RGTk2`. This
 is a somehow difficult operation. The following instructions should help:
