### 2018-05-24 bc1a85f
Re-add covr installation
