### 2018-05-24 dbd0547
Bump version, update website, update news
