### 2018-05-25 9952b9a
Updated cran-comments.md
