### 2018-05-25 #184 @lbusett
@karthik

I finally managed to put out a new version (1.3.4), including changes required to address changes in NASA policies. The new version is available on github, and I plan to upload it to CRAN soon.

I'd like therefore to ask:

1) What are the next steps for approval?
2) Can I proceed with CRAN release or do you prefer I wait for the end of the review/approval process?

Lorenzo
