### 2018-05-26 d841ee1
Hyperlink DOIs against preferred resolver
