+++ README.Rmd
@@ -17,7 +17,7 @@ knitr::opts_chunk$set(
 [![](http://cranlogs.r-pkg.org/badges/MODIStsp)](https://cran.r-project.org/package=MODIStsp)
 [![Travis-CI Build Status](https://travis-ci.org/lbusett/MODIStsp.svg?branch=master)](https://travis-ci.org/lbusett/MODIStsp)
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/lbusett/MODIStsp?branch=master&svg=true)](https://ci.appveyor.com/project/lbusett/MODIStsp)
-[![DOI](http://zenodo.org/badge/DOI/10.5281/zenodo.290683.svg)](http://doi.org/10.5281/zenodo.290683)
+[![DOI](http://zenodo.org/badge/DOI/10.5281/zenodo.290683.svg)](https://doi.org/10.5281/zenodo.290683)
 [![Coverage Status](http://img.shields.io/codecov/c/github/lbusett/MODIStsp/master.svg)](http://codecov.io/github/lbusett/MODIStsp?branch=master)
 
 MODIStsp  <img src="man/figures/logo.png" width="147" height="170" align="right" />
@@ -50,7 +50,7 @@ To cite MODIStsp please use:
 
 L. Busetto, L. Ranghetti (2016) MODIStsp: An R package for automatic preprocessing of MODIS
   Land Products time series, Computers & Geosciences, Volume 97, Pages
-  40-48, ISSN 0098-3004, http://dx.doi.org/10.1016/j.cageo.2016.08.020, URL: https://github.com/lbusett/MODIStsp. 
+  40-48, ISSN 0098-3004, https://doi.org/10.1016/j.cageo.2016.08.020, URL: https://github.com/lbusett/MODIStsp. 
   
 
 ## Important News ! 
