+++ index.Rmd
@@ -91,7 +91,7 @@ To cite MODIStsp in publications, please use:
 
 L. Busetto, L. Ranghetti (2016) MODIStsp: An R package for automatic preprocessing of MODIS
   Land Products time series, Computers & Geosciences, Volume 97, Pages
-  40-48, ISSN 0098-3004, http://dx.doi.org/10.1016/j.cageo.2016.08.020, URL: https://github.com/lbusett/MODIStsp. 
+  40-48, ISSN 0098-3004, https://doi.org/10.1016/j.cageo.2016.08.020, URL: https://github.com/lbusett/MODIStsp. 
   
 A BibTeX entry for LaTeX users is:
 
