### 2018-06-13 #184 @lwasser
woo hoo! great work @lbusett ... seems like it's ready for final acceptance into ropensci by @karthik  ?
