### 2018-06-20 77f5693
Fixes a bug leading to errors in processing extent
When switching products with different Native projection (4008 vs sinusoidal), the projection string was not properly updated.
