### 2018-06-20 ffb44e8
Fixes a bug leading to errors in processing extent
When switching products with different Native projection (4008 vs sinusoidal), the projection string was not properly updated.
