### 2018-06-20 532677c
Various changes for 1.3.5 submission
- skipped test_addindex on CRAN and implemented request for permission
- updated readmes, description, etcetera
- bumped version
- modified travis.yml to test on R devel
