+++ README.Rmd
@@ -17,7 +17,7 @@ knitr::opts_chunk$set(
 [![](http://cranlogs.r-pkg.org/badges/MODIStsp)](https://cran.r-project.org/package=MODIStsp)
 [![Travis-CI Build Status](https://travis-ci.org/lbusett/MODIStsp.svg?branch=master)](https://travis-ci.org/lbusett/MODIStsp)
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/lbusett/MODIStsp?branch=master&svg=true)](https://ci.appveyor.com/project/lbusett/MODIStsp)
-[![DOI](http://zenodo.org/badge/DOI/10.5281/zenodo.290683.svg)](http://doi.org/10.5281/zenodo.290683)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1286981.svg)](https://doi.org/10.5281/zenodo.1286981)
 [![Coverage Status](http://img.shields.io/codecov/c/github/lbusett/MODIStsp/master.svg)](http://codecov.io/github/lbusett/MODIStsp?branch=master)
 
 MODIStsp  <img src="man/figures/logo.png" width="147" height="170" align="right" />
@@ -39,9 +39,11 @@ execution outside an "R" environment is also possible, allowing to use scheduled
 execution of MODIStsp to automatically update time series related to a MODIS
 product and extent whenever a new image is available. 
 
-_MODIStsp is developed and maintained by L.Busetto and L.Ranghetti, from the
-Institute of Remote Sensing of Environment - National Research Council - Italy 
-(CNR-IREA)_
+<img src="man/figures/logo_irea.png" height="60" align="left" /> 
+
+___MODIStsp is developed and maintained by L.Busetto and L.Ranghetti, from the
+[Institute of Remote Sensing of Environment](http://www.irea.cnr.it/en/) - National Research Council - Italy 
+(CNR-IREA)___
 
 ## Citation
   
@@ -54,6 +56,8 @@ L. Busetto, L. Ranghetti (2016) MODIStsp: An R package for automatic preprocessi
   
 
 ## Important News ! 
+
+- 20/062018 - MODIStsp v. 1.3.5 is out. Check out the [Release Notes](https://github.com/lbusett/MODIStsp/releases/tag/v1.3.5) for further details !
 
 - 11/04/2018 - Due to new NASA Policies the MODIS FTP servers were shut down 
 starting, April 2, 2018. **FTP download is therefore no longer working** and will
