+++ README.md
@@ -1,110 +1,181 @@
 
--   [MODIStsp](#modistsp)
-    -   [Citation](#citation)
-    -   [Important News !](#important-news)
-    -   [Problems and Issues](#problems-and-issues)
-    -   [<i class="fa fa-desktop" aria-hidden="true"></i> System Requirements](#system-requirements)
--   [Installation Instructions](#installation-instructions)
-    -   [<i class="fa fa-windows" aria-hidden="true"></i> Installing on Windows](#installing-on-windows)
-    -   [<i class="fa fa-linux" aria-hidden="true"></i> Installing on Linux Systems](#installing-on-linux-systems)
-    -   [<i class="fa fa-apple" aria-hidden="true"></i> Installing on Mac](#installing-on-mac)
--   [Usage](#usage)
--   [Code of Conduct](#code-of-conduct)
+[![](https://www.r-pkg.org/badges/version-ago/MODIStsp)](https://cran.r-project.org/package=MODIStsp)
+[![](http://cranlogs.r-pkg.org/badges/MODIStsp)](https://cran.r-project.org/package=MODIStsp)
+[![Travis-CI Build
+Status](https://travis-ci.org/lbusett/MODIStsp.svg?branch=master)](https://travis-ci.org/lbusett/MODIStsp)
+[![AppVeyor Build
+Status](https://ci.appveyor.com/api/projects/status/github/lbusett/MODIStsp?branch=master&svg=true)](https://ci.appveyor.com/project/lbusett/MODIStsp)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1286981.svg)](https://doi.org/10.5281/zenodo.1286981)
+[![Coverage
+Status](http://img.shields.io/codecov/c/github/lbusett/MODIStsp/master.svg)](http://codecov.io/github/lbusett/MODIStsp?branch=master)
 
-[![](https://www.r-pkg.org/badges/version-ago/MODIStsp)](https://cran.r-project.org/package=MODIStsp) [![](http://cranlogs.r-pkg.org/badges/MODIStsp)](https://cran.r-project.org/package=MODIStsp) [![Travis-CI Build Status](https://travis-ci.org/lbusett/MODIStsp.svg?branch=master)](https://travis-ci.org/lbusett/MODIStsp) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/lbusett/MODIStsp?branch=master&svg=true)](https://ci.appveyor.com/project/lbusett/MODIStsp) [![DOI](http://zenodo.org/badge/DOI/10.5281/zenodo.290683.svg)](http://doi.org/10.5281/zenodo.290683) [![Coverage Status](http://img.shields.io/codecov/c/github/lbusett/MODIStsp/master.svg)](http://codecov.io/github/lbusett/MODIStsp?branch=master)
+# MODIStsp <img src="man/figures/logo.png" width="147" height="170" align="right" />
 
-MODIStsp
-========
+MODIStsp is a “R” package devoted to automatizing the creation of time
+series of rasters derived from MODIS Land Products data. MODIStsp allows
+to perform several preprocessing steps (e.g., download, mosaicing,
+reprojection and resize) on MODIS data available within a given time
+period. Users have the ability to select which specific layers of the
+original MODIS HDF files they want to process. They also can select
+which additional Quality Indicators should be extracted from the
+aggregated MODIS Quality Assurance layers and, in the case of Surface
+Reflectance products, which Spectral Indexes should be computed from the
+original reflectance bands. For each output layer, outputs are saved as
+single-band raster files corresponding to each available acquisition
+date. Virtual files allowing access to the entire time series as a
+single file can be also created. All processing parameters can be easily
[…]
