+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: MODIStsp
 Title: A Tool for Automating Download and Preprocessing of MODIS Land Products
     Data
 Type: Package
-Version: 1.3.4
+Version: 1.3.5
 Authors@R: c(person("Lorenzo", "Busetto", email = "lbusett@gmail.com", role = c("aut", "cre"),
     comment = c(ORCID = '0000-0001-9634-6038')), 
     person("Luigi", "Ranghetti", email = "ranghetti.l@irea.cnr.it", role = c("aut")))
