### 2018-06-20 c87a12b
updated news and travis.yml
