+++ index.md
@@ -1,61 +1,138 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-[![](https://www.r-pkg.org/badges/version-ago/MODIStsp)](http://cran.rstudio.com/web/packages/MODIStsp/index.html) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.290683.svg)](https://doi.org/10.5281/zenodo.290683) [![Downloads](http://cranlogs.r-pkg.org/badges/grand-total/MODIStsp?color=red)](http://cran.rstudio.com/web/packages/MODIStsp/index.html) [![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](http://www.gnu.org/licenses/gpl-3.0) [![Travis-CI Build Status](https://travis-ci.org/lbusett/MODIStsp.svg?branch=master)](https://travis-ci.org/lbusett/MODIStsp) [![Coverage Status](https://img.shields.io/codecov/c/github/lbusett/MODIStsp/master.svg)](https://codecov.io/github/lbusett/MODIStsp?branch=master)
 
-<i class="fa fa-globe" aria-hidden="true"></i> MODIStsp: a "R" package for preprocessing MODIS time series <img src="docs/logo.png" width="147" height="170" align="right" />
-=============================================================================================================================================================================
+[![](https://www.r-pkg.org/badges/version-ago/MODIStsp)](http://cran.rstudio.com/web/packages/MODIStsp/index.html)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.290683.svg)](https://doi.org/10.5281/zenodo.290683)
+[![Downloads](http://cranlogs.r-pkg.org/badges/grand-total/MODIStsp?color=red)](http://cran.rstudio.com/web/packages/MODIStsp/index.html)
+[![License: GPL
+v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](http://www.gnu.org/licenses/gpl-3.0)
+[![Travis-CI Build
+Status](https://travis-ci.org/lbusett/MODIStsp.svg?branch=master)](https://travis-ci.org/lbusett/MODIStsp)
+[![Coverage
+Status](https://img.shields.io/codecov/c/github/lbusett/MODIStsp/master.svg)](https://codecov.io/github/lbusett/MODIStsp?branch=master)
 
-**MODIStsp** is a "R" package devoted to **automatizing the creation of time series of raster images derived from MODIS Land Products data**.
+# <i class="fa fa-globe" aria-hidden="true"></i> MODIStsp <img src="man/figures/logo.png" width="100" height="100" align="right"/>
 
-**MODIStsp** allows to perform several preprocessing steps *(e.g., download, mosaicing, reprojection, resize, data extraction)* on MODIS data available within a given time period. Users have the ability to select which specific layers of the original MODIS HDF files they want to process. They also can select which additional **Quality Indicators** should be extracted from the aggregated MODIS Quality Assurance layers and, in the case of Surface Reflectance products, which **Spectral Indexes** should be computed from the original reflectance bands.
+**MODIStsp** is a “R” package devoted to **automatizing the creation of
+time series of raster images derived from MODIS Land Products data**.
 
-All processing parameters can be easily selected with a **powerful and user-friendly GUI**, although non-interactive execution exploiting a previously created Options File is possible. Stand-alone execution outside an "R" environment is also possible, allowing to use scheduled execution of MODIStsp to automatically update time series related to a MODIS product and extent whenever a new image is available.
+**MODIStsp** allows to perform several preprocessing steps *(e.g.,
+download, mosaicing, reprojection, resize, data extraction)* on MODIS
+data available within a given time period. Users have the ability to
[…]
