+++ man/MODIStsp_addindex.Rd
@@ -55,8 +55,8 @@ Default: retrieved from package installation folder}
 }
 \value{
 The function is called for its side effects. On success, the
-MODIStsp_Previous.RData file is modified so to allow computation of the
-additional indexes.
+MODIStsp_Previous.json or the opts_jfile specified by the user
+is modified so to allow computation of the additional indexes.
 }
 \description{
 Function used to add a user-defined Spectral Index to the
