+++ NAMESPACE
@@ -11,6 +11,7 @@ export(MODIStsp_reset_options)
 export(MODIStsp_resetindexes)
 export(check_proj4string)
 export(install_MODIStsp_launcher)
+import(gWidgetsRGtk2)
 importFrom(bitops,bitAnd)
 importFrom(bitops,bitShiftR)
 importFrom(data.table,data.table)
@@ -91,6 +92,7 @@ importFrom(rgdal,getGDALVersionInfo)
 importFrom(rgdal,ogrInfo)
 importFrom(rgdal,readOGR)
 importFrom(rgdal,writeOGR)
+importFrom(rgeos,gCentroid)
 importFrom(sp,CRS)
 importFrom(sp,Polygon)
 importFrom(sp,Polygons)
