### 2018-06-20 eb703df
Removed "library" calls from tests
