### 2018-06-21 94bcb8c
Final changes for v 1.3.5
