+++ cran-comments.md
@@ -11,7 +11,7 @@ the MODIStsp_previous.json file
 * ubuntu 18.04 on local install, R 3.4.4
 * ubuntu 14.04 on travis-ci (https://travis-ci.org/lbusett/MODIStsp/builds/394612258)
   (R-devel, R-release)
-* win-builder (R-devel, R-release) (r-release and devel)
+* win-builder (R-devel, R-release)
 * windows 10 on local install, R 3.5
 
 ## R CMD check results
@@ -21,6 +21,18 @@ There were no ERRORs, WARNINGs or NOTEs
 
 MODIStsp 1.3.4
 ================
+
+## Resubmission 
+ 
+This is a resubmission. In this version I have: 
+ 
+* Re-added package "testthat" in Suggests and removed an unnecessary call to  
+library(testthat) to avoid the following WARNING on Debian: 
+ 
+  "* checking for unstated dependencies in ‘tests’ ... WARNING 
+  '::' or ':::' import not declared from: ‘testthat’ 
+  'library' or 'require' call not declared from: ‘testthat’" 
+  
 ## Test environments
 
 * ubuntu 18.04 on local install, R 3.4.4
