### 2018-07-09 df0e6ff
Fixes #133
Fixes bug leading to error if only "Aqua" download is selected.
