### 2018-07-10 174b7a8
Update website for 1.3.6 release
