+++ README.Rmd
@@ -59,6 +59,8 @@ L. Busetto, L. Ranghetti (2016) MODIStsp: An R package for automatic preprocessi
 
 ## Important News ! 
 
+- 10/07/2018 - MODIStsp v. 1.3.6 is out. Check out the [Release Notes](https://github.com/lbusett/MODIStsp/releases/tag/v1.3.5) for further details !
+
 - 20/06/2018 - MODIStsp v. 1.3.5 is out. Check out the [Release Notes](https://github.com/lbusett/MODIStsp/releases/tag/v1.3.5) for further details !
 
 - 11/04/2018 - Due to new NASA Policies the MODIS FTP servers were shut down 
