### 2018-07-10 ebfc058
update readme and news
