### 2018-08-02 #184 @lwasser
... i was just talking to someone about it and looking for this on  ropensci and didn't see it and was wondering the same thing? @maelle  @karthik just curious.
