### 2018-08-06 ff1be48
final changes to transfer repo to ropensci
