### 2018-08-07 dd73520
updated rbuildignore to prevent previous options file to be installed
