+++ tests/testthat/test_MODIStsp.R
@@ -232,7 +232,9 @@ testthat::test_that("Tests on MODIStsp", {
     gui = FALSE, n_retries = 2, verbose = FALSE)
 
   out_files_rts <- list.files(
-    file.path(tempdir(), "MODIStsp/Snow_cov_mnt_005dg_v6/Time_Series/RData/Aqua/SN_COV_MNT"),
+    file.path(
+      tempdir(),
+      "MODIStsp/Snow_cov_mnt_005dg_v6/Time_Series/RData/Aqua/SN_COV_MNT"),
     pattern = "\\.RData$", recursive = TRUE, full.names = TRUE)
 
   #check that rts files are properly created
