+++ index.utf8.md
@@ -1,112 +1,0 @@
----
-output: github_document
----
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
-
-
-[![](https://www.r-pkg.org/badges/version-ago/MODIStsp)](http://cran.rstudio.com/web/packages/MODIStsp/index.html)
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.290683.svg)](https://doi.org/10.5281/zenodo.290683)
-[![Downloads](http://cranlogs.r-pkg.org/badges/grand-total/MODIStsp?color=red)](http://cran.rstudio.com/web/packages/MODIStsp/index.html)
-[![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](http://www.gnu.org/licenses/gpl-3.0)
-[![Travis-CI Build Status](https://travis-ci.org/lbusett/MODIStsp.svg?branch=master)](https://travis-ci.org/lbusett/MODIStsp)
-[![Coverage Status](https://img.shields.io/codecov/c/github/lbusett/MODIStsp/master.svg)](https://codecov.io/github/lbusett/MODIStsp?branch=master)
-
-
-# <i class="fa fa-globe" aria-hidden="true"></i> MODIStsp <img src="man/figures/logo.png" width="100" height="100" align="right"/>
-
-**MODIStsp** is a "R" package devoted to **automatizing the creation of time series of raster images derived from MODIS Land Products data**. 
-
-**MODIStsp** allows to perform several preprocessing steps _(e.g., download, mosaicing, 
-reprojection, resize, data extraction)_ on MODIS data available within a given time period.
-Users have the ability to select which specific layers of the original MODIS HDF files they 
-want to process. They also can select which additional **Quality Indicators** should be 
-extracted from the aggregated MODIS Quality Assurance layers and, in the case of Surface 
-Reflectance products, which **Spectral Indexes** should be computed from the original reflectance 
-bands. 
-
-All processing parameters can be easily selected with a **powerful and user-friendly GUI**, 
-although non-interactive execution exploiting a previously created Options File is possible. 
-Stand-alone execution outside an "R" environment is also possible, allowing to use scheduled 
-execution of MODIStsp to automatically update time series related to a MODIS product and extent 
-whenever a new image is available. 
-
-For each output layer, outputs are saved as **single-band raster** files corresponding to 
-each available acquisition date. **Virtual files** allowing access to the entire time series 
-as a single file can be also created. 
-
-
-<a href="http://www.irea.cnr.it/en/"> <img src="man/figures/logo_irea.png" height="40" align="left" /></a>
-
-
-___`MODIStsp` is developed and maintained by Lorenzo Busetto and Luigi Ranghetti,
-[Institute of Remote Sensing of Environment](http://www.irea.cnr.it/en/) - National Research Council - Italy (CNR-IREA)___
-
-
-____________________________________________________________________________________
-
-## <i class="fa fa-cog" aria-hidden="true"></i> Getting Started
-
-- __To install `MODIStsp`__, please follow instructions reported [here](articles/installation.html),
-both for [<i class="fa fa-windows" aria-hidden="true"></i>](articles/installation.html#installing-on-windows) , [<i class="fa fa-linux" aria-hidden="true"></i>](articles/installation.html#installing-on-linux-systems) and [<i class="fa fa-apple" aria-hidden="true"></i>](articles/installation.html#installing-on-mac)
-
-- `MODIStsp` can be used either in [interactive mode](articles/interactive_execution.html) 
-exploiting its user-friendly GUI, or in [non-interactive mode](articles/noninteractive_execution.html) 
[…]
