### 2018-08-07 e727285
re-updated all links for ropensci
