+++ README.Rmd
@@ -15,13 +15,13 @@ knitr::opts_chunk$set(
 ```
 [![](https://www.r-pkg.org/badges/version-ago/MODIStsp)](https://cran.r-project.org/package=MODIStsp)
 [![](http://cranlogs.r-pkg.org/badges/MODIStsp)](https://cran.r-project.org/package=MODIStsp)
-[![Travis-CI Build Status](https://travis-ci.org/lbusett/MODIStsp.svg?branch=master)](https://travis-ci.org/lbusett/MODIStsp)
+[![Travis-CI Build Status](https://travis-ci.org/ropensci/MODIStsp.svg?branch=master)](https://travis-ci.org/ropensci/MODIStsp)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1286981.svg)](https://doi.org/10.5281/zenodo.1286981)
-[![Coverage Status](http://img.shields.io/codecov/c/github/lbusett/MODIStsp/master.svg)](http://codecov.io/github/lbusett/MODIStsp?branch=master)
+[![Coverage Status](http://img.shields.io/codecov/c/github/ropensci/MODIStsp/master.svg)](http://codecov.io/github/ropensci/MODIStsp?branch=master)
 
 MODIStsp  <img src="man/figures/logo.png" width="147" height="170" align="right" />
 =========
-[MODIStsp](http://lbusett.github.io/MODIStsp/) is a "R" package devoted to
+[MODIStsp](http://ropensci.github.io/MODIStsp/) is a "R" package devoted to
 automatizing the creation of time series of rasters derived from MODIS Land Products
 data.
 MODIStsp allows to perform several preprocessing steps (e.g., download, mosaicing,
@@ -39,7 +39,7 @@ File is possible. Stand-alone execution outside an "R" environment is also possi
 allowing to use scheduled execution of MODIStsp to automatically update time series 
 related to a MODIS product and extent whenever a new image is available. 
 
-__For more information, documentation and examples of use, please see MODIStsp website at [lbusett.github.io/MODIStsp](http://lbusett.github.io/MODIStsp/)__ 
+__For more information, documentation and examples of use, please see MODIStsp website at [ropensci.github.io/MODIStsp](http://ropensci.github.io/MODIStsp/)__ 
 
 <a href="http://www.irea.cnr.it/en/"> <img src="man/figures/logo_irea.png" height="40" align="left" /></a> 
 
@@ -53,14 +53,19 @@ To cite `MODIStsp` please use:
 
 L. Busetto, L. Ranghetti (2016) MODIStsp: An R package for automatic preprocessing of MODIS
   Land Products time series, Computers & Geosciences, Volume 97, Pages
-  40-48, ISSN 0098-3004, http://dx.doi.org/10.1016/j.cageo.2016.08.020, URL: https://github.com/lbusett/MODIStsp. 
+  40-48, ISSN 0098-3004, http://dx.doi.org/10.1016/j.cageo.2016.08.020, URL: https://github.com/ropensci/MODIStsp. 
   
 
 ## Important News ! 
 
-- 10/07/2018 - MODIStsp v. 1.3.6 is out. Check out the [Release Notes](https://github.com/lbusett/MODIStsp/releases/tag/v1.3.6) for further details !
+- 07/08/2018 - We are glad to report that MODIStsp is now included in the 
+[rOpenSci](https://ropensci.org/about/) packages' ecosystem. We thank reviewers
+Leah Wasser and Jeffrey Hanson for their valuable reviews, which helped us to 
+further improve the package!
 
-- 20/06/2018 - MODIStsp v. 1.3.5 is out. Check out the [Release Notes](https://github.com/lbusett/MODIStsp/releases/tag/v1.3.5) for further details !
[…]
