+++ DESCRIPTION
@@ -57,8 +57,8 @@ Suggests:
     mapedit (>= 0.4.1)
 SystemRequirements: Cairo >= 1.0.0, ATK (>= 1.10.0), Pango (>= 1.10.0), GTK+ (>=
     2.8.0), GLib (>= 2.8.0), Curl, GDAL (>= 1.6.3), PROJ.4 (>= 4.4.9)
-URL: https://github.com/lbusett/MODIStsp
-BugReports: https://github.com/lbusett/MODIStsp/issues
+URL: https://github.com/ropensci/MODIStsp
+BugReports: https://github.com/ropensci/MODIStsp/issues
 LazyData: true
 VignetteBuilder: knitr
 RoxygenNote: 6.1.0

[…]
