+++ docs/articles/analyze.md
@@ -1,241 +1,0 @@
-Accessing and Analyzing Processed Data from R
-================
-
-Accessing the processed time series from R
-------------------------------------------
-
-Preprocessed MODIS data can be retrieved within `R` either by accessing the single-date raster files, or by loading the saved `RasterStack` objects (see [HERE](articles/output.html) for a description of `MODIStsp` output folder structure and naming conventions).
-
-To test this functionality, you can run the following example:
-
-``` r
-library(MODIStsp)
-opts_file <- system.file("testdata/test_extract.json", package = "MODIStsp")
-MODIStsp(options_file = opts_file, gui = TRUE)
-```
-
-, specifying a folder on your PC to store the outputs from the GUI.
-
-This will download a yearly time series of MODIS NDVI data and subset it over the region of the Como Lake in Italy.
-
-After the download and processing finishes (it will take a while), the MODIS time series will be placed in `youroutfolder/VI_16Days_500m_v6/`. (In this case, I have them in "`D:/Temp/VI_16Days_500m_v6/`").
-
-Any single-date image can be accessed by simply opening it with a `raster` command:
-
-``` r
-modistsp_file <- "D:/Temp/VI_16Days_500m_v6/NDVI/MOD13A1_NDVI_2016_177.tif"
-my_raster <- raster(modistsp_file)
-plot(my_raster)
-```
-
-![](analyze_files/figure-markdown_github/unnamed-chunk-2-1.png)
-
-`RasterStack` (or GDAL vrt) time series containing all the processed data for a given parameter are saved in the "Time Series/RData" subfolder, and can be accessed by:
-
-``` r
-# Load the NDVI time series
-ts_folder <- "D:/Temp/VI_16Days_500m_v6/Time_Series/"
-in_virtual_file <- file.path(ts_folder, "RData/Terra/NDVI/MOD13A1_NDVI_1_2016_353_2016_RData.RData")
-ts_data <- get(load(in_virtual_file))
-ts_data
-```
-
-    ## class       : RasterStack 
-    ## dimensions  : 122, 65, 7930, 23  (nrow, ncol, ncell, nlayers)
-    ## resolution  : 463.3127, 463.3127  (x, y)
-    ## extent      : 711648.3, 741763.7, 5076054, 5132578  (xmin, xmax, ymin, ymax)
-    ## coord. ref. : +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs 
-    ## names       : MOD13A1_NDVI_2016_001, MOD13A1_NDVI_2016_017, MOD13A1_NDVI_2016_033, MOD13A1_NDVI_2016_049, MOD13A1_NDVI_2016_065, MOD13A1_NDVI_2016_081, MOD13A1_NDVI_2016_097, MOD13A1_NDVI_2016_113, MOD13A1_NDVI_2016_129, MOD13A1_NDVI_2016_145, MOD13A1_NDVI_2016_161, MOD13A1_NDVI_2016_177, MOD13A1_NDVI_2016_193, MOD13A1_NDVI_2016_209, MOD13A1_NDVI_2016_225, ... 
-    ## time        : 2016-01-01 - 2016-12-18 (range)
-
-, which gives us a 23-band `RasterStack` in the `ts_data` variable.
-
-This `RasterStack` can be analyzed using the functionalities for raster/raster time series analysis, extraction and plotting provided for example by the `raster` or `rasterVis` packages:
-
-``` r
-# plot some dates
-plot(ts_data[[c(1,5,10,15,20)]], axes = FALSE, horizontal = T)
-```
-
-![](analyze_files/figure-markdown_github/unnamed-chunk-4-1.png)
-
-``` r
[…]
