+++ index.Rmd
@@ -66,7 +66,7 @@ from within `R` scripts
 automatic updating of MODIS time series through scheduled jobs, both on [<i class="fa fa-windows" aria-hidden="true"></i>](articles/standalone_execution.html#on-windows) and [<i class="fa fa-linux" aria-hidden="true"></i>](articles/standalone_execution.html#on-linux)
 
 - Solutions to common **installation, data download and processing problems** can be found 
-in our [<i class="fa fa-question-circle-o aria-hidden="true"></i> faq](articles/faq.html)
+in our [<i class="fa fa-question-circle-o" aria-hidden="true"></i> faq](articles/faq.html)
 
 - Please **report any issues** you may encounter in our [issues page on github <i class="fa fa-github-square" aria-hidden="true"></i>](https://github.com/ropensci/MODIStsp/issues) .
 
@@ -74,7 +74,12 @@ ________________________________________________________________________________
 
 ## <i class="fa fa-newspaper-o" aria-hidden="true"></i> What's New 
 
-- 10/07/2018 - MODIStsp v. 1.3.6 is out. Check out the [Release Notes](https://github.com/ropensci/MODIStsp/releases/tag/v1.3.5) for further details !
+- 07/08/2018 - We are glad to report that MODIStsp is now included in the 
+[rOpenSci](https://ropensci.org/about/) packages' ecosystem. We thank reviewers
+Leah Wasser and Jeffrey Hanson for their valuable reviews, which helped us  
+further improving the package!
+
+- 10/07/2018 - MODIStsp v. 1.3.6 is out. Check out the [Release Notes](https://github.com/ropensci/MODIStsp/releases/tag/1.3.6) for further details !
 
 - 20/06/2018 - MODIStsp v. 1.3.5 is out. Check out the [Release Notes](https://github.com/ropensci/MODIStsp/releases/tag/v1.3.5) for further details !
 

[…]
