### 2018-08-07 #184 @lbusett
Hi @karthik ,

 I just transferred the repo to https://github.com/ropensci/MODIStsp after making the necessary changes.

(Note that the build on Travis is currently failing apparently due to a downtime of the NSIDC server)
